### 2023-07-24 #121 @NickleDave
Linking back to discussion on the presubmission issue:
https://github.com/pyOpenSci/software-submission/issues/114#issuecomment-1648683592

We will hold off on starting this review for just a bit (two weeks maximum) while we help @jagerber48 with some questions about development and scope.
