### 2023-08-03 #121 @jagerber48
I don't have any non-trivial breaking changes in mind for the package at this time (though stuff always seems to come up). The only API changes I have in mind are possible name changes to options or classes. There are a few I'm not 100% happy with so I may poll pyopensci or e.g. code review stack exchange to try to get some broader set of opinions on some naming choices.

Just two I've written down:
- `precision` is overloaded to mean (1) the number of digits after the decimal point to round to in `RoundMode.PREC` mode and (2) the number of sig figs to round to in `RoundMode.SIG_FIG`. It isn't really the best name for either of these. Maybe something like `num_round_digits`?
- `SciNum` and `SciNumUnc` used to be `sfloat` and `vufloat` before the code was refactored to use `Decimal` over `float`. I'm not really in deep love with any of these names.

Those are just two I had written down, but pretty much any name in the API is fair game for naming improvement suggestions.
