### 2023-08-18 #121 @jagerber48
`sciform` is now on version `0.26.2`. Version `0.26.0` introduces a number of trivial name changes to various options. `0.26.1` introduced more unit tests for better coverage (I learned how to measure unit test coverage) and helped me discover one bug and another bug/unexpected behavior. The first bug is fixed in `0.26.2`. The second bug requires some convention decisions to be made about the format specification mini language. See https://github.com/jagerber48/sciform/issues/29.
