### 2023-09-22 #121 @jagerber48
Here's an update on the current status of the package and some questions. Since making this submission I modified the code quite a bit moving up to version `0.29.0`. Since then I've taken a break from updating `sciform` because I think all of the core functionality I envisioned is in place. The tasks I see (I think it's likely other will arise during the review) in front of me are

- Trying to fully stabilize the API. I think it's close but there are few "opinion-based" questions I have about e.g. the naming of some options and similar scope. See below.
- Adding non-core features. Importantly there is one important feature I want to add which is pre-defined `FormatOptions` and corresponding `Formatter` classes. Hopefully these can exist and make it so that users rarely if ever have to actually interface with configuring their own options and can just select a pre-configured formatter and use that. Or maybe set pre-configured global default options and use a generic formatter.

For adding the non-core features I plan to wait until after the pyopensci review and possibly until after I release version 1.0.0 with a more stable API (which will also happen after the pyopensci review).
For stabilizing the API since I feel a lot of the question I have are opinion-based I'm excited to get opinions from other people than me! I'd hope that some of these questions can be debated or discussed during the review if they're in scope. Here are the current questions I have listed:

- `Enum`s. Right now when configuring a `FormatOptions` object the user has to import various `enum` objects like the `ExpMode` or `ExpFormat` `enum` to configure certain settings. Is it too much of a pain to have to import all of these objects to choose settings? What would be a better alternative? (I know bare strings could be used but these are error-prone...). Also in some cases a sentinel like `AutoDigits` or `AutoExpVal` needs to be imported. (@Batalex already suggested using `Literal`. There's also the possibility of a `StrEnum` which might support passing an enum instance or the string representation..)
- `Formatter` and `FormatOptions`. Right now creating a `Formatter` object is a two step process which first involves creating `FormatOptions` object then passing it to the `Formatter` constructor. This can be done in one line but takes up a lot of characters. The `FormatOptions` can't be dispensed with because it is also used to conveniently configure globa[l options and new sets of format options. I considered allowing passing the kwargs to form `FormatOptions` directly to the `Formatter` constructor, but there were various downsides to that approach because of just how many options there are. [See this discussion for further details.](https://pyopensci.discourse.group/t/how-to-avoid-repeating-a-long-list-of-keyword-options-throughout-a-package/331/15) Is this approach to burdensome? How could it be improved (taking into account all of the points made in the linked discussion).
- By default value/uncertainty pairs are formatted as `123 +/- 42`. But using the `unicode_pm=True` option it is possible to format this as `123 ± 42`. Should the `unicode_pm` option be `True` by default? Or even further, should `unicode_pm=True` just be the only behavior and the option is dispensed with entirely??
- I'm not in love with some of the names chosen for various objects and options. Looking for suggestions for improvements on naming ANYWHERE throughout the package (even possibly the package name). But a few that I've singled out are:
    - Is `top_dig_place` an ok name?
    - Is `superscript_exp` an ok name? Maybe just `superscript`?
    - Is `bracket_unc` an ok name?
    - Is `SciNum` an ok name?
    - Is `SciNumUnc` an ok name?
- Originally I wasn't planning support for binary formatting modes but I was inspired to do so by the  [prefixed](https://github.com/Rockhopper-Technologies/prefixed) package. Right now binary formatting modes are accessed via dedicated exponent modes. I wonder if the two binary modes should instead be accessed via a separate option indicating the exponent base (10 or 2 at the moment). I'm also curious  to query interest levels in the binary formatting. I personally have no interest in the binary formatting since my work is mostly in physical sciences and I use decimals and SI. But I imagine people more in computer sciences may have more need for binary formatting.
- Right now there's an issue with the `sciform` FSML where the separator options and rounding mode option collide. [See the github issue.](https://github.com/jagerber48/sciform/issues/29) This issue came about from when i was trying to include all options in the FSML. But since then I've built out `FormatOptions` much more, setup global configuration options, and abandoned the idea that all options should be configurable via the mini language. I'm leaning towards just dropping the ability to configure separators from the FSML. Objections to this approach?

These are some of the main *existing* api features that I have questions about. I.e. making decisions on the above will constitute breaking changes, so I want to address these before 1.0.0. I am ALSO open to feature requests but will be implementing those with lower priority than stabilizing the API.

Thank you so much for taking the time to read this and also thank you very much for any feedback you are able to provide! I'm also curious to hear suggestions for other communities where I could seek feedback.
