### 2023-10-05 #121 @Batalex
:wave: Hi @isabelizimm  and @machow! Thank you for volunteering to review for pyOpenSci. I am truly excited by the team we have on this review.

@jagerber48, Isabel and Michael are involved with quarto, and I believe they have some pretty good opinions on how to format numbers for publishing. I am looking forward to seeing their review!

Please get in touch with any questions or concerns! Your review is due: 27 October 2023.
---

Reviewers: @isabelizimm, @machow
Due date: 2023/10/27
