### 2023-10-17 jagerber48/sciform#65 @jagerber48
`pdg_sig_figs` fails with 0 uncertainty when value is non-positive
```
from sciform import Formatter, FormatOptions
sform = Formatter(FormatOptions(pdg_sig_figs=True))
sform(-1, 0)
sform(0, 0)
```
Both the `sform(-1, 0)` and `sform(0, 0)` reach code in `get_pdg_round_digit()` that should be unreachable! The issue is that in this case the value is being used to determine the number of sig figs since the error is zero and this case isn't handled properly yet.
