### 2023-10-21 4d872aa
Don't use pdg sig figs unless the uncertainty drives the number of sig figs.
