+++ src/sciform/formatting.py
@@ -161,8 +161,8 @@ def format_val_unc(val: Decimal, unc: Decimal,
     else:
         round_driver = val
         '''
-        Don't use pdg sig figs unless the uncertainty drivers the number of sig
-        figs.
+        Don't use pdg sig figs if the uncertainty doesn't drive the number of 
+        sig figs.
         '''
         use_pdg_sig_figs = False
 
