+++ tests/test_val_unc_formatter.py
@@ -41,6 +41,20 @@ class TestFormatting(unittest.TestCase):
 
         self.run_val_unc_formatter_cases(cases_list)
 
+    def test_bracket_unc_invalid_unc(self):
+        cases_list = [
+            ((123, 0), [
+                (FormatOptions(bracket_unc=True), '123(0)')
+            ]),
+            ((123, float('nan')), [
+                (FormatOptions(bracket_unc=True), '123(nan)')
+            ]),
+            ((123, float('inf')), [
+                (FormatOptions(bracket_unc=True), '123(inf)')
+            ])
+        ]
+
+        self.run_val_unc_formatter_cases(cases_list)
     def test_prefix(self):
         cases_list = [
             ((123.456, 0.789), [
@@ -239,7 +253,7 @@ class TestFormatting(unittest.TestCase):
             ]),
             ((0, float('nan')), [
                 (FormatOptions(pdg_sig_figs=True), '0 +/- nan')
-            ]),
+            ])
         ]
 
         self.run_val_unc_formatter_cases(cases_list)
