### 2023-10-22 96e0571
More tests of uncertainty formatting with positive and negative values and uncertainty larger than value.
