### 2023-10-22 6f0524f
use abs(val) for magnitude comparisons with unc and don't strip single 0 for zero uncertainty
