### 2023-10-23 043cadf
compress bugfix changelog and reference issues
