### 2023-10-23 #121 @jagerber48
Started running the code on real applications myself some more and quickly found some important bugs. Just made a release of version `0.29.1` that fixes these bugs.
https://github.com/jagerber48/sciform/releases/tag/0.29.1
