### 2023-10-23 jagerber48/sciform#73 @jagerber48
`pdg_sig_figs` makes it impossible to control number of digits for single value formatting.
If `pdg_sig_figs=True` then an exception is raised if `ndigits != AutoDigits`. This limits the functionally for using one formatter for both uncertainty formatting and single-value formatting. It's not clear what the desired behavior should be here.
