### 2023-10-25 #121 @isabelizimm
# 🚧 WIP REVIEW 🚧

I'll finish this up tomorrow (10/25), but didn't want to lose my notes here 😄

It might be useful to show how to install a development version (eg, fork the repository and install with `pip install  -e .` or `pip install git+https://github.com/jagerber48/sciform`)

has examples page

includes documentation, but some functions do not have parameters documented. eg, `set_global_defaults_rendered` has the parameter `rendered_format_options` but [it is not shown in the API docs](https://sciform.readthedocs.io/en/stable/api.html#sciform.set_global_defaults)

no contributing guidelines

- [X] Any additional setup required to use the package (authentication tokens, etc.) (N/A for sciform)

Nit: to make documentation even easier to find,

Has 100% test coverage! :tada:

---

- for more readable code, you might want to add black formatting. when running black on this codebase, I had the following output

(px) isabelzimmerman@Isabels-MBP-3 sciform % black src/sciform/
reformatted /Users/isabelzimmerman/code/github/sciform/src/sciform/__init__.py
reformatted /Users/isabelzimmerman/code/github/sciform/src/sciform/formatter.py
reformatted /Users/isabelzimmerman/code/github/sciform/src/sciform/prefix.py
reformatted /Users/isabelzimmerman/code/github/sciform/src/sciform/scinum.py
reformatted /Users/isabelzimmerman/code/github/sciform/src/sciform/grouping.py
reformatted /Users/isabelzimmerman/code/github/sciform/src/sciform/modes.py
reformatted /Users/isabelzimmerman/code/github/sciform/src/sciform/fsml.py
reformatted /Users/isabelzimmerman/code/github/sciform/src/sciform/formatting.py
reformatted /Users/isabelzimmerman/code/github/sciform/src/sciform/format_utils.py
reformatted /Users/isabelzimmerman/code/github/sciform/src/sciform/format_options.py

All done! ✨ 🍰 ✨
10 files reformatted.
```
-

#### Response to maintainer questions

- Enums. Right now when configuring a FormatOptions object the user has to import various enum objects like the ExpMode or ExpFormat enum to configure certain settings. Is it too much of a pain to have to import all of these objects to choose settings? What would be a better alternative? (I know bare strings could be used but these are error-prone...). Also in some cases a sentinel like AutoDigits or AutoExpVal needs to be imported. (@Batalex already suggested using Literal. There's also the possibility of a StrEnum which might support passing an enum instance or the string representation..)
- Formatter and FormatOptions. Right now creating a Formatter object is a two step process which first involves creating FormatOptions object then passing it to the Formatter constructor. This can be done in one line but takes up a lot of characters. The FormatOptions can't be dispensed with because it is also used to conveniently configure globa[l options and new sets of format options. I considered allowing passing the kwargs to form FormatOptions directly to the Formatter constructor, but there were various downsides to that approach because of just how many options there are. See this discussion for further details. Is this approach to burdensome? How could it be improved (taking into account all of the points made in the linked discussion).
- By default value/uncertainty pairs are formatted as 123 +/- 42. But using the unicode_pm=True option it is possible to format this as 123 ± 42. Should the unicode_pm option be True by default? Or even further, should unicode_pm=True just be the only behavior and the option is dispensed with entirely??

I think setting `unicode_pm=True` as a default is the right move! I think it is a delight `sciform` offers to users to not have to type ± :smile: But, there are a lot of configuration parameters and it could be missed, so this makes the out-of-the-box experience feel a little extra polished.

- Is top_dig_place an ok name?

I wonder if you could somehow work padding into the name? My thought process behind that: `top_dig_place` looks to be used in the general purpose `Formatter` where arguments have a variety of use cases (padding, decimal places, separators), so it's use for padding might not be apparent. (These are not very good suggestions, but something like `pad_to` or `pad_to_exp`?)

- Is superscript_exp an ok name? Maybe just superscript?

I have a slight preference to `superscript`.

- Is bracket_unc an ok name?

For `bracket_unc` and `SciNumUnc` alike, my brain did not go to uncertainty for `unc`. I do realize that a other arguments build off this name (`bracket_unc_remove_seps`), so brevity is important. (`bracket_uncertain` + `bracket_uncertain_seps` maybe? you would have to flip the boolean for the separators then since you drop the negative)

- Is SciNum an ok name?

I like SciNum.

- Is SciNumUnc an ok name?

It was not immediately apparent to me that `SciNumUnc` was for uncertainty, but once I made that connection, it made a lot of sense. `SciNumUncertain` is a bit longer, but more clear. I do think `Uncertainty` alone isn't verbose _enough_ since you still need your scientific number. FWIW I ended up typing SciNumU and then auto-completing the rest, so I'm not too afraid of longer names.

- I wonder if the two binary modes should instead be accessed via a separate option indicating the exponent base (10 or 2 at the moment). I'm also curious to query interest levels in the binary formatting. I personally have no interest in the binary formatting since my work is mostly in physical sciences and I use decimals and SI.

For binary formatting, maybe open an issue/discussion about it and ask people to :thumbsup: if they would like to see it implemented? That way you're not building things you don't have interest in. However, if you already have the plumbing in place, and it's just a matter of exposing an argument, and you feel compelled to do so, then that seems like a good interrim move!

- Right now there's an issue with the sciform FSML where the separator options and rounding mode option collide. See the github issue. This issue came about from when i was trying to include all options in the FSML. But since then I've built out FormatOptions much more, setup global configuration options, and abandoned the idea that all options should be configurable via the mini language. I'm leaning towards just dropping the ability to configure separators from the FSML. Objections to this approach?

- Should I include changelog even for minor version bumps?

I think yes! For every release, it is super useful to see the changes. If it feels like too much effort to go back and write everything manually, even something small like using the `Generate Release Notes` button when creating a new release is sufficient. :smile:

- Right now on my github PRs I have github actions to run tests and linting on different python versions. Is there a way for me to run these tests all on my local system with github? Or is this type of automation typically done with some third party build/automation tool?

Something you can do locally is install and use `pre-commit`. You can have pre-commit hooks do a variety of things, but some common use cases are running black and flake. You can also configure testing to run on pre-commit, but this is less common to do since sometimes tests can take a while to run.
