### 2023-10-28 #121 @Batalex
Thanks a bunch, @isabelizimm and @machow, for the reviews!

@jagerber48 I also took a few notes (without the suitable format, editor's privilege 🐈‍⬛ )

# Project documentation

## README

Instructions are clear, and the statement of need as well.
Lacking a few elements from our guide, nothing too serious.
Python version compatibility needs to be clarified.

> Most of those issues have been solved in the meantime.

## Contributing

Missing as of now.

## License

Ok.

## Changelog

Ok.

# Installation

Installation resolves correctly.

# pyproject.toml

The classifiers are lacking (python version).
Nice setup for the dynamic version & readme.

> More classifiers in recent releases.

# General remarks and QoC

- I would advise using a code formatter, such as black.
- I would also strongly advise sorting imports using isort or ruff (❤️ ).
(Maybe isort since the linter used is flake8, but I sure won't stop you from migrating to Ruff)

- Is there any reason for the comments to start with `#:`?
- A few type hints are wrong; a parameter cannot take a default value of `None` if its type does not allow it.
- There are a lot of exceptions raised throughout the code base.
We could simplify it by removing those checks in private functions.

## `modes.py`

- FillMode: the exception raised is unneeded. Since enums can only take a finite set of values, we cannot have anything other than space or zero.
- GroupingSeparator: same as above.

I am on the fence concerning using Literal type hints with the values.

## `format_utils`

- Maybe use a type alias for Number
- A few type hints are missing (get_top_digit)
- `get_mantissa_exp_base` seems a little too complex. Maybe it can be simplified.
 - `get_round_digit`. The two innermost conditions lead to the same result.

# API design

I want to propose something regarding the use of the sciform. Currently, the return types for the format functions are strings, which are acceptable for the project's current goal.
I would like to propose making a `FormattedValue` class inheriting from `str`.
The reason is that we could then include methods such as `_repr_html_` to use html syntax in notebooks, or other mime types (LaTeX export to use with quarto?).
