### 2023-10-30 #121 @machow
I've been chewing on the question of SciformOptions and Enums more. In case it's helpful, here's a 1-minute screencast showing some of the nice autocomplete Literals can provide!

https://www.loom.com/share/3ba1eadec7e84689b20d39406c27fa1e
