### 2023-11-01 #121 @jagerber48
@isabelizimm Thanks so much for your thorough feedback! Responding here and using this as a todo list.

- [ ] I see that `set_global_defaults` is not documented properly in the API docs. I wonder why that is? Those are being auto generated by sphinx. I will look into it.
- [ ] I can try to add examples in the API docs for `FormatOptions` and the `FormatOptions.merge()` method. I haven't put much effort into the API docs yet, I've focused more on some of the other docs. This is a good suggestion to take a closer look at the API docs.
- [ ] You mention annotated examples about the FSML. Yes, I agree that would be very helpful. `_,_.4` is a super complicated formatting specification. I have a number of thoughts here. The first is a comment: This package *started* only as a format specification mini language to extend the python built in FSML, motivated by the uncertainties package and [this python discourse discussion](https://discuss.python.org/t/new-format-specifiers-for-string-formatting-of-floats-with-si-and-iec-prefixes/26914/1). But then a suggestion came up to use a class or functional approach to doing the formatting. This seemed interesting and lifted some of the shackles of trying to work within the constraints of an FSML. The class-based approach could handle many more options and complicated option combinations. The FSML grew to reflect *some* of this complexity but not all of it. Looking back I have two ideas on this. One is to strip down the `sciform` FSML, specifically excluding the separator control from the FSML. The other is to just remove the FSML completely. It seems just as easy to pass a number into a `Formatter` as it is to convert a number to `SciNum` or `SciNumUnc` and then use a format string. There are some trade offs that I think I'll discuss below. But, I agree that some annotated FSML examples would be very helpful.
- [ ] I can add a `CODE_OF_CONDUCT` file. Thank you for the links!
- [ ] I can add a `CONTRIBUTING` file! Again thank you for the link!
- [ ] Thanks for the input that `unicode_pm=True` can be default. I agree that `unicode_pm=True` seems fine and won't cause people issues. I wonder if I should go even further and remove it as an option and only bring it back if I hear a complaint that someone can't display unicode characters and wants an options to display it as `+/-`?
- [ ] `top_dig_place`: Yes, I agree with you that some name about padding would be much clearer. Not sure exactly what it should be. In many words: This parameter specifies the digits place (left of the decimal point) to which either empty spaces or zeros should be padded to a number. `pad_digit`, `pad_digit_place`, `left_pad_digit`.
- [ ] Thanks for the preference on `superscript`
- [ ] Ok regarding the "`unc`" options. [this issue](https://github.com/jagerber48/sciform/issues/73) has made me realize there are issues with trying to use the same set of `FormatOptions` for both single values and for value/uncertainty pairs. So I'm considering something like having a separate data class for value/uncertainty pair formatting options and this may open the door to better named options. That aside, I could see renaming `bracket_unc` to `bracket_uncertainty`. Not sure what to do about `bracket_unc_remove_seps`, I find that to just be a terrible name.
- [ ] Thanks for the feedback on `SciNum`.
- [ ] Thanks also for the feedback here. One idea would be `SciNumErr`, but I think, scientifically/statistically, we are really talking about uncertainty and not error. Maybe I can make it more obvious in the documentation that `SciNumUnc` is modelling a SCIentific NUMber along with UNCertainty.
- [ ] Thanks for the tip to make an issue/discussion on binary formatting. I think that's exactly what I'll do.
- [ ] Ok, I think I'll keep doing a changelog for all version bumps. I like recording the changes in the changelog as well. I've found it to be a useful history for myself.
- [ ] Ok, I've started dabbling with pre-commit. It does look like I could use this to run tests and to formatting.
- [ ] I'll add the documentation url on github.
- [ ] I can add development install instructions. That may go well along with the contributing instructions.
- [ ] I'll look into black. I did use flake8 for the first time on this project. I'm personally not yet comfortable with a formatter *actually* making changes to my code, but I do like following code styles. I use pycharm and follow its style guidelines/warnings, but I did find that some stuff that got past pycharm didn't get past flake8. And it sounds like some stuff that got past pycharm and flake8 doesn't get past black?
