### 2023-11-01 #121 @jagerber48
@machow here's my response + todo list for your review, thanks again for taking the time to do this review!

- [ ] Ok thanks for the feedback on gigt branching
- [ ] I think since asking these questions I have set up some CI around this... Let me recall.. Yes, I overhauled my versioning and it helped with this issue. So now I use `setuptools_scm` to help manage the package version. There is no longer any instances of the package version appearing in the source code (except for in the changelog, but that is handled special, see below). I make PRs into the `main` branch and update the `CHANGELOG.rst` in an unreleased changes section (so I don't yet need to commit to a version number). Then once enough changes are collected I decide (in my brain) to make a new release. I make a PR whose only main job is to set the version appropriately in the changelog, though other cleanup could in principle happen in that PR. Then when that changelog version bump PR is merged in a make a github release with the same version number as was put in the changelog. This creates a tag which is then picked up by `setuptools_scm` to control the version number. There is a github action that publishes to pypi when a release like this happens (during the built setuptool_scm picks up the version from the tag) and readthedocs also picks up on the new version tag and automatically builds a new version of the docs.
- [ ] I'll checkout tox and act, thanks for the recs!
- [ ] thanks for feedback on names
- [ ] Enums and Literal, I saw you posted a video about this, I will check that out soon and give more of a response. Thank you for helping me think about this, it's been giving me trouble and I think it is very important for usability. Yes, it feels like a burden having the user import like 2-4 different Enums for standard use cases. I asked about stuff like this on StackExchange but kept getting answers along the lines of "it's not a burden on the user to have to import Enums to configure options", but I think they may have been imagining importing a *single* enum, and not like 4.
- [ ] `Formatter` and `FormatOptions`. Yeah, I appreciate the feedback that `Formatter` and `FormatOptions` can be one class. Or at least one class from the user's perspective. Originally it was combined and I split it out. But I think the split didn't happen exactly in the right way. I'll consider how I can merge this functionality. I think it would be a big improvement in usability, but may be a moderate to major change.
- [ ] Thanks for more thoughts on names!
- [ ] Yeah thanks for the feedback on FSML. See my comment above to isabelizimm. I agree that the access via `FormatOptions` seems to be easier and more user-friendly. Considering at least slimming down the FSML, or even possibly dropping it entirely.

Thanks again for the great feedback!
