### 2023-11-01 #121 @jagerber48
@Batalex ok, and my response to your comments:

- [ ] will add contributing information
- [ ] will add python version compatibility. Note: right now I think python 3.8 is not supported because of some typing feature I use. Debating if I want to make it compatible with 3.8 or not. Seems typing feature are a major culprit for breaking compatibility with even very recent versions of python.
- [ ] I'll look into black
- [ ] I'll look into import sorting
- [ ] Those comments that start `#:` have something to do with sphinx adding comments in as documentation for Enum member or something. So yes, there is a reason, I don't remember the exact details off hand.
- [ ] I will look into incorrect type hints. I think I got some mixed signals. I saw something that I interpreted to mean that a default value of `None` would coerce the type to be unioned with `None`, but then I saw that that was maybe an old behavior that was removed. I did have one round of adding `Optional` to lots of types like this, but may not have hit them all.
- [ ] I'm interested in what you say about removing checks in private functions. Maybe you or I could find an example and we could discuss. I asked specifically about this kind of thing on [python discourse](https://discuss.python.org/t/how-should-i-terminate-if-elif-elif-when-all-options-are-exhausted/31796/1). One of these exact exceptions actually ended up catching [this bug](https://github.com/jagerber48/sciform/issues/65).
- [ ] `FillMode` and `GroupingSeparator`, maybe I already changed this? I don't see exceptions in those classes in my current source code.
- [ ] type alias for Number is probably a good idea
- [ ] I'll check type hints, thanks!
- [ ] I can look at simplifying `get_mantissa_exp_base` and `get_round_digit`. These functions are part of the meat of the package that is doing the actual formatting algorithm. I was hoping to improve upon existing codebases approaches to formatting which I found to be pretty complex and hard to follow, but in the end, I think `sciform`'s functions can also be complex and hard to follow. But I think you're right that improvements can probably be made.
- [ ] API design suggestion for having a `FormattedValue` that could be converted to html or latex. Thanks for this suggestion. I only partially understand the use case and the implications but it seems like it could be very useful for broadening the scope of `sciform` beyond jupyter notebooks and terminal printouts. Would it make sense to open a [discussion idea topic](https://github.com/jagerber48/sciform/discussions) on the sciform page with this idea and we could discuss in more detail there? I'm interested in this, but this sounds like something further out that won't get completed within the timeline of this review (or in other words, I'd like to prioritize the other items before looking into this possible new feature).
