### 2023-11-01 #121 @jagerber48
> I've been chewing on the question of SciformOptions and Enums more. In case it's helpful, here's a 1-minute screencast showing some of the nice autocomplete Literals can provide!
>
> https://www.loom.com/share/3ba1eadec7e84689b20d39406c27fa1e

@machow  thank you so much for putting together this quick screencast. Ok, that was a pretty compelling comparison between `Enum` and `Literal`. I've been searching online for comparisons just like this between `Literal` and `Enum` in python but haven't found much yet. I guess this is because `Literal` is somewhat new. But it seems like, at least for this use case, `Literal` might just be better than `Enum`. The `Enum` feel like some sort of "proper" coding, but I can just imagine myself and users being like, why can't I just pass in a string? The `Literal` input fixes all of the shortcomings of using "magic" strings as inputs while still being able to accept string inputs.

I will definitely at least mock up a branch of `sciform` that uses literals instead of enums and see how it goes.
