### 2023-11-01 #121 @Batalex
Thank you for your very neat checklists, which will significantly reduce the mental charge of everyone here.
I see that you use a PR-based workflow even though you are the only contributor to the project. That's great! Feel free to request reviews directly on the PRs to avoid cluttering this issue.

In addition to Michael's loom on the subjet of Enum / Literal, since you are targeting Python 3.8+, it might be worth using Literal instead of Enum, because you have access to the `get_args` [function](https://docs.python.org/3/library/typing.html#typing.get_args) in the typing module. Here is an example of how you could use it.

```py
from typing import get_args, Literal, TypeAlias

DecimalSep: TypeAlias = Literal[".", ","]

def some_function(sep: DecimalSep):
    if sep not in get_args(DecimalSep):
        raise ValueError()

some_function(sep=";")  # raise
some_function(sep=",")  # OK

```

I think this approach is interesting, because we both have the ease of use of the string argument instead of a verbose Enum, as well as a runtime check.
