### 2023-11-02 #121 @jagerber48
@machow, @Batalex to be clear, from a user API perspective, the shift to `Literal` from `Enum` will involve ONLY accepting Literal inputs to configure options and no longer accepting Enums, correct? It would be a bad choice to accept either a Literal or the corresponding Enum from the user?
