### 2023-11-03 #121 @Batalex
Per [PEP 506](https://peps.python.org/pep-0586/#illegal-parameters-for-literal-at-type-check-time), there is no way to build a `Literal` from an expression.
Accepting both a `Literal` and a `Enum` would require that you duplicate the values throughout the code base.

```python
from enum import Enum
from typing import get_args, Literal, TypeAlias

class DecimalEnum(str, Enum):
    period = "."
    comma = ","

DecimalLit = Literal[".", ","]
DecimalSep: TypeAlias = DecimalLit | DecimalEnum

def some_function(sep: DecimalSep):
    if sep not in get_args(DecimalLit):
        raise ValueError()

some_function(sep=",")  # OK
some_function(sep=DecimalEnum.comma)  # OK
```
That is a valid possibility, I am just concerned by the confusion of the three symbols we need
