### 2023-11-04 #121 @jagerber48
@Batalex re: your last comment about an `Enum` + `Literal` approach. Thank you for putting together that example. So yes, the string representations would need to be duplicated in two spots. I also have a feeling the context type hint provided by IDE code completion would be long and less helpful in this case. It feels like `Enum` was a nice controlled strategy for this use case prior to `Literal`, but I think it's just going to easier for users if they can pass in strings. I's not like its unusual at all in scientific programming to use strings to select function options. `scipy` and `numpy` do that sort of thing all over the place. I think it's very rare that I've used a package that actually uses `Enum` for options. Just some people on stack exchange say it's a best practice. Now with `Literal`, it seems a lot of the downsides of using strings as inputs have gone away. Specifically, while a format string like `engineering` (for engineering exponent mode) would have been a "magic" string in the past, using `Literal['engineering', ...]` makes it no longer as much of a magic string. I will hash out an example that only uses `Literal`. I'll make a decision while I'm coding whether I want to convert the `Literal` string user input into an `Enum` for use by the backend code.
