### 2023-11-04 2bcf234
docstring for set_global_defaults
