### 2023-11-04 8588dd4
GlobalDefaultsContext docstring
