+++ docs/source/api.rst
@@ -48,6 +48,9 @@ Global Configuration
 ====================
 
 .. autoclass:: GlobalDefaultsContext()
+   :members:
+
+   .. automethod:: __init__
 
 .. autofunction:: print_global_defaults()
 
