+++ CHANGELOG.rst
@@ -5,7 +5,10 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 Unreleased
 ----------
 
-* There are no unreleased changes
+Fixed
+^^^^^
+
+* Cleaned up API documentation.
 
 ----
 
