### 2023-11-04 6e09d89
repr asdict(self) for RenderedFormat options and repr self._render() for unrendered format options
