+++ CHANGELOG.rst
@@ -11,6 +11,7 @@ Fixed
 * Fixed a bug where the `repr` for `FormatOptions` would return a string
   containing information about the global format options rather than
   about the specific `FormatOptions` instance.
+  [`#75 <https://github.com/jagerber48/sciform/issues/75>`_]
 
 ----
 
