### 2023-11-04 143174e
replace +/- with ± in tests
