+++ tests/test_val_unc_fsml.py
@@ -23,30 +23,30 @@ class TestFormatting(unittest.TestCase):
     def test_fixed(self):
         cases_list = [
             ((123.456, 0.789), [
-                ('f', '123.456 +/- 0.789'),
-                ('!1f', '123.5 +/- 0.8'),
-                ('!2f', '123.46 +/- 0.79'),
-                ('!3f', '123.456 +/- 0.789'),
-                ('!4f', '123.4560 +/- 0.7890'),
+                ('f', '123.456 ± 0.789'),
+                ('!1f', '123.5 ± 0.8'),
+                ('!2f', '123.46 ± 0.79'),
+                ('!3f', '123.456 ± 0.789'),
+                ('!4f', '123.4560 ± 0.7890'),
                 ('!2f()', '123.46(79)')
             ]),
             ((-123.456, 0.789), [
-                ('f', '-123.456 +/- 0.789'),
-                ('!1f', '-123.5 +/- 0.8'),
-                ('!2f', '-123.46 +/- 0.79'),
-                ('!3f', '-123.456 +/- 0.789'),
-                ('!4f', '-123.4560 +/- 0.7890'),
+                ('f', '-123.456 ± 0.789'),
+                ('!1f', '-123.5 ± 0.8'),
+                ('!2f', '-123.46 ± 0.79'),
+                ('!3f', '-123.456 ± 0.789'),
+                ('!4f', '-123.4560 ± 0.7890'),
                 ('!2f()', '-123.46(79)')
             ]),
             ((0.789, 123.456), [
-                ('f', '0.789 +/- 123.456'),
-                ('!1f', '0 +/- 100'),
-                ('!2f', '0 +/- 120'),
-                ('!3f', '1 +/- 123'),
-                ('!4f', '0.8 +/- 123.5'),
-                ('!5f', '0.79 +/- 123.46'),
-                ('!6f', '0.789 +/- 123.456'),
-                ('!7f', '0.7890 +/- 123.4560'),
+                ('f', '0.789 ± 123.456'),
+                ('!1f', '0 ± 100'),
+                ('!2f', '0 ± 120'),
+                ('!3f', '1 ± 123'),
+                ('!4f', '0.8 ± 123.5'),
+                ('!5f', '0.79 ± 123.46'),
+                ('!6f', '0.789 ± 123.456'),
+                ('!7f', '0.7890 ± 123.4560'),
                 ('!1f()', '0(100)'),
                 ('!2f()', '0(120)'),
                 ('!3f()', '1(123)'),
@@ -56,14 +56,14 @@ class TestFormatting(unittest.TestCase):
                 ('!7f()', '0.7890(123.4560)')
             ]),
             ((-0.789, 123.456), [
-                ('f', '-0.789 +/- 123.456'),
-                ('!1f', '0 +/- 100'),
-                ('!2f', '0 +/- 120'),
-                ('!3f', '-1 +/- 123'),
-                ('!4f', '-0.8 +/- 123.5'),
-                ('!5f', '-0.79 +/- 123.46'),
-                ('!6f', '-0.789 +/- 123.456'),
-                ('!7f', '-0.7890 +/- 123.4560'),
+                ('f', '-0.789 ± 123.456'),
+                ('!1f', '0 ± 100'),
+                ('!2f', '0 ± 120'),
+                ('!3f', '-1 ± 123'),
+                ('!4f', '-0.8 ± 123.5'),
+                ('!5f', '-0.79 ± 123.46'),
+                ('!6f', '-0.789 ± 123.456'),
+                ('!7f', '-0.7890 ± 123.4560'),
                 ('!1f()', '0(100)'),
[…]
