### 2023-11-05 3402fba
replace +/- by ± in docs
