### 2023-11-05 cee8bd5
fix doctests
