+++ CHANGELOG.rst
@@ -13,6 +13,16 @@ Fixed
   about the specific `FormatOptions` instance.
   [`#75 <https://github.com/jagerber48/sciform/issues/75>`_]
 
+Removed
+^^^^^^^
+
+* Removed the ``unicode_pm`` feature which allowed toggling between
+  using ``'+/-'`` or ``'±'`` in value/uncertainty strings. Previously
+  ``unicode_pm`` defaulted to ``False`` so that ``'+/-'`` was the
+  default behavior. Now the default behavior is to use ``'±'`` and there
+  is no way to change to the old ``'+/-'`` behavior.
+  [`#10 <https://github.com/jagerber48/sciform/discussions/10>`_]
+
 ----
 
 0.29.1 (2023-10-22)
