### 2023-11-06 2e59065
Fix some changelog formatting
