+++ src/sciform/format_options.py
@@ -263,7 +263,7 @@ class FormatOptions:
         return rendered_format_options
 
     def __repr__(self):
-        return repr(self.render())
+        return repr(self._render())
 
 
 def validate_options(options: Union[FormatOptions, RenderedFormatOptions]):
