### 2023-11-07 2c12374
make render a public (but non-api-exposed) function (as opposed to public or private class/instance method)
