### 2023-11-07 #121 @jagerber48
I'm working on contributing guidelines for `sciform`. I haven't made many if any PRs on other people's repos on github. Trying to figure out how PRs from other people would work on `sciform`. My basic understanding is that someone would make a fork of `sciform` that lives in their user space, they would branch off `main` and make their changes. Then I think they PR their forked version into the main version in my user space. Questions:

- I can't tell if this would work for `sciform` for any user or if I would have to grant specific users the ability to make a PR from a fork.
- How would this process look in the git history for `sciform`? If the PR is approved/merged in would it just appear as a new commit on the `main` branch? Since the contributors feature branch lives in their space instead of in the main `sciform` repo in my user space?
- Do I need to describe this process in the contributing guidelines if I want to welcome PRs from other devs?
