### 2023-11-07 3e938aa
add CONTRIBUTING.rst
