+++ CONTRIBUTING.rst
@@ -1,0 +1,23 @@
+Contributing Guidelines
+=======================
+
+Contributions to ``sciform`` are welcome.
+Here are some ways you can contribute:
+
+* Create a
+  `discussion topic <https://github.com/jagerber48/sciform/discussions>`_
+  if you have an idea for a new feature or a general topic for
+  discussion.
+* Create an `issue <https://github.com/jagerber48/sciform/issues>`_ if
+  you find a bug with ``sciform`` or a challenge with the interface.
+* If you would like to make improvements to the source code or
+  documentation then you may do so directly by opening a
+  `pull request <https://github.com/jagerber48/sciform/pulls>`_.
+
+The current main goal for the development of ``sciform`` is
+stabilization of the public user interface.
+To this end, ``sciform`` is seeking feedback/suggestions about
+ease-of-use for the user interface as well as naming suggestions for the
+various objects and options that users interact with.
+New feature request may be backlogged until the interface is stabilized,
+however, new feature requests are always welcomed!
