### 2023-11-07 ca1f6cc
Add notes about compatibility with python >=3.9
