### 2023-11-07 #121 @NickleDave
> I can't tell if this would work for sciform for any user or if I would have to grant specific users the ability to make a PR from a fork.

No, you don't need to grant anyone permissions for them to be able to fork and make a PR. If only you could 😇

> How would this process look in the git history for sciform? If the PR is approved/merged in would it just appear as a new commit on the main branch? Since the contributors feature branch lives in their space instead of in the main sciform repo in my user space?

Basically: yes, a new commit on main if you "squash and merge", or all the commits if you merge the entire branch
More detail: it depends on how you merge
https://docs.github.com/en/repositories/configuring-branches-and-merges-in-your-repository/configuring-pull-request-merges/about-merge-methods-on-github
I recommend to let contributors do whatever in a branch and then just squash and merge as one commit: https://jacobtomlinson.dev/posts/2022/dont-prematurely-squash/rebase-and-force-push-your-prs/

> Do I need to describe this process in the contributing guidelines if I want to welcome PRs from other devs?

Ideally although you can probably point to one of many existing resources and say "we do this", e.g. "we use a standard PR workflow similar to [NumPy](https://numpy.org/doc/stable/dev/index.html) and [scikit-learn](https://www.youtube.com/watch?v=5OL8XoMMOfA)"

You might find this example from the PyGMT docs helpful:
https://github.com/GenericMappingTools/pygmt/blob/main/CONTRIBUTING.md
Note how they split out a separate guide that lives in the rendered docs:
https://github.com/GenericMappingTools/pygmt/blob/main/doc/contributing.md
