### 2023-11-18 #121 @jagerber48
I want to give an update. I've checked off a couple of the minor changes, but I'm now delving into an experiment on a major refactor. This refactor involves two thing that I think are major improvements to the UI, but, for now, seem to make developing on the back end a little more of a headache. That said, my goal is to release a version 1.0 of this project and my main criteria for doing that is I want to have a nice user interface. "easiness of maintainability" is not a must-have for releasing 1.0, so I've made the decision, at the moment, to explore what things would look like if I make the UI as nice as possible and then just work around that on the back end. The changes are:

- Move from using `Enum`s for users to specify options to using `Literal`s for users to specify options. My sense is that people naively using Python would just naturally use literal strings to specify different options, especially since this is a pattern in packages like `numpy`, `scipy`, `pandas`, etc. But the "proper" way to do this (if you look it up) is to use `Enum`s. This fixes some maintainability issues around using magic strings but it really an annoyance on the user to have to import more things into their name space and probably to even have to learn what `Enum`s are and how to use them if they don't already know. `Literal` fixes a lot of the shortcomings of both the simple magic string strategy and the `Enum` strategy. I think it will just be a lot more straightforward for users. There are downsides, however. My IDE, Pycharm, doesn't have autocompletions for `Literal` yet but there is an open issue about it, I'm optimistic it will come at some point. Also, `Literal` doesn't pass type checking if you first do something like `exp_mode = 'engineering'` and then pass `exp_mode` as a kwarg into the formatter constructor unless you explicitly specify `exp_mode` as a `Literal` in the right way.
- The second major change is I am eliminating `FormatOptions` from the user interface. Instead users just pass kwargs directly into either the `Formatter` constructor or the various global options settings functions. I introduced the `FormatOptions` approach previously because it meant the full list of options could be repeated many fewer times whiles till keeping everything nicely statically typed. See the conversation here: https://pyopensci.discourse.group/t/how-to-avoid-repeating-a-long-list-of-keyword-options-throughout-a-package/331/15. However, the `FormatOptions` object really doesn't provide any value to the user interface. In fact, it just further pollutes the user's namespace. So for now I've reverted to a proliferation of kwargs all over the source code. However, when PEP 692 (kwargs unpacking type annotations) is supported on pycharm (it's probably already supported by other typing tools) I can move to just using `**kwargs` everywhere which will compress all those repetitions of all the options while still having good static type annotations.

So basically, I had made a few decisions for better static typing and maintainability on the back end, but they were at the cost to the user interface. I've now decided the user interface is more important and I'm refactoring the code to reflect this change. Once these changes are more finalized I'll link to a PR in this thread. I would appreciate getting feedbacks from folks here on (1) the changes in user-friendliness and (2) the changes to the source code.
