### 2023-11-18 136652c
Major refactor
- Remove FormatOptions from UI
- Remove Enums from UI
- Update unit tests to reflect these changes
