+++ tests/test_float_fsml.py
@@ -1,6 +1,6 @@
 import unittest
 
-from sciform import SciNum, GlobalDefaultsContext, FormatOptions
+from sciform import SciNum, GlobalDefaultsContext
 
 
 FloatFSMLCases = list[tuple[float, list[tuple[str, str]]]]
@@ -419,7 +419,7 @@ class TestFormatting(unittest.TestCase):
             ])
         ]
 
-        with GlobalDefaultsContext(FormatOptions(nan_inf_exp=True)):
+        with GlobalDefaultsContext(nan_inf_exp=True):
             self.run_float_fsml_cases(cases_list)
 
     def test_separators(self):
