### 2023-11-18 #121 @jagerber48
https://github.com/jagerber48/sciform/pull/82 here's the PR. As of right now the source code has been re-worked and the tests in the `tests` directory have been re-written and are passing on my system. The main task ahead is rewriting a bit of documentation and the doctests.
