### 2023-11-18 14f0427
eliminate comments for auto documentation of modes (these will no longer need to be documented)
