### 2023-11-19 dd6a8c7
Remove FormatOptions and Enums from API documentation
