+++ docs/source/api.rst
@@ -7,35 +7,11 @@ API
 Formatting
 ==========
 
-.. autoclass:: FormatOptions()
-   :members: merge
-
 .. autoclass:: Formatter()
 
 .. autoclass:: SciNum()
 
 .. autoclass:: SciNumUnc()
-
-Mode Enums
-==========
-
-.. autoclass:: FillMode()
-   :members:
-
-.. autoclass:: SignMode()
-   :members:
-
-.. autoclass:: GroupingSeparator()
-   :members:
-
-.. autoclass:: RoundMode()
-   :members:
-
-.. autoclass:: ExpMode()
-   :members:
-
-.. autoclass:: ExpFormat()
-   :members:
 
 Auto Options
 ============
