### 2023-11-19 02e9963
Move option kwarg documentation into Formatter (and update text)
