+++ src/sciform/formatter.py
@@ -13,17 +13,17 @@ from sciform.format_utils import Number
 class Formatter:
     """
     :class:`Formatter` is used to convert numbers and pairs of numbers
-    into formatted strings. Formatting options are configured using
-    :class:`FormatOptions`. Any unpopulated format options will be
-    populated from the global default options at format time. See
-    :ref:`formatting_options` for more details on the available options.
-    If no options are provided then the global default options are used
-    for all options.
+    into formatted strings according to a variety of formatting options.
+    See :ref:`formatting_options` for more details on the available
+    options. Any options which are unpopulated (have the value ``None``)
+    will be populated at format time by the corresponding values in the
+    globally configured default options. See :ref:`global_config` for
+    details about how to view and modify the global default options.
 
     >>> from sciform import Formatter
     >>> sform = Formatter(exp_mode='engineering',
     ...                   round_mode='sig_fig',
-    ...                   ndigits=4))
+    ...                   ndigits=4)
     >>> print(sform(12345.678))
     12.35e+03
 
@@ -32,25 +32,12 @@ class Formatter:
 
     >>> sform = Formatter(exp_mode='engineering',
     ...                   round_mode='sig_fig',
-    ...             ndigits=2))
+    ...             ndigits=2)
     >>> print(sform(12345.678, 3.4))
     (12.3457 ± 0.0034)e+03
 
-    :class:`FormatOptions` instances store all the configuration options
-    used to format numbers and number/uncertainty pairs. See
-    :ref:`formatting_options` for more details on the available options.
-    :class:`FormatOptions` instances are used to create
-    :class:`Formatter` instances and to modify the global default
-    configuration.
-
-    It is not necessary to provide input for all options. At format
-    time, any un-populated options (indicated by a ``None`` value) will
-    be populated with the corresponding options from the global default
-    options. See :ref:`global_config` for details about how to view and
-    modify the global default options.
-
     The following checks are performed when creating a new
-    :class:`FormatOptions` object:
+    :class:`Formatter` object:
 
     * ``ndigits`` >= 1 for significant figure rounding mode
     * ``exp_val`` must be consistent with the exponent mode. If
@@ -62,100 +49,110 @@ class Formatter:
         engineering modes
       * ``exp_val`` must be a multiple of 10 for binary iec mode
 
-    * ``upper_separator`` may be any :class:`GroupingSeparator` but must
-      be different from ``decimal_separator``
-    * ``decimal_separator`` may only be :class:`GroupingSeparator.POINT`
-      or :class:`GroupingSeparator.COMMA`
-    * ``lower_separator`` may only be :class:`GroupingSeparator.NONE`,
-      :class:`GroupingSeparator.SPACE`, or
-      :class:`GroupingSeparator.UNDERSCORE`
+    * ``upper_separator`` may be any of ``['', ',', '.', ' ', '_']`` but
+      must be different from ``decimal_separator``
+    * ``decimal_separator`` may be any of ``['.', ',']``
+    * ``lower_separator`` may be any of ``['', ' ', '_']``
     * if ``pdg_sig_figs=True`` then ``ndigits=None`` or
       ``ndigits=AutoDigits``.
 
-    :param exp_mode: :class:`ExpMode` indicating the formatting
-      mode to be used.
-    :param exp_val: :class:`int` or :class:`AutoExpVal` sentinel
-      indicating a value which must be used for the exponent. If
-      specified, this value must be 0 for fixed point and percent modes,
[…]
