### 2023-11-19 799c76a
Formatter doesn't need the format method which is non-public anyways.
