### 2023-11-19 fa24ad7
some global configuration doc updating
