### 2023-11-19 f6192cc
Update Formatter documentation, remove FormatOptions in usage.rst and start doing some enum -> literal replacements in doctests.
