### 2023-11-19 26369ba
More replacements, usage.rst doctest passing.
