### 2023-11-19 a49d3c4
fix doctests
