+++ docs/source/options.rst
@@ -35,8 +35,8 @@ Fixed Point
 Fixed point notation is standard notation in which a number is displayed
 directly with no extra exponent.
 
->>> from sciform import Formatter, FormatOptions, ExpMode, RoundMode
->>> sform = Formatter(FormatOptions(exp_mode=ExpMode.FIXEDPOINT))
+>>> from sciform import Formatter
+>>> sform = Formatter(exp_mode='fixed_point')
 >>> print(sform(123.456))
 123.456
 >>> print(sform(123.456, 0.001))
@@ -51,7 +51,7 @@ Percent mode is similar to fixed point mode.
 For percent mode, the number is multiplied by 100 and a ``%`` symbol is
 appended to the end of the formatted string.
 
->>> sform = Formatter(FormatOptions(exp_mode=ExpMode.PERCENT))
+>>> sform = Formatter(exp_mode='percent')
 >>> print(sform(0.12345))
 12.345%
 >>> print(sform(0.12345, 0.001))
@@ -66,7 +66,7 @@ Scientific notation is used to display base-10 decimal numbers.
 In scientific notation, the exponent is uniquely chosen so that the
 mantissa ``m`` satisfies ``1 <= m < 10``.
 
->>> sform = Formatter(FormatOptions(exp_mode=ExpMode.SCIENTIFIC))
+>>> sform = Formatter(exp_mode='scientific')
 >>> print(sform(123.456))
 1.23456e+02
 >>> print(sform(123.456, 0.001))
@@ -99,7 +99,7 @@ Here it may be more convenient to display the number with an exponent of
 to an exponent of 8 which would be chosen for standard scientific
 notation.
 
->>> sform = Formatter(FormatOptions(exp_mode='engineering'))
+>>> sform = Formatter(exp_mode='engineering')
 >>> print(sform(123.456))
 123.456e+00
 >>> print(sform(123.456, 0.001))
@@ -114,7 +114,7 @@ Shifted engineering notation is the same as engineering notation except
 the exponent is chosen so that the mantissa ``m`` satisfies
 ``0.1 <= m < 100``.
 
->>> sform = Formatter(FormatOptions(exp_mode='engineering'))
+>>> sform = Formatter(exp_mode='engineering_shifted')
 >>> print(sform(123.456))
 0.123456e+03
 >>> print(sform(123.456, 0.001))
@@ -128,7 +128,7 @@ Binary
 Binary formatting can be chosen to display a number in scientific
 notation in base-2.
 
->>> sform = Formatter(FormatOptions(exp_mode=ExpMode.BINARY))
+>>> sform = Formatter(exp_mode='binary')
 >>> print(sform(256))
 1b+08
 
@@ -146,7 +146,7 @@ exponent is constrained to be a multiple of 10, consistent with the
 IEC binary prefixes.
 The mantissa ``m`` satisfies ``1 <= m < 1024``.
 
->>> sform = Formatter(FormatOptions(exp_mode=ExpMode.BINARY_IEC))
+>>> sform = Formatter(exp_mode='binary_iec')
 >>> print(sform(2048))
 2b+10
 
@@ -157,9 +157,9 @@ Fixed Exponent
 
 The user can coerce the exponent for the formatting to a fixed value.
 
->>> sform = Formatter(FormatOptions(
-...             exp_mode=ExpMode.SCIENTIFIC,
-...             exp_val=3))
+>>> sform = Formatter(
+...             exp_mode='scientific',
+...             exp_val=3)
 >>> print(sform(123.456))
 0.123456e+03
 
@@ -197,22 +197,21 @@ Furthermore, it is possible to customize :class:`FormatOptions`
 objects or the global configuration settings to map additional
[…]
