+++ docs/source/exp replacement.rst
@@ -148,11 +148,11 @@ For this reason, the usage of this notation is sometimes discouraged.
 Note that it is possible, using the ``extra_parts_per_forms`` option to
 override the standard mappings listed above.
 
->>> from sciform import Formatter, FormatOptions, ExpMode, ExpFormat
->>> sform = Formatter(FormatOptions(
+>>> from sciform import Formatter
+>>> sform = Formatter(
 ...             exp_mode='engineering',
-...             exp_format=ExpFormat.PARTS_PER,
-...             extra_parts_per_forms={-9: None, -12: 'ppb'}))
+...             exp_format='parts_per',
+...             extra_parts_per_forms={-9: None, -12: 'ppb'})
 >>> print(sform(33e-9))
 33e-09
 >>> print(sform(33e-12))
