### 2023-11-19 bfa5338
rename RenderedFormatOptions to RenderedOptions
