### 2023-11-19 #121 @jagerber48
Ok I would say the PR is mostly complete now. All tests are passing and documentation and changelog have been updated to reflect (1) the removal of `FormatOptions` from the public interface along with (2) the replacement of `Enum` options by string `Literal` options. This is a big enough change that I would appreciate another set of eyes on it before merging into `main`. Based on the comments so far I think this is the most significant change I'll be making as part of this review. I would appreciate if anyone could let me know if they could or wouldn't be able to have a look at this in, say, the next 2-3 weeks?

https://github.com/jagerber48/sciform/pull/82
