+++ CONTRIBUTING.rst
@@ -9,7 +9,7 @@ Here are some ways you can contribute:
   if you have an idea for a new feature or a general topic for
   discussion.
 * Create an `issue <https://github.com/jagerber48/sciform/issues>`_ if
-  you find a bug with ``sciform`` or a challenge with the interface.
+  you find a bug with ``sciform`` or another challenge with the package.
 * If you would like to make improvements to the source code or
   documentation then you may do so directly by opening a
   `pull request <https://github.com/jagerber48/sciform/pulls>`_.
@@ -19,5 +19,5 @@ stabilization of the public user interface.
 To this end, ``sciform`` is seeking feedback/suggestions about
 ease-of-use for the user interface as well as naming suggestions for the
 various objects and options that users interact with.
-New feature request may be backlogged until the interface is stabilized,
-however, new feature requests are always welcomed!
+While stabilizing the interface will take precedence for the time being,
+new feature requests are always welcome!
