### 2023-11-20 e0c7369
add to project.rst
