+++ tests/test_val_unc_formatter.py
@@ -4,7 +4,7 @@ from sciform import Formatter
 
 
 ValUncFormatterCases = list[tuple[tuple[float, float],
-                                      list[tuple[Formatter, str]]]]
+                                  list[tuple[Formatter, str]]]]
 
 
 class TestFormatting(unittest.TestCase):
@@ -23,17 +23,17 @@ class TestFormatting(unittest.TestCase):
             ((123.456, 0.789), [
                 (Formatter(bracket_unc=True), '123.456(789)'),
                 (Formatter(exp_mode='scientific',
-                               bracket_unc=True), '(1.23456(789))e+02'),
+                           bracket_unc=True), '(1.23456(789))e+02'),
                 (Formatter(exp_mode='engineering',
-                               bracket_unc=True), '(123.456(789))e+00'),
+                           bracket_unc=True), '(123.456(789))e+00'),
                 (Formatter(exp_mode='engineering_shifted',
-                               bracket_unc=True), '(0.123456(789))e+03'),
+                           bracket_unc=True), '(0.123456(789))e+03'),
                 (Formatter(exp_mode='scientific',
-                               exp_val=+1,
-                               bracket_unc=True), '(12.3456(789))e+01'),
+                           exp_val=+1,
+                           bracket_unc=True), '(12.3456(789))e+01'),
                 (Formatter(exp_mode='scientific',
-                               exp_val=-1,
-                               bracket_unc=True), '(1234.56(7.89))e-01'),
+                           exp_val=-1,
+                           bracket_unc=True), '(1234.56(7.89))e-01'),
             ])
         ]
 
@@ -64,37 +64,37 @@ class TestFormatting(unittest.TestCase):
         cases_list = [
             ((123.456, 0.789), [
                 (Formatter(bracket_unc=True,
-                               exp_mode='engineering',
-                               exp_format='prefix'), '123.456(789)'),
+                           exp_mode='engineering',
+                           exp_format='prefix'), '123.456(789)'),
                 (Formatter(bracket_unc=True,
-                               exp_mode='engineering_shifted',
-                               exp_format='prefix'),
+                           exp_mode='engineering_shifted',
+                           exp_format='prefix'),
                  '0.123456(789) k'),
             ]),
             ((123456, 789), [
                 (Formatter(bracket_unc=True,
-                               exp_mode='fixed_point',
-                               exp_format='prefix'), '123456(789)'),
+                           exp_mode='fixed_point',
+                           exp_format='prefix'), '123456(789)'),
                 (Formatter(bracket_unc=True,
-                               exp_mode='engineering',
-                               exp_format='prefix'), '123.456(789) k'),
+                           exp_mode='engineering',
+                           exp_format='prefix'), '123.456(789) k'),
                 (Formatter(bracket_unc=True,
-                               exp_mode='engineering_shifted',
-                               exp_format='prefix'), '0.123456(789) M')
+                           exp_mode='engineering_shifted',
+                           exp_format='prefix'), '0.123456(789) M')
             ]),
             ((123456e-9, 789e-9), [
                 (Formatter(bracket_unc=True,
-                               exp_mode='engineering',
-                               exp_format='parts_per'),
+                           exp_mode='engineering',
+                           exp_format='parts_per'),
                  '123.456(789) ppm'),
                 (Formatter(bracket_unc=True,
-                               exp_mode='engineering',
-                               exp_format='prefix'),
+                           exp_mode='engineering',
+                           exp_format='prefix'),
                  '123.456(789) μ'),
                 (Formatter(bracket_unc=True,
-                               exp_mode='engineering_shifted',
-                               exp_format='parts_per',
-                               add_ppth_form=True),
[…]
