### 2023-11-23 39ea4d7
changelog wording
