### 2023-11-23 2022fbf
remove merge method (using dataclasses.replace instead)
