### 2023-11-24 e40b71d
Add breaking flag to unicode pm entry in changelog
