+++ CHANGELOG.rst
@@ -27,11 +27,11 @@ Fixed
 Removed
 ^^^^^^^
 
-* Removed the ``unicode_pm`` feature which allowed toggling between
-  using ``'+/-'`` or ``'±'`` in value/uncertainty strings. Previously
-  ``unicode_pm`` defaulted to ``False`` so that ``'+/-'`` was the
-  default behavior. Now the default behavior is to use ``'±'`` and there
-  is no way to change to the old ``'+/-'`` behavior.
+* **[BREAKING]** Removed the ``unicode_pm`` feature which allowed
+  toggling between using ``'+/-'`` or ``'±'`` in value/uncertainty
+  strings. Previously ``unicode_pm`` defaulted to ``False`` so that
+  ``'+/-'`` was the default behavior. Now the default behavior is to use
+  ``'±'`` and there is no way to change to the old ``'+/-'`` behavior.
   [`#10 <https://github.com/jagerber48/sciform/discussions/10>`_]
 
 ----
