### 2023-11-24 a5f7085
add test to check that user mode strings agree with string values for backend mode enums
