### 2023-11-24 059df04
str, Enum
