### 2023-11-24 cfccb2d
user_options is a "private" attribute
