### 2023-11-24 3e93dae
Clean up options printing
