+++ tests/test_print.py
@@ -12,17 +12,17 @@ class TestPrint(unittest.TestCase):
             print_global_defaults()
         actual_printout = sout.getvalue()
         expected_printout = (
-            "{'exp_mode': <ExpMode.FIXEDPOINT: 'fixed_point'>,\n"
-            " 'exp_val': <class 'sciform.modes.AutoExpVal'>,\n"
-            " 'round_mode': <RoundMode.SIG_FIG: 'sig_fig'>,\n"
-            " 'ndigits': <class 'sciform.modes.AutoDigits'>,\n"
-            " 'upper_separator': <Separator.NONE: ''>,\n"
-            " 'decimal_separator': <Separator.POINT: '.'>,\n"
-            " 'lower_separator': <Separator.NONE: ''>,\n"
-            " 'sign_mode': <SignMode.NEGATIVE: '-'>,\n"
-            " 'fill_mode': <FillMode.SPACE: ' '>,\n"
+            "{'exp_mode': 'fixed_point',\n"
+            " 'exp_val': AutoExpVal,\n"
+            " 'round_mode': 'sig_fig',\n"
+            " 'ndigits': AutoDigits,\n"
+            " 'upper_separator': '',\n"
+            " 'decimal_separator': '.',\n"
+            " 'lower_separator': '',\n"
+            " 'sign_mode': '-',\n"
+            " 'fill_mode': ' ',\n"
             " 'top_dig_place': 0,\n"
-            " 'exp_format': <ExpFormat.STANDARD: 'standard'>,\n"
+            " 'exp_format': 'standard',\n"
             " 'extra_si_prefixes': {},\n"
             " 'extra_iec_prefixes': {},\n"
             " 'extra_parts_per_forms': {},\n"
@@ -45,17 +45,17 @@ class TestPrint(unittest.TestCase):
 
         actual_printout = sout.getvalue()
         expected_printout = (
-            "{'exp_mode': <ExpMode.FIXEDPOINT: 'fixed_point'>,\n"
-            " 'exp_val': <class 'sciform.modes.AutoExpVal'>,\n"
-            " 'round_mode': <RoundMode.SIG_FIG: 'sig_fig'>,\n"
-            " 'ndigits': <class 'sciform.modes.AutoDigits'>,\n"
-            " 'upper_separator': <Separator.NONE: ''>,\n"
-            " 'decimal_separator': <Separator.POINT: '.'>,\n"
-            " 'lower_separator': <Separator.NONE: ''>,\n"
-            " 'sign_mode': <SignMode.NEGATIVE: '-'>,\n"
-            " 'fill_mode': <FillMode.SPACE: ' '>,\n"
+            "{'exp_mode': 'fixed_point',\n"
+            " 'exp_val': AutoExpVal,\n"
+            " 'round_mode': 'sig_fig',\n"
+            " 'ndigits': AutoDigits,\n"
+            " 'upper_separator': '',\n"
+            " 'decimal_separator': '.',\n"
+            " 'lower_separator': '',\n"
+            " 'sign_mode': '-',\n"
+            " 'fill_mode': ' ',\n"
             " 'top_dig_place': 3,\n"
-            " 'exp_format': <ExpFormat.STANDARD: 'standard'>,\n"
+            " 'exp_format': 'standard',\n"
             " 'extra_si_prefixes': {},\n"
             " 'extra_iec_prefixes': {},\n"
             " 'extra_parts_per_forms': {},\n"
