### 2023-11-24 0522b7c
remove user options __str__ method until we have a complete interface setup for how the user will access this information.
