### 2023-11-24 0e47e0a
update usage.rst global defaults printing test
