+++ docs/source/usage.rst
@@ -127,17 +127,17 @@ package default settings):
 
 >>> from sciform import print_global_defaults
 >>> print_global_defaults()
-{'exp_mode': <ExpMode.FIXEDPOINT: 'fixed_point'>,
- 'exp_val': <class 'sciform.modes.AutoExpVal'>,
- 'round_mode': <RoundMode.SIG_FIG: 'sig_fig'>,
- 'ndigits': <class 'sciform.modes.AutoDigits'>,
- 'upper_separator': <Separator.NONE: ''>,
- 'decimal_separator': <Separator.POINT: '.'>,
- 'lower_separator': <Separator.NONE: ''>,
- 'sign_mode': <SignMode.NEGATIVE: '-'>,
- 'fill_mode': <FillMode.SPACE: ' '>,
+{'exp_mode': 'fixed_point',
+ 'exp_val': AutoExpVal,
+ 'round_mode': 'sig_fig',
+ 'ndigits': AutoDigits,
+ 'upper_separator': '',
+ 'decimal_separator': '.',
+ 'lower_separator': '',
+ 'sign_mode': '-',
+ 'fill_mode': ' ',
  'top_dig_place': 0,
- 'exp_format': <ExpFormat.STANDARD: 'standard'>,
+ 'exp_format': 'standard',
  'extra_si_prefixes': {},
  'extra_iec_prefixes': {},
  'extra_parts_per_forms': {},
@@ -164,17 +164,17 @@ unchanged.
 ...     ndigits=4,
 ...     decimal_separator=',')
 >>> print_global_defaults()
-{'exp_mode': <ExpMode.ENGINEERING_SHIFTED: 'engineering_shifted'>,
- 'exp_val': <class 'sciform.modes.AutoExpVal'>,
- 'round_mode': <RoundMode.SIG_FIG: 'sig_fig'>,
+{'exp_mode': 'engineering_shifted',
+ 'exp_val': AutoExpVal,
+ 'round_mode': 'sig_fig',
  'ndigits': 4,
- 'upper_separator': <Separator.NONE: ''>,
- 'decimal_separator': <Separator.COMMA: ','>,
- 'lower_separator': <Separator.NONE: ''>,
- 'sign_mode': <SignMode.NEGATIVE: '-'>,
- 'fill_mode': <FillMode.ZERO: '0'>,
+ 'upper_separator': '',
+ 'decimal_separator': ',',
+ 'lower_separator': '',
+ 'sign_mode': '-',
+ 'fill_mode': '0',
  'top_dig_place': 0,
- 'exp_format': <ExpFormat.STANDARD: 'standard'>,
+ 'exp_format': 'standard',
  'extra_si_prefixes': {},
  'extra_iec_prefixes': {},
  'extra_parts_per_forms': {},
