### 2023-11-24 b143199
test mode_str_to_enum failure
