+++ CHANGELOG.rst
@@ -17,6 +17,8 @@ Fixed
 
 * Fixed Changelog.
 
+----
+
 0.30.0 (2023-11-24)
 -------------------
 
