### 2023-11-25 #121 @Batalex
I apologize for sounding somewhat pedantic; I just love explaining things in detail.

What is a code formatter, and why do we need it?

Ultimately, the goals of a code formatter fall into two categories: (1) reducing the cognitive charge of reading code and (2) asserting that code diffs add value to the code base.

As you might know, the [PEP 8](https://peps.python.org/pep-0008/) style guide is prevalent throughout the Python ecosystem. Its specifications are not precise enough, so there are no ambiguities, or some rules are not as relevant as others. Hence, multiple implementations.
Nonetheless, using a code format based on PEP8 lowers the entry barrier for contributors since the code is easier to read.

We suggested `black` because we can address some of the caveats of PyCharm's integrated formatter.

First, make it easier for potential contributors to work on your project. Though PyCharm is quite popular, that means that VSCode, (neo)vim, helix, etc., users cannot easily format their code contributions with the format you use.
The good news is since PyCharm [2023.2](https://www.jetbrains.com/help/pycharm/2023.2/reformat-and-rearrange-code.html#format-python-code-with-black), you can use `black` as the default formatter.
Using a CLI-based tool means you can format the code (or check if the code is properly formatted) outside the IDE, i.e., in a GH action workflow. `flake8` checks are pretty limited regarding the format.
Finally, the `black` code format aims to produce smaller diffs.

`black` is pretty opinionated. One of the things that people do not like about `black` is that it uses double quotes for strings. You probably have to add a few exceptions to your linter configuration so that it plays nicely with `black`.
