### 2023-11-25 c81abdd
tentative ruff config
