### 2023-11-27 5928f49
ruff configuration
