+++ pyproject.toml
@@ -31,6 +31,11 @@ readme = {file = "README.rst"}
 
 [tool.ruff.lint]
 select = ['ALL']
+ignore = ['PLR2004']
 
 [tool.ruff.format]
 quote-style = "single"
+
+[tool.ruff.lint.flake8-quotes]
+inline-quotes = "single"
+multiline-quotes = "single"
