+++ src/sciform/__init__.py
@@ -1,18 +1,33 @@
 from sciform.formatter import Formatter
-from sciform.scinum import SciNum, SciNumUnc
-from sciform.modes import AutoExpVal, AutoDigits
 from sciform.global_configuration import (
-    print_global_defaults, set_global_defaults, reset_global_defaults,
-    global_add_c_prefix, global_add_small_si_prefixes, global_add_ppth_form,
-    global_reset_si_prefixes, global_reset_iec_prefixes,
-    global_reset_parts_per_forms, GlobalDefaultsContext
+    GlobalDefaultsContext,
+    global_add_c_prefix,
+    global_add_ppth_form,
+    global_add_small_si_prefixes,
+    global_reset_iec_prefixes,
+    global_reset_parts_per_forms,
+    global_reset_si_prefixes,
+    print_global_defaults,
+    reset_global_defaults,
+    set_global_defaults,
 )
+from sciform.modes import AutoDigits, AutoExpVal
+from sciform.scinum import SciNum, SciNumUnc
 
 __all__ = [
-    'Formatter', 'SciNum', 'SciNumUnc', 'AutoExpVal', 'AutoDigits',
-    'print_global_defaults', 'set_global_defaults', 'reset_global_defaults',
-    'global_add_c_prefix', 'global_add_small_si_prefixes',
-    'global_add_ppth_form', 'global_reset_si_prefixes',
-    'global_reset_iec_prefixes', 'global_reset_parts_per_forms',
-    'GlobalDefaultsContext'
+    'Formatter',
+    'GlobalDefaultsContext',
+    'global_add_c_prefix',
+    'global_add_ppth_form',
+    'global_add_small_si_prefixes',
+    'global_reset_iec_prefixes',
+    'global_reset_parts_per_forms',
+    'global_reset_si_prefixes',
+    'print_global_defaults',
+    'reset_global_defaults',
+    'set_global_defaults',
+    'AutoDigits',
+    'AutoExpVal',
+    'SciNum',
+    'SciNumUnc',
 ]
