### 2023-11-27 dd3ab1d
refactor on format_utils
