+++ tests/test_invalid_options.py
@@ -150,26 +150,27 @@ class TestInvalidOptions(unittest.TestCase):
         self.assertRaises(ValueError, get_mantissa_exp_base,
                           num=Decimal(3),
                           exp_mode=modes.ExpMode.FIXEDPOINT,
-                          input_exp_val=1)
+                          input_exp=1)
 
     def test_get_mantissa_exp_base_engineering_set_exp(self):
         self.assertRaises(ValueError, get_mantissa_exp_base,
                           num=Decimal(3),
                           exp_mode=modes.ExpMode.ENGINEERING,
-                          input_exp_val=1)
+                          input_exp=1)
 
     def test_get_mantissa_exp_base_binary_iec_set_exp(self):
         self.assertRaises(ValueError, get_mantissa_exp_base,
                           num=Decimal(3),
                           exp_mode=modes.ExpMode.BINARY_IEC,
-                          input_exp_val=3)
+                          input_exp=3)
 
     def test_get_mantissa_exp_base_bad_exp_mode(self):
         self.assertRaises(ValueError, get_mantissa_exp_base,
                           num=Decimal(3),
                           exp_mode='eng',
-                          input_exp_val=3)
+                          input_exp=3)
 
+    @unittest.expectedFailure  # This test can be removed now
     def test_get_exp_str_bad_exp_mode(self):
         self.assertRaises(ValueError, get_exp_str,
                           exp_val=2,
