+++ src/sciform/format_utils.py
@@ -1,127 +1,167 @@
-from typing import Union
-from math import floor, log10, log2
+from __future__ import annotations
+
 import re
 from decimal import Decimal
+from math import floor, log2
+from typing import Literal, Union, cast
 
-from sciform.modes import (ExpMode, ExpFormat, RoundMode, SignMode, AutoExpVal,
-                           AutoDigits)
-from sciform.prefix import (si_val_to_prefix_dict, iec_val_to_prefix_dict,
-                            pp_val_to_prefix_dict)
-
+from sciform.modes import (
+    AutoDigits,
+    AutoExpVal,
+    ExpFormat,
+    ExpMode,
+    RoundMode,
+    SignMode,
+)
+from sciform.prefix import (
+    iec_val_to_prefix_dict,
+    pp_val_to_prefix_dict,
+    si_val_to_prefix_dict,
+)
 
 Number = Union[Decimal, float, int, str]
 
 
-def get_top_digit(num: Decimal, binary=False) -> int:
-    if not num.is_finite():
+def get_top_digit(num: Decimal) -> int:
+    """Get the decimal place of a decimal's most significant digit."""
+    if not num.is_finite() or num == 0:
         return 0
-    if num == 0:
+    _, digits, exp = num.as_tuple()
+    return len(digits) + exp - 1
+
+
+def get_top_digit_binary(num: Decimal) -> int:
+    """Get the decimal place of a decimal's most significant digit."""
+    if not num.is_finite() or num == 0:
         return 0
-    if not binary:
-        return floor(log10(abs(num)))
-    else:
-        return floor(log2(abs(num)))
+    return floor(log2(abs(num)))
 
 
 def get_bottom_digit(num: Decimal) -> int:
+    """Get the decimal plac of a decimal's least significant digit."""
     if not num.is_finite():
         return 0
+    _, _, exp = num.as_tuple()
+    return exp
+
+
+def get_fixed_exp(
+    input_exp: int | type(AutoExpVal),
+) -> Literal[0]:
+    """Get the exponent for fixed or percent format modes."""
+    if input_exp is not AutoExpVal and input_exp != 0:
+        msg = 'Cannot set non-zero exponent in fixed point or percent exponent mode.'
+        raise ValueError(msg)
+    return 0
+
+
+def get_scientific_exp(
+    num: Decimal,
+    input_exp: int | type(AutoExpVal),
+) -> int:
+    """Get the exponent for scientific formatting mode."""
+    return get_top_digit(num) if input_exp is AutoExpVal else input_exp
+
+
+def get_engineering_exp(
+    num: Decimal,
+    input_exp: int | type(AutoExpVal),
+    *,
+    shifted: bool = False,
+) -> int:
+    """Get the exponent for engineering formatting modes."""
+    if input_exp is AutoExpVal:
+        exp_val = get_top_digit(num)
+        exp_val = exp_val // 3 * 3 if not shifted else (exp_val + 1) // 3 * 3
     else:
-        _, _, exp = num.as_tuple()
-        return exp
+        if input_exp % 3 != 0:
+            msg = (
+                f'Exponent must be an integer multiple of 3 in engineering modes, not '
+                f'{input_exp}.'
+            )
+            raise ValueError(msg)
+        exp_val = input_exp
+    return exp_val
+
+
+def get_binary_exp(
+    num: Decimal,
+    input_exp: int | type(AutoExpVal),
+    *,
+    iec: bool = False,
+) -> int:
+    """Get the exponent for binary formatting modes."""
+    if input_exp is AutoExpVal:
+        exp_val = get_top_digit_binary(num)
+        if iec:
+            exp_val = (exp_val // 10) * 10
+    else:
[…]
