+++ src/sciform/format_utils.py
@@ -52,7 +52,7 @@ def get_fixed_exp(
 ) -> Literal[0]:
     """Get the exponent for fixed or percent format modes."""
     if input_exp is not AutoExpVal and input_exp != 0:
-        msg = 'Cannot set non-zero exponent in fixed point or percent exponent mode.'
+        msg = "Cannot set non-zero exponent in fixed point or percent exponent mode."
         raise ValueError(msg)
     return 0
 
@@ -78,8 +78,8 @@ def get_engineering_exp(
     else:
         if input_exp % 3 != 0:
             msg = (
-                f'Exponent must be an integer multiple of 3 in engineering modes, not '
-                f'{input_exp}.'
+                f"Exponent must be an integer multiple of 3 in engineering modes, not "
+                f"{input_exp}."
             )
             raise ValueError(msg)
         exp_val = input_exp
@@ -100,8 +100,8 @@ def get_binary_exp(
     else:
         if iec and input_exp % 10 != 0:
             msg = (
-                f'Exponent must be an integer multiple of 10 in binary IEC mode, not '
-                f'{input_exp}.'
+                f"Exponent must be an integer multiple of 10 in binary IEC mode, not "
+                f"{input_exp}."
             )
             raise ValueError(msg)
         exp_val = input_exp
@@ -133,7 +133,7 @@ def get_mantissa_exp_base(
         elif exp_mode is ExpMode.BINARY_IEC:
             exp = get_binary_exp(num, input_exp, iec=True)
         else:
-            msg = f'Unhandled exponent mode {exp_mode}.'
+            msg = f"Unhandled exponent mode {exp_mode}."
             raise ValueError(msg)
         mantissa = num * Decimal(base) ** Decimal(-exp)
     mantissa = mantissa.normalize()
@@ -142,18 +142,18 @@ def get_mantissa_exp_base(
 
 def get_standard_exp_str(base: int, exp_val: int, *, capitalize: bool = False) -> str:
     """Get standard (eg. 'e+02') exponent string."""
-    base_exp_symbol_dict = {10: 'e', 2: 'b'}
+    base_exp_symbol_dict = {10: "e", 2: "b"}
     exp_symbol = base_exp_symbol_dict[base]
     if capitalize:
         exp_symbol = exp_symbol.capitalize()
-    return f'{exp_symbol}{exp_val:+03d}'
+    return f"{exp_symbol}{exp_val:+03d}"
 
 
 def get_superscript_exp_str(base: int, exp_val: int) -> str:
     """Get superscript (e.g. '×10⁺²') exponent string."""  # noqa: RUF002
-    sup_trans = str.maketrans('+-0123456789', '⁺⁻⁰¹²³⁴⁵⁶⁷⁸⁹')
-    exp_val_str = f'{exp_val}'.translate(sup_trans)
-    return f'×{base}{exp_val_str}'  # noqa: RUF001
+    sup_trans = str.maketrans("+-0123456789", "⁺⁻⁰¹²³⁴⁵⁶⁷⁸⁹")
+    exp_val_str = f"{exp_val}".translate(sup_trans)
+    return f"×{base}{exp_val_str}"  # noqa: RUF001
 
 
 def get_prefix_dict(
@@ -172,13 +172,13 @@ def get_prefix_dict(
             prefix_dict = iec_val_to_prefix_dict.copy()
             prefix_dict.update(extra_iec_prefixes)
         else:
-            msg = f'Unhandled base {base}'
+            msg = f"Unhandled base {base}"
             raise ValueError(msg)
     elif exp_format is ExpFormat.PARTS_PER:
         prefix_dict = pp_val_to_prefix_dict.copy()
         prefix_dict.update(extra_parts_per_forms)
     else:
[…]
