### 2023-11-27 6819287
Updates
- some changes to ruff configuration
- formatter.py into compliance
- Update formatter.py documentation
