+++ src/sciform/formatter.py
@@ -1,17 +1,24 @@
+"""Main Formatter class."""
+
+from __future__ import annotations
+
 from decimal import Decimal
-from typing import Optional, Union
+from typing import TYPE_CHECKING
 
-from sciform import modes
-from sciform.user_options import UserOptions
 from sciform.formatting import format_num, format_val_unc
-from sciform.format_utils import Number
+from sciform.user_options import UserOptions
 
+if TYPE_CHECKING:
+    from sciform import modes
+    from sciform.format_utils import Number
 
 # TODO: Support SciNum/SciNumUnc input to formatter?
 
 
 class Formatter:
     """
+    Class to format numbers and pairs of numbers into strings.
+
     :class:`Formatter` is used to convert numbers and pairs of numbers
     into formatted strings according to a variety of formatting options.
     See :ref:`formatting_options` for more details on the available
@@ -35,154 +42,164 @@ class Formatter:
     ...             ndigits=2)
     >>> print(sform(12345.678, 3.4))
     (12.3457 ± 0.0034)e+03
-
-    The following checks are performed when creating a new
-    :class:`Formatter` object:
-
-    * ``ndigits`` >= 1 for significant figure rounding mode
-    * ``exp_val`` must be consistent with the exponent mode. If
-      ``exp_val`` is specified (i.e. not ``None``) and ``exp_val`` is
-      not ``AutoExpVal`` then
-
-      * ``exp_val`` must be 0 for fixed point and percent modes
-      * ``exp_val`` must be a multiple of 3 for engineering and shifted
-        engineering modes
-      * ``exp_val`` must be a multiple of 10 for binary iec mode
-
-    * ``upper_separator`` may be any of ``['', ',', '.', ' ', '_']`` but
-      must be different from ``decimal_separator``
-    * ``decimal_separator`` may be any of ``['.', ',']``
-    * ``lower_separator`` may be any of ``['', ' ', '_']``
-    * if ``pdg_sig_figs=True`` then ``ndigits=None`` or
-      ``ndigits=AutoDigits``.
-
-    :param exp_mode: Specify the formatting mode.
-    :type exp_mode: ``Literal['fixed_point', 'percent', 'scientific',
-      'engineering', 'engineering_shifted', 'binary', 'binary_iec']``
-    :param exp_val: Indicates how the exponent value should be chosen.
-      If an integer is specified, the value must be 0 for fixed point
-      and percent modes, an integer multiple of 3 for engineering and
-      engineering shifted modes, and an integer multiple of 10 for
-      binary IEC mode.
-    :type exp_val: ``Union[int, type[AutoExpVal]]``
-    :param round_mode: Indicate how to round numbers during formatting.
-    :type round_mode: ``Literal['sig_fig', 'dec_place']``
-    :param ndigits: Indicate how the many significant digits or the
-      decimal place to use for rounding. Must be >= 1 for significant
-      figure rounding. Can be any integer for decimal place rounding.
-    :type ndigits: ``Union[int, type[AutoDigits]]``
-    :param upper_separator: Separator character to be used to group
-      digits above the decimal symbol.
-    :type upper_separator: ``Literal['', ',', '.', ' ', '_']``
-    :param decimal_separator: Separator character to be used as the
-      decimal symbol. Note that ``decimal_separator`` cannot be the same
-      as ``upper_separator``
-    :type decimal_separator: ``Literal['.', ',']``
-    :param lower_separator: Separator character to be used to group
-      digits below the decimal symbol.
-    :type lower_separator: ``Literal['', ' ', '_']``
-    :param sign_mode: :class:`SignMode` indicating sign symbol behavior.
-    :type sign_mode: ``Literal['-', '+', ' ']``
[…]
