+++ pyproject.toml
@@ -35,6 +35,7 @@ ignore = [
     'D212',
     'D211',
     'PLR2004',
+    'SIM108',
     'TD002',
     'TD003',
     'FIX002',
