### 2023-11-30 aec20ec
refactor formatting.py, including moving lots of code out into format_utils
