+++ src/sciform/formatting.py
@@ -1,26 +1,39 @@
-from dataclasses import replace
-from warnings import warn
-import re
-from decimal import Decimal
+"""Main formatting functions."""
 
-from sciform.modes import ExpMode, SignMode, AutoExpVal, RoundMode, ExpFormat
-from sciform.rendered_options import RenderedOptions
+from dataclasses import replace
+from decimal import Decimal
+from typing import cast
+from warnings import warn
+
 from sciform.format_utils import (
-    get_mantissa_exp_base, get_exp_str, get_top_digit, get_round_digit,
-    format_num_by_top_bottom_dig, latex_translate, parse_standard_exp_str
+    construct_val_unc_exp_str,
+    construct_val_unc_str,
+    format_num_by_top_bottom_dig,
+    get_exp_str,
+    get_mantissa_exp_base,
+    get_round_digit,
+    get_val_unc_exp,
+    get_val_unc_mantissa_exp_strs,
+    get_val_unc_top_digit,
+    latex_translate,
+    round_val_unc,
 )
 from sciform.grouping import add_separators
+from sciform.modes import AutoExpVal, ExpFormat, ExpMode, RoundMode, SignMode
+from sciform.rendered_options import RenderedOptions
 
 
 def format_non_finite(num: Decimal, options: RenderedOptions) -> str:
+    """Format non-finite numbers."""
     if num.is_nan():
-        num_str = 'nan'
-    elif num == Decimal('inf'):
-        num_str = 'inf'
-    elif num == Decimal('-inf'):
-        num_str = '-inf'
+        num_str = "nan"
+    elif num == Decimal("inf"):
+        num_str = "inf"
+    elif num == Decimal("-inf"):
+        num_str = "-inf"
     else:
-        raise ValueError(f'format_non_finite() cannot format {num}.')
+        msg = f"format_non_finite() cannot format {num}."
+        raise ValueError(msg)
 
     if options.nan_inf_exp:
         exp_mode = options.exp_mode
@@ -39,15 +52,15 @@ def format_non_finite(num: Decimal, options: RenderedOptions) -> str:
             superscript=options.superscript_exp,
             extra_si_prefixes=options.extra_si_prefixes,
             extra_iec_prefixes=options.extra_iec_prefixes,
-            extra_parts_per_forms=options.extra_parts_per_forms
+            extra_parts_per_forms=options.extra_parts_per_forms,
         )
     else:
-        exp_str = ''
+        exp_str = ""
 
-    if exp_str != '':
-        result = f'({num_str}){exp_str}'
+    if exp_str != "":
+        result = f"({num_str}){exp_str}"
     else:
-        result = f'{num_str}'
+        result = f"{num_str}"
 
     if options.capitalize:
         result = result.upper()
@@ -61,6 +74,7 @@ def format_non_finite(num: Decimal, options: RenderedOptions) -> str:
 
 
 def format_num(num: Decimal, options: RenderedOptions) -> str:
+    """Format a single number according to input options."""
     if not num.is_finite():
         return format_non_finite(num, options)
 
@@ -77,23 +91,24 @@ def format_num(num: Decimal, options: RenderedOptions) -> str:
     round_digit = get_round_digit(mantissa, round_mode, ndigits)
     mantissa_rounded = round(mantissa, -round_digit)
 
-    '''
+    """
     Repeat mantissa + exponent discovery after rounding in case rounding
     altered the required exponent.
-    '''
+    """
     rounded_num = mantissa_rounded * Decimal(base)**Decimal(temp_exp_val)
     mantissa, exp_val, base = get_mantissa_exp_base(rounded_num, exp_mode,
                                                     exp_val)
     round_digit = get_round_digit(mantissa, round_mode, ndigits)
     mantissa_rounded = round(mantissa, -int(round_digit))
[…]
