+++ src/sciform/format_utils.py
@@ -10,9 +10,11 @@ from typing import Literal, Union, cast
 from sciform.modes import (
     AutoDigits,
     AutoExpVal,
+    ExpDriver,
     ExpFormat,
     ExpMode,
     RoundMode,
+    Separator,
     SignMode,
 )
 from sciform.prefix import (
@@ -375,3 +377,195 @@ def latex_translate(input_str: str) -> str:
     for old_chars, new_chars in replacements:
         result_str = result_str.replace(old_chars, new_chars)
     return result_str
+
+
+def round_val_unc(
+        val: Decimal,
+        unc: Decimal,
+        ndigits: int | type[AutoDigits],
+        *,
+        use_pdg_sig_figs: bool = False,
+) -> tuple[Decimal, Decimal, int]:
+    """Simultaneously round the value and uncertainty."""
+    if unc.is_finite() and unc != 0:
+        round_digit = get_round_digit(
+            unc,
+            RoundMode.SIG_FIG,
+            ndigits,
+            pdg_sig_figs=use_pdg_sig_figs,
+        )
+        unc_rounded = round(unc, -round_digit)
+    else:
+        round_digit = get_round_digit(
+            val,
+            RoundMode.SIG_FIG,
+            ndigits,
+            pdg_sig_figs=False,
+        )
+        unc_rounded = unc
+    if val.is_finite():
+        val_rounded = round(val, -round_digit)
+    else:
+        val_rounded = val
+    return val_rounded, unc_rounded, round_digit
+
+
+def get_val_unc_exp(
+        val: Decimal,
+        unc: Decimal,
+        exp_mode: ExpMode,
+        input_exp: int,
+) -> tuple[int, ExpDriver]:
+    """Get exponent for value/uncertainty formatting."""
+    if val.is_finite() and unc.is_finite():
+        if abs(val) >= unc:
+            exp_driver_type = ExpDriver.VAL
+        else:
+            exp_driver_type = ExpDriver.UNC
+    elif val.is_finite():
+        exp_driver_type = ExpDriver.VAL
+    else:
+        exp_driver_type = ExpDriver.UNC
+
+    if exp_driver_type is ExpDriver.VAL:
+        exp_driver_val = val
+    elif exp_driver_type is ExpDriver.UNC:
+        exp_driver_val = unc
+    else:
+        raise ValueError
+
+    _, exp_val, _ = get_mantissa_exp_base(
+        exp_driver_val,
+        exp_mode=exp_mode,
+        input_exp=input_exp,
+    )
+
+    return exp_val, exp_driver_type
+
+
+def get_val_unc_top_digit(
+        val_mantissa: Decimal,
+        unc_mantissa: Decimal,
+        input_top_digit: int | AutoDigits,
+        *,
+        val_unc_match_widths: bool,
+) -> int | AutoDigits:
+    """Get top digit place for value/uncertainty formatting."""
+    if val_unc_match_widths:
+        val_top_digit = get_top_digit(val_mantissa)
+        unc_top_digit = get_top_digit(unc_mantissa)
+        new_top_digit = max(
+            input_top_digit,
+            val_top_digit,
+            unc_top_digit,
+        )
+    else:
+        new_top_digit = input_top_digit
+    return new_top_digit
+
+
+def get_val_unc_mantissa_exp_strs(
+        val_mantissa_exp_str: str,
+        unc_mantissa_exp_str: str,
+        exp_driver_type: ExpDriver,
+) -> tuple[str, str, str]:
+    """Break val/unc mantissa/exp strings into mantissa strings and an exp string."""
+    # Optional parentheses needed to handle (nan)e+00 case
[…]
