### 2023-11-30 4c2fdc7
reformat format_utils and formatting.py
