### 2023-11-30 dd3bc27
fix bracket_unc_remove_seps bug
