+++ src/sciform/format_utils.py
@@ -519,8 +519,10 @@ def construct_val_unc_str(  # noqa: PLR0913
             for separator in Separator:
                 if separator == decimal_separator:
                     continue
-                unc_mantissa_str.replace(separator, "")
+                unc_mantissa_str = unc_mantissa_str.replace(separator, "")
             if unc_mantissa < abs(val_mantissa):
+                # TODO: I think this raises an error if bracket_unc=True but either
+                #   unc_mantissa or val_mantissa is non-finite.
                 # Only removed "embedded" decimal symbol for unc < val
                 unc_mantissa_str = unc_mantissa_str.replace(
                     decimal_separator,
