+++ src/sciform/fsml.py
@@ -1,9 +1,12 @@
+"""Parse Format Specification Mini Language."""
+
+from __future__ import annotations
+
 import re
 
 from sciform.user_options import UserOptions
 
-
-pattern = re.compile(r'''^
+pattern = re.compile(r"""^
                          (?:(?P<fill_mode>[ 0])=)?
                          (?P<sign_mode>[-+ ])?
                          (?P<alternate_mode>\#)?
@@ -16,21 +19,52 @@ pattern = re.compile(r'''^
                          (?:x(?P<exp_val>[+-]?\d+))?
                          (?P<prefix_mode>p)?
                          (?P<bracket_unc>\(\))?
-                         $''', re.VERBOSE)
+                         $""", re.VERBOSE)
+
+
+def parse_exp_mode(
+        exp_mode: str | None,
+        *,
+        alternate_mode: bool,
+) -> tuple[str | None, bool]:
+    """Pase the exp_mode."""
+    if exp_mode is not None:
+        capitalize = exp_mode.isupper()
+        if exp_mode in ["f", "F"]:
+            exp_mode = "fixed_point"
+        elif exp_mode == "%":
+            exp_mode = "percent"
+        elif exp_mode in ["e", "E"]:
+            exp_mode = "scientific"
+        elif exp_mode in ["r", "R"]:
+            if alternate_mode:
+                exp_mode = "engineering_shifted"
+            else:
+                exp_mode = "engineering"
+        elif exp_mode in ["b", "B"]:
+            if alternate_mode:
+                exp_mode = "binary_iec"
+            else:
+                exp_mode = "binary"
+    else:
+        capitalize = None
+    return exp_mode, capitalize
 
 
 def format_options_from_fmt_spec(fmt_spec: str) -> UserOptions:
+    """Resolve UserOptions form format specification string."""
     match = pattern.match(fmt_spec)
     if match is None:
-        raise ValueError(f'Invalid format specifier: \'{fmt_spec}\'')
+        msg = f"Invalid format specifier: \'{fmt_spec}\'"
+        raise ValueError(msg)
 
-    fill_mode = match.group('fill_mode')
-    sign_mode = match.group('sign_mode')
+    fill_mode = match.group("fill_mode")
+    sign_mode = match.group("sign_mode")
 
-    alternate_mode = match.group('alternate_mode')
+    alternate_mode = match.group("alternate_mode")
     alternate_mode = alternate_mode is not None
 
-    top_dig_place = match.group('top_dig_place')
+    top_dig_place = match.group("top_dig_place")
     if top_dig_place is not None:
         top_dig_place = int(top_dig_place)
         val_unc_match_widths = True
@@ -38,64 +72,47 @@ def format_options_from_fmt_spec(fmt_spec: str) -> UserOptions:
         top_dig_place = None
         val_unc_match_widths = None
 
-    upper_separator = match.group('upper_separator')
+    upper_separator = match.group("upper_separator")
     if upper_separator is not None:
-        upper_separator = upper_separator.replace('n', '')
-        upper_separator = upper_separator.replace('s', ' ')
+        upper_separator = upper_separator.replace("n", "")
+        upper_separator = upper_separator.replace("s", " ")
 
-    decimal_separator = match.group('decimal_separator')
+    decimal_separator = match.group("decimal_separator")
 
-    lower_separator = match.group('lower_separator')
+    lower_separator = match.group("lower_separator")
     if lower_separator is not None:
-        lower_separator = lower_separator.replace('n', '')
-        lower_separator = lower_separator.replace('s', ' ')
+        lower_separator = lower_separator.replace("n", "")
[…]
