### 2023-11-30 516ba2b
reformat global_configuration.py
