+++ src/sciform/global_configuration.py
@@ -1,47 +1,56 @@
-from typing import Optional, Union
+"""Global configuration functions and context manager."""
+
+from __future__ import annotations
+
+from typing import TYPE_CHECKING
+
 from sciform import global_options, modes
 from sciform.user_options import UserOptions
-from sciform.rendered_options import RenderedOptions
+
+if TYPE_CHECKING:
+    from types import TracebackType
+
+    from sciform.rendered_options import RenderedOptions
 
 
-def print_global_defaults():
+def print_global_defaults() -> None:
+    """Print current global default formatting options as a dictionary."""
+    print(str(global_options.GLOBAL_DEFAULT_OPTIONS))  # noqa: T201
+
+
+def set_global_defaults(  # noqa: PLR0913
+    *,
+    exp_mode: modes.UserExpMode | None = None,
+    exp_val: int | type(modes.AutoExpVal) | None = None,
+    round_mode: modes.UserRoundMode | None = None,
+    ndigits: int | type(modes.AutoDigits) | None = None,
+    upper_separator: modes.UserUpperSeparators | None = None,
+    decimal_separator: modes.UserDecimalSeparators | None = None,
+    lower_separator: modes.UserLowerSeparators | None = None,
+    sign_mode: modes.UserSignMode | None = None,
+    fill_mode: modes.UserFillMode | None = None,
+    top_dig_place: int | None = None,
+    exp_format: modes.UserExpFormat | None = None,
+    extra_si_prefixes: dict[int, str] | None = None,
+    extra_iec_prefixes: dict[int, str] | None = None,
+    extra_parts_per_forms: dict[int, str] | None = None,
+    capitalize: bool | None = None,
+    superscript_exp: bool | None = None,
+    latex: bool | None = None,
+    nan_inf_exp: bool | None = None,
+    bracket_unc: bool | None = None,
+    pdg_sig_figs: bool | None = None,
+    val_unc_match_widths: bool | None = None,
+    bracket_unc_remove_seps: bool | None = None,
+    unc_pm_whitespace: bool | None = None,
+    add_c_prefix: bool = False,
+    add_small_si_prefixes: bool = False,
+    add_ppth_form: bool = False,
+) -> None:
     """
-    Print current global default formatting options as a dictionary.
-    """
-    print(str(global_options.GLOBAL_DEFAULT_OPTIONS))
+    Configure global default format options.
 
-
-def set_global_defaults(
-            exp_mode: Optional[modes.UserExpMode] = None,
-            exp_val: Optional[Union[int, type(modes.AutoExpVal)]] = None,
-            round_mode: Optional[modes.UserRoundMode] = None,
-            ndigits: Optional[Union[int, type(modes.AutoDigits)]] = None,
-            upper_separator: Optional[modes.UserUpperSeparators] = None,
-            decimal_separator: Optional[modes.UserDecimalSeparators] = None,
-            lower_separator: Optional[modes.UserLowerSeparators] = None,
-            sign_mode: Optional[modes.UserSignMode] = None,
-            fill_mode: Optional[modes.UserFillMode] = None,
-            top_dig_place: Optional[int] = None,
-            exp_format: Optional[modes.UserExpFormat] = None,
-            extra_si_prefixes: Optional[dict[int, str]] = None,
-            extra_iec_prefixes: Optional[dict[int, str]] = None,
-            extra_parts_per_forms: Optional[dict[int, str]] = None,
-            capitalize: Optional[bool] = None,
-            superscript_exp: Optional[bool] = None,
-            latex: Optional[bool] = None,
-            nan_inf_exp: Optional[bool] = None,
-            bracket_unc: Optional[bool] = None,
-            pdg_sig_figs: Optional[bool] = None,
[…]
