### 2023-12-05 bfdc078
refactor and reformat grouping.py
simplify algorithm using regex instead of "split"
