### 2023-12-05 f184c10
lint on modes.py
