+++ src/sciform/modes.py
@@ -1,16 +1,21 @@
-from typing import Literal, TypeVar
+"""Enums and Literals for the various formatting options."""
+
+from __future__ import annotations
+
 from enum import Enum
+from typing import Literal, TypeVar
 
 
 class SentinelMeta(type):
-    def __repr__(self):
-        return self.__name__
+    """Sentinel metaclass, __repr__ returns class name."""
+
+    def __repr__(cls: SentinelMeta) -> str:  # noqa: D105
+        return cls.__name__
 
 
 class AutoExpVal(metaclass=SentinelMeta):
     """
-    Flag for auto-exponent calculation mode. Set ``exp_val=AutoExpVal`` or an
-    integer.
+    Flag for auto-exponent calculation mode.
 
       * For scientific exponent mode the base-10 exponent is selected so
         that the mantissa ``m`` satisfies ``1 <= m < 10``.
@@ -25,13 +30,11 @@ class AutoExpVal(metaclass=SentinelMeta):
       * For binary IEC exponent mode the base-2 exponent is chosen so
         that the mantissa ``m`` satisfies ``1 <= m < 1024 = 2**10``.
     """
-    pass
 
 
 class AutoDigits(metaclass=SentinelMeta):
     """
-    Flag for auto ndigits calculation mode. Set ``ndigits=AutoDigits``
-    or an integer.
+    Flag for auto ndigits calculation mode.
 
     In both sig fig and ndigits round modes this auto ndigits
     option chooses the ndigits so that the least significant digit of
@@ -45,37 +48,42 @@ class AutoDigits(metaclass=SentinelMeta):
     chosen to be one or two in accordance with the Particle Data Group
     algorithm.
     """
-    pass
 
 
-UserFillMode = Literal[' ', '0']
+UserFillMode = Literal[" ", "0"]
 
 
 class FillMode(str, Enum):
-    SPACE = ' '
-    ZERO = '0'
+    """Fill mode Enum."""
+
+    SPACE = " "
+    ZERO = "0"
 
 
-UserSignMode = Literal['-', '+', ' ']
+UserSignMode = Literal["-", "+", " "]
 
 
 class SignMode(str, Enum):
-    NEGATIVE = '-'
-    ALWAYS = '+'
-    SPACE = ' '
+    """Sign mode Enum."""
+
+    NEGATIVE = "-"
+    ALWAYS = "+"
+    SPACE = " "
 
 
-UserUpperSeparators = Literal['', ',', '.', ' ', '_']
-UserDecimalSeparators = Literal['.', ',']
-UserLowerSeparators = Literal['', ' ', '_']
+UserUpperSeparators = Literal["", ",", ".", " ", "_"]
+UserDecimalSeparators = Literal[".", ","]
+UserLowerSeparators = Literal["", " ", "_"]
 
 
 class Separator(str, Enum):
-    NONE = ''
-    COMMA = ','
-    POINT = '.'
-    UNDERSCORE = '_'
-    SPACE = ' '
+    """Separator type Enum."""
+
+    NONE = ""
+    COMMA = ","
+    POINT = "."
+    UNDERSCORE = "_"
+    SPACE = " "
 
 
 UpperSeparators = Literal[Separator.NONE,
@@ -90,47 +98,57 @@ LowerSeparators = Literal[Separator.NONE,
                           Separator.SPACE]
 
 
-UserRoundMode = Literal['sig_fig', 'dec_place']
+UserRoundMode = Literal["sig_fig", "dec_place"]
 
 
 class RoundMode(str, Enum):
-    SIG_FIG = 'sig_fig'
-    DEC_PLACE = 'dec_place'
+    """Round mode Enum."""
+
[…]
