+++ src/sciform/prefix.py
@@ -1,36 +1,44 @@
-si_val_to_prefix_dict = {30: 'Q',
-                         27: 'R',
-                         24: 'Y',
-                         21: 'Z',
-                         18: 'E',
-                         15: 'P',
-                         12: 'T',
-                         9: 'G',
-                         6: 'M',
-                         3: 'k',
-                         0: '',
-                         -3: 'm',
-                         -6: 'μ',
-                         -9: 'n',
-                         -12: 'p',
-                         -15: 'f',
-                         -18: 'a',
-                         -21: 'z',
-                         -24: 'y',
-                         -27: 'r',
-                         -30: 'q'}
+"""Exponent prefix translations."""
 
-iec_val_to_prefix_dict = {0: '',
-                          10: 'Ki',
-                          20: 'Mi',
-                          30: 'Gi',
-                          40: 'Ti',
-                          50: 'Pi',
-                          60: 'Ei',
-                          70: 'Zi',
-                          80: 'Yi'}
+si_val_to_prefix_dict = {
+    30: "Q",
+    27: "R",
+    24: "Y",
+    21: "Z",
+    18: "E",
+    15: "P",
+    12: "T",
+    9: "G",
+    6: "M",
+    3: "k",
+    0: "",
+    -3: "m",
+    -6: "μ",
+    -9: "n",
+    -12: "p",
+    -15: "f",
+    -18: "a",
+    -21: "z",
+    -24: "y",
+    -27: "r",
+    -30: "q",
+}
 
-pp_val_to_prefix_dict = {-6: 'ppm',
-                         -9: 'ppb',
-                         -12: 'ppt',
-                         -15: 'ppq'}
+iec_val_to_prefix_dict = {
+    0: "",
+    10: "Ki",
+    20: "Mi",
+    30: "Gi",
+    40: "Ti",
+    50: "Pi",
+    60: "Ei",
+    70: "Zi",
+    80: "Yi",
+}
+
+pp_val_to_prefix_dict = {
+    -6: "ppm",
+    -9: "ppb",
+    -12: "ppt",
+    -15: "ppq",
+}
