### 2023-12-05 4c56925
lint rendered_options.py
