### 2023-12-05 9c214f0
format scinum.py
