+++ src/sciform/scinum.py
@@ -1,12 +1,21 @@
+"""SciNum and SciNumUnc classes give users access to sciform FSML."""
+
+from __future__ import annotations
+
 from decimal import Decimal
+from typing import TYPE_CHECKING
 
 from sciform.formatting import format_num, format_val_unc
-from sciform.format_utils import Number
 from sciform.fsml import format_options_from_fmt_spec
+
+if TYPE_CHECKING:
+    from sciform.format_utils import Number
 
 
 class SciNum:
     """
+    Single number to be used with FSML.
+
     :class:`SciNum` objects are used in combination with the
     :mod:`sciform` format specification mini-language for scientific
     formatting of numbers. Any options not configured by the format
@@ -18,21 +27,24 @@ class SciNum:
     >>> print(f'{snum:,._.7f}')
     123,456.654_321_0
     """
-    def __init__(self, value: Number, /):
+
+    def __init__(self: SciNum, value: Number, /) -> None:
         self.value = Decimal(str(value))
 
-    def __format__(self, fmt: str):
+    def __format__(self: SciNum, fmt: str) -> str:
         user_options = format_options_from_fmt_spec(fmt)
         rendered_options = user_options.render()
         return format_num(self.value,
                           rendered_options)
 
-    def __repr__(self):
-        return f'{self.__class__.__name__}({self.value})'
+    def __repr__(self: SciNum) -> str:
+        return f"{self.__class__.__name__}({self.value})"
 
 
 class SciNumUnc:
     """
+    Value/uncertainty pair to be used with FSML.
+
     A :class:`SciNumUnc` objects stores a pair of numbers, a value and
     an uncertainty, for scientific formatting. This class is used in
     combination with the :mod:`sciform` format specification mini
@@ -45,17 +57,18 @@ class SciNumUnc:
     >>> print(f'{snumunc:,._!1f()}')
     123,456.654_321(2)
     """
-    def __init__(self, value: Number,
-                 uncertainty: Number, /):
+
+    def __init__(self: SciNum, value: Number,
+                 uncertainty: Number, /) -> None:
         self.value = Decimal(str(value))
         self.uncertainty = Decimal(str(uncertainty))
 
-    def __format__(self, fmt: str):
+    def __format__(self: SciNum, fmt: str) -> str:
         user_options = format_options_from_fmt_spec(fmt)
         rendered_options = user_options.render()
         return format_val_unc(self.value,
                               self.uncertainty,
                               rendered_options)
 
-    def __repr__(self):
-        return f'{self.__class__.__name__}({self.value}, {self.uncertainty})'
+    def __repr__(self: SciNum) -> str:
+        return f"{self.__class__.__name__}({self.value}, {self.uncertainty})"
