+++ src/sciform/scinum.py
@@ -58,17 +58,17 @@ class SciNumUnc:
     123,456.654_321(2)
     """
 
-    def __init__(self: SciNum, value: Number,
+    def __init__(self: SciNumUnc, value: Number,
                  uncertainty: Number, /) -> None:
         self.value = Decimal(str(value))
         self.uncertainty = Decimal(str(uncertainty))
 
-    def __format__(self: SciNum, fmt: str) -> str:
+    def __format__(self: SciNumUnc, fmt: str) -> str:
         user_options = format_options_from_fmt_spec(fmt)
         rendered_options = user_options.render()
         return format_val_unc(self.value,
                               self.uncertainty,
                               rendered_options)
 
-    def __repr__(self: SciNum) -> str:
+    def __repr__(self: SciNumUnc) -> str:
         return f"{self.__class__.__name__}({self.value}, {self.uncertainty})"
