+++ src/sciform/user_options.py
@@ -1,146 +1,196 @@
-from typing import Optional, Union, get_args
-from dataclasses import dataclass, InitVar, asdict
-from pprint import pformat
+"""UserOptions Dataclass which stores user input."""
 
-from sciform import modes
-from sciform import global_options
+
+from __future__ import annotations
+
+from dataclasses import InitVar, asdict, dataclass
+from typing import get_args
+
+from sciform import global_options, modes
 from sciform.rendered_options import RenderedOptions
 
 
 @dataclass(frozen=True)
 class UserOptions:
-    exp_mode: Optional[modes.UserExpMode] = None
-    exp_val: Optional[Union[int, type(modes.AutoExpVal)]] = None
-    round_mode: Optional[modes.UserRoundMode] = None
-    ndigits: Optional[Union[int, type(modes.AutoDigits)]] = None
-    upper_separator: Optional[modes.UserUpperSeparators] = None
-    decimal_separator: Optional[modes.UserDecimalSeparators] = None
-    lower_separator: Optional[modes.UserLowerSeparators] = None
-    sign_mode: Optional[modes.UserSignMode] = None
-    fill_mode: Optional[modes.UserFillMode] = None
-    top_dig_place: Optional[int] = None
-    exp_format: Optional[modes.UserExpFormat] = None
-    extra_si_prefixes: Optional[dict[int, str]] = None
-    extra_iec_prefixes: Optional[dict[int, str]] = None
-    extra_parts_per_forms: Optional[dict[int, str]] = None
-    capitalize: Optional[bool] = None
-    superscript_exp: Optional[bool] = None
-    latex: Optional[bool] = None
-    nan_inf_exp: Optional[bool] = None
-    bracket_unc: Optional[bool] = None
-    pdg_sig_figs: Optional[bool] = None
-    val_unc_match_widths: Optional[bool] = None
-    bracket_unc_remove_seps: Optional[bool] = None
-    unc_pm_whitespace: Optional[bool] = None
+    """Dataclass storing user input."""
+
+    exp_mode: modes.UserExpMode | None = None
+    exp_val: int | type(modes.AutoExpVal) | None = None
+    round_mode: modes.UserRoundMode | None = None
+    ndigits: int | type(modes.AutoDigits) | None = None
+    upper_separator: modes.UserUpperSeparators | None = None
+    decimal_separator: modes.UserDecimalSeparators | None = None
+    lower_separator: modes.UserLowerSeparators | None = None
+    sign_mode: modes.UserSignMode | None = None
+    fill_mode: modes.UserFillMode | None = None
+    top_dig_place: int | None = None
+    exp_format: modes.UserExpFormat | None = None
+    extra_si_prefixes: dict[int, str] | None = None
+    extra_iec_prefixes: dict[int, str] | None = None
+    extra_parts_per_forms: dict[int, str] | None = None
+    capitalize: bool | None = None
+    superscript_exp: bool | None = None
+    latex: bool | None = None
+    nan_inf_exp: bool | None = None
+    bracket_unc: bool | None = None
+    pdg_sig_figs: bool | None = None
+    val_unc_match_widths: bool | None = None
+    bracket_unc_remove_seps: bool | None = None
+    unc_pm_whitespace: bool | None = None
 
     add_c_prefix: InitVar[bool] = False
     add_small_si_prefixes: InitVar[bool] = False
     add_ppth_form: InitVar[bool] = False
 
-    def __post_init__(self, add_c_prefix, add_small_si_prefixes,
-                      add_ppth_form):
-        self.populate_dicts(add_c_prefix, add_small_si_prefixes,
-                            add_ppth_form)
+    def __post_init__(
+            self: UserOptions,
[…]
