### 2023-12-08 #121 @jagerber48
@Batalex thanks for the ping. So third party linter/formatting is something I've been needing in my life. I'm currently working on a PR getting `sciform` in line with the `ruff` formatter: https://github.com/jagerber48/sciform/pull/88. Hopefully I'll have it merged in a few days.

I think with the completion of this PR I will have checked off almost everything on the list except for opinions on naming convention. I do think I still may be seeking stronger opinions on naming conventions so I'm open to stronger opinions there and also suggestions for how to get finalized on the names (one suggestion can be: "don't worry about it so much, just go with the names here"). I'm worried because I'm a bit of a perfectionist and I often do things one way thinking it's good, but then a few months later realize there was a better way. If I'm in a situation with `sciform` where I feel like I can't change the names because users are relying on the stable api, then I'll be frustrated.
