+++ tests/test_val_unc_fsml.py
@@ -1,12 +1,11 @@
 import unittest
 
-from sciform import SciNumUnc, GlobalDefaultsContext
-
+from sciform import GlobalDefaultsContext, SciNumUnc
 
 ValUncFSMLCases = list[tuple[tuple[float, float],
                              list[tuple[str, str]]]]
-NAN = float('nan')
-INF = float('inf')
+NAN = float("nan")
+INF = float("inf")
 
 
 class TestFormatting(unittest.TestCase):
@@ -14,7 +13,7 @@ class TestFormatting(unittest.TestCase):
         for (val, unc), formats_list in cases_list:
             for format_spec, expected_str in formats_list:
                 snum = SciNumUnc(val, unc)
-                snum_str = f'{snum:{format_spec}}'
+                snum_str = f"{snum:{format_spec}}"
                 with self.subTest(val=val, unc=unc, format_spec=format_spec,
                                   expected_str=expected_str,
                                   actual_str=snum_str):
@@ -23,55 +22,55 @@ class TestFormatting(unittest.TestCase):
     def test_fixed(self):
         cases_list = [
             ((123.456, 0.789), [
-                ('f', '123.456 ± 0.789'),
-                ('!1f', '123.5 ± 0.8'),
-                ('!2f', '123.46 ± 0.79'),
-                ('!3f', '123.456 ± 0.789'),
-                ('!4f', '123.4560 ± 0.7890'),
-                ('!2f()', '123.46(79)')
+                ("f", "123.456 ± 0.789"),
+                ("!1f", "123.5 ± 0.8"),
+                ("!2f", "123.46 ± 0.79"),
+                ("!3f", "123.456 ± 0.789"),
+                ("!4f", "123.4560 ± 0.7890"),
+                ("!2f()", "123.46(79)"),
             ]),
             ((-123.456, 0.789), [
-                ('f', '-123.456 ± 0.789'),
-                ('!1f', '-123.5 ± 0.8'),
-                ('!2f', '-123.46 ± 0.79'),
-                ('!3f', '-123.456 ± 0.789'),
-                ('!4f', '-123.4560 ± 0.7890'),
-                ('!2f()', '-123.46(79)')
+                ("f", "-123.456 ± 0.789"),
+                ("!1f", "-123.5 ± 0.8"),
+                ("!2f", "-123.46 ± 0.79"),
+                ("!3f", "-123.456 ± 0.789"),
+                ("!4f", "-123.4560 ± 0.7890"),
+                ("!2f()", "-123.46(79)"),
             ]),
             ((0.789, 123.456), [
-                ('f', '0.789 ± 123.456'),
-                ('!1f', '0 ± 100'),
-                ('!2f', '0 ± 120'),
-                ('!3f', '1 ± 123'),
-                ('!4f', '0.8 ± 123.5'),
-                ('!5f', '0.79 ± 123.46'),
-                ('!6f', '0.789 ± 123.456'),
-                ('!7f', '0.7890 ± 123.4560'),
-                ('!1f()', '0(100)'),
-                ('!2f()', '0(120)'),
-                ('!3f()', '1(123)'),
-                ('!4f()', '0.8(123.5)'),
-                ('!5f()', '0.79(123.46)'),
-                ('!6f()', '0.789(123.456)'),
-                ('!7f()', '0.7890(123.4560)')
+                ("f", "0.789 ± 123.456"),
+                ("!1f", "0 ± 100"),
+                ("!2f", "0 ± 120"),
[…]
