+++ tests/test_invalid_options.py
@@ -1,101 +1,105 @@
 import unittest
 from decimal import Decimal
 
-from sciform import Formatter
+from sciform import Formatter, modes
+from sciform.format_utils import (
+    get_exp_str,
+    get_mantissa_exp_base,
+    get_prefix_dict,
+    get_round_digit,
+    get_sign_str,
+    get_top_digit,
+    parse_standard_exp_str,
+)
 from sciform.formatting import format_non_finite
 from sciform.user_options import UserOptions
-from sciform import modes
-from sciform.format_utils import (
-    get_top_digit, get_mantissa_exp_base, get_exp_str, get_sign_str,
-    get_round_digit, get_prefix_dict, parse_standard_exp_str
-)
 
 
 class TestInvalidOptions(unittest.TestCase):
     def test_sig_fig_ndigits(self):
         self.assertRaises(
-            ValueError, Formatter, round_mode='sig_fig',
-            ndigits=0
+            ValueError, Formatter, round_mode="sig_fig",
+            ndigits=0,
         )
 
     def test_pdg_sig_figs_ndigits(self):
         self.assertRaises(
-            ValueError, Formatter, pdg_sig_figs=True, ndigits=3
+            ValueError, Formatter, pdg_sig_figs=True, ndigits=3,
         )
 
     def test_fixed_point(self):
         self.assertRaises(
-            ValueError, Formatter, exp_mode='fixed_point',
-            exp_val=1
+            ValueError, Formatter, exp_mode="fixed_point",
+            exp_val=1,
         )
 
     def test_percent(self):
         self.assertRaises(
-            ValueError, Formatter, exp_mode='percent',
-            exp_val=1
+            ValueError, Formatter, exp_mode="percent",
+            exp_val=1,
         )
 
     def test_engineering(self):
         self.assertRaises(
-            ValueError, Formatter, exp_mode='engineering',
-            exp_val=1
+            ValueError, Formatter, exp_mode="engineering",
+            exp_val=1,
         )
 
     def test_engineering_shifted(self):
         self.assertRaises(
-            ValueError, Formatter, exp_mode='engineering',
-            exp_val=1
+            ValueError, Formatter, exp_mode="engineering",
+            exp_val=1,
         )
 
     def test_binary_iec(self):
         self.assertRaises(
-            ValueError, Formatter, exp_mode='binary_iec',
-            exp_val=5
+            ValueError, Formatter, exp_mode="binary_iec",
+            exp_val=5,
         )
 
     def test_upper_separator_non_option(self):
         self.assertRaises(
-            ValueError, Formatter, upper_separator='-'
+            ValueError, Formatter, upper_separator="-",
         )
 
     def test_decimal_separator_non_option(self):
         self.assertRaises(
-            ValueError, Formatter, decimal_separator='-'
+            ValueError, Formatter, decimal_separator="-",
         )
 
     def test_decimal_separator_underscore(self):
         self.assertRaises(
             ValueError, Formatter,
-            decimal_separator='_'
+            decimal_separator="_",
         )
 
     def test_decimal_separator_space(self):
         self.assertRaises(
             ValueError, Formatter,
-            decimal_separator=' '
+            decimal_separator=" ",
         )
 
     def test_decimal_separator_none(self):
         self.assertRaises(
             ValueError, Formatter,
-            decimal_separator=''
+            decimal_separator="",
         )
 
     def test_lower_separator_non_option(self):
         self.assertRaises(
-            ValueError, Formatter, lower_separator='-'
+            ValueError, Formatter, lower_separator="-",
         )
 
     def test_lower_separator_comma(self):
         self.assertRaises(
             ValueError, Formatter,
-            lower_separator=','
+            lower_separator=",",
         )
 
     def test_lower_separator_point(self):
[…]
