+++ pyproject.toml
@@ -36,6 +36,7 @@ ignore = [
     'D212',
     'D211',
     'PLR2004',
+    'RUF001',
     'SIM108',
     'TD002',
     'TD003',
