+++ tests/test_sci_num_and_sci_num_unc.py
@@ -9,5 +9,5 @@ class TestSciNumAndSciNumUnc(unittest.TestCase):
 
     def test_sci_num_unc_repr(self):
         self.assertEqual(
-            repr(SciNumUnc(123.456, 0.0000023)),
-            "SciNumUnc(123.456, 0.0000023)")
+            repr(SciNumUnc(123.456, 0.0000023)), "SciNumUnc(123.456, 0.0000023)"
+        )

[…]
