+++ tests/test_val_unc_formatter.py
@@ -2,245 +2,408 @@ import unittest
 
 from sciform import Formatter
 
-ValUncFormatterCases = list[tuple[tuple[float, float],
-                                  list[tuple[Formatter, str]]]]
+ValUncFormatterCases = list[tuple[tuple[float, float], list[tuple[Formatter, str]]]]
 
 
 class TestFormatting(unittest.TestCase):
-    def run_val_unc_formatter_cases(self,
-                                    cases_list: ValUncFormatterCases):
+    def run_val_unc_formatter_cases(self, cases_list: ValUncFormatterCases):
         for (val, unc), formats_list in cases_list:
             for formatter, expected_val_unc_str in formats_list:
                 snum_str = formatter(val, unc)
-                with self.subTest(val=val, unc=unc,
-                                  expected_num_str=expected_val_unc_str,
-                                  actual_num_str=snum_str):
+                with self.subTest(
+                    val=val,
+                    unc=unc,
+                    expected_num_str=expected_val_unc_str,
+                    actual_num_str=snum_str,
+                ):
                     self.assertEqual(snum_str, expected_val_unc_str)
 
     def test_bracket_unc(self):
         cases_list = [
-            ((123.456, 0.789), [
-                (Formatter(bracket_unc=True), "123.456(789)"),
-                (Formatter(exp_mode="scientific",
-                           bracket_unc=True), "(1.23456(789))e+02"),
-                (Formatter(exp_mode="engineering",
-                           bracket_unc=True), "(123.456(789))e+00"),
-                (Formatter(exp_mode="engineering_shifted",
-                           bracket_unc=True), "(0.123456(789))e+03"),
-                (Formatter(exp_mode="scientific",
-                           exp_val=+1,
-                           bracket_unc=True), "(12.3456(789))e+01"),
-                (Formatter(exp_mode="scientific",
-                           exp_val=-1,
-                           bracket_unc=True), "(1234.56(7.89))e-01"),
-            ]),
+            (
+                (123.456, 0.789),
+                [
+                    (Formatter(bracket_unc=True), "123.456(789)"),
+                    (
+                        Formatter(exp_mode="scientific", bracket_unc=True),
+                        "(1.23456(789))e+02",
+                    ),
+                    (
+                        Formatter(exp_mode="engineering", bracket_unc=True),
+                        "(123.456(789))e+00",
+                    ),
+                    (
+                        Formatter(exp_mode="engineering_shifted", bracket_unc=True),
+                        "(0.123456(789))e+03",
+                    ),
+                    (
+                        Formatter(exp_mode="scientific", exp_val=+1, bracket_unc=True),
+                        "(12.3456(789))e+01",
+                    ),
+                    (
+                        Formatter(exp_mode="scientific", exp_val=-1, bracket_unc=True),
+                        "(1234.56(7.89))e-01",
+                    ),
+                ],
+            ),
         ]
 
         self.run_val_unc_formatter_cases(cases_list)
 
     def test_bracket_unc_invalid_unc(self):
         cases_list = [
-            ((123, 0), [
-                (Formatter(bracket_unc=True), "123(0)"),
-            ]),
-            ((-123, 0), [
-                (Formatter(bracket_unc=True), "-123(0)"),
-            ]),
-            ((123, float("nan")), [
-                (Formatter(bracket_unc=True), "123(nan)"),
-            ]),
-            ((123, float("inf")), [
-                (Formatter(bracket_unc=True), "123(inf)"),
-            ]),
-            ((0, 0), [
-                (Formatter(bracket_unc=True), "0(0)"),
-            ]),
+            (
+                (123, 0),
+                [
+                    (Formatter(bracket_unc=True), "123(0)"),
+                ],
+            ),
[…]
