+++ tests/test_float_formatter.py
@@ -10,138 +10,232 @@ class TestFormatting(unittest.TestCase):
         for num, formats_list in cases_list:
             for formatter, expected_num_str in formats_list:
                 snum_str = formatter(num)
-                with self.subTest(num=num,
-                                  expected_num_str=expected_num_str,
-                                  actual_num_str=snum_str):
+                with self.subTest(
+                    num=num, expected_num_str=expected_num_str, actual_num_str=snum_str
+                ):
                     self.assertEqual(snum_str, expected_num_str)
 
     def test_superscript_exp(self):
         cases_list = [
-            (789, [
-                (Formatter(exp_mode="scientific",
-                           superscript_exp=True), "7.89×10²"),
-            ]),
-
+            (
+                789,
+                [
+                    (
+                        Formatter(exp_mode="scientific", superscript_exp=True),
+                        "7.89×10²",
+                    ),
+                ],
+            ),
             # Superscript in prefix mode when there's no replacement
-            (789, [
-                (Formatter(exp_mode="scientific",
-                           exp_format="prefix",
-                           superscript_exp=True), "7.89×10²"),
-            ]),
-            (1024, [
-                (Formatter(exp_mode="binary",
-                           superscript_exp=True), "1×2¹⁰"),
-            ]),
+            (
+                789,
+                [
+                    (
+                        Formatter(
+                            exp_mode="scientific",
+                            exp_format="prefix",
+                            superscript_exp=True,
+                        ),
+                        "7.89×10²",
+                    ),
+                ],
+            ),
+            (
+                1024,
+                [
+                    (Formatter(exp_mode="binary", superscript_exp=True), "1×2¹⁰"),
+                ],
+            ),
         ]
 
         self.run_float_formatter_cases(cases_list)
 
     def test_fill_and_separators(self):
         cases_list = [
-            (123456789.654321, [
-                (Formatter(
-                    upper_separator="_",
-                    lower_separator="_",
-                    fill_mode="0",
-                    top_dig_place=14), "000_000_123_456_789.654_321"),
-                (Formatter(
-                    upper_separator="_",
-                    lower_separator="_",
-                    fill_mode=" ",
-                    top_dig_place=14), "      123_456_789.654_321"),
-            ]),
-            (4567899.7654321, [
-                (Formatter(
-                    upper_separator="_",
-                    lower_separator="_",
-                    fill_mode="0",
-                    top_dig_place=14), "000_000_004_567_899.765_432_1"),
-                (Formatter(
-                    upper_separator="_",
-                    lower_separator="_",
-                    fill_mode=" ",
-                    top_dig_place=14), "        4_567_899.765_432_1"),
-            ]),
+            (
+                123456789.654321,
+                [
+                    (
+                        Formatter(
+                            upper_separator="_",
+                            lower_separator="_",
+                            fill_mode="0",
+                            top_dig_place=14,
+                        ),
+                        "000_000_123_456_789.654_321",
+                    ),
+                    (
+                        Formatter(
+                            upper_separator="_",
+                            lower_separator="_",
+                            fill_mode=" ",
+                            top_dig_place=14,
+                        ),
+                        "      123_456_789.654_321",
+                    ),
+                ],
+            ),
+            (
+                4567899.7654321,
+                [
+                    (
+                        Formatter(
+                            upper_separator="_",
+                            lower_separator="_",
+                            fill_mode="0",
+                            top_dig_place=14,
+                        ),
+                        "000_000_004_567_899.765_432_1",
+                    ),
+                    (
+                        Formatter(
+                            upper_separator="_",
[…]
