+++ examples/fit_plot_with_sciform.py
@@ -11,36 +11,36 @@ from sciform import Formatter
 
 
 def get_scale_and_offset_from_offset_str(
-        ax: plt.Axes, axis: Literal['x', 'y']) -> tuple[float, float]:
+        ax: plt.Axes, axis: Literal["x", "y"]) -> tuple[float, float]:
     """
     Extract the scale and offset for a particular axis from the existing
     offset text when the axis is formatted in scientific mode.
     """
     plt.draw()
-    if axis == 'x':
+    if axis == "x":
         offset_text_obj = ax.xaxis.get_offset_text()
-    elif axis == 'y':
+    elif axis == "y":
         offset_text_obj = ax.yaxis.get_offset_text()
     else:
-        raise ValueError(f'axis must be \'x\' or \'y\', not '
-                         f'\'{axis}\'.')
+        raise ValueError(f"axis must be \'x\' or \'y\', not "
+                         f"\'{axis}\'.")
 
-    ax.ticklabel_format(axis=axis, style='sci')
+    ax.ticklabel_format(axis=axis, style="sci")
     ax.get_figure().canvas.draw()  # Redraw canvas to update offset text
     offset_text = offset_text_obj.get_text()
 
     # Replace minus sign with hyphen minus sign
-    offset_text = offset_text.replace('\u2212', '-')
+    offset_text = offset_text.replace("\u2212", "-")
 
-    pattern = re.compile(r'^(?P<scale>1e[+-]?\d+)?(?P<offset>[+-]1e\d+)?$')
+    pattern = re.compile(r"^(?P<scale>1e[+-]?\d+)?(?P<offset>[+-]1e\d+)?$")
     match = re.match(pattern, offset_text)
-    scale = float(match.group('scale') or 1)
-    offset = float(match.group('offset') or 0)
+    scale = float(match.group("scale") or 1)
+    offset = float(match.group("offset") or 0)
 
     return scale, offset
 
 
-def prefix_exp_ticks(ax: plt.Axes, axis: Literal['x', 'y'],
+def prefix_exp_ticks(ax: plt.Axes, axis: Literal["x", "y"],
                      shifted: bool = False) -> None:
     """
     Use prefix notation for axis tick labels. Scale the tick labels by
@@ -49,25 +49,25 @@ def prefix_exp_ticks(ax: plt.Axes, axis: Literal['x', 'y'],
     offset text into SI prefix format as well.
     """
     if not shifted:
-        exp_mode = 'engineering'
+        exp_mode = "engineering"
     else:
-        exp_mode = 'engineering_shifted'
+        exp_mode = "engineering_shifted"
     tick_formatter = Formatter(
         exp_mode=exp_mode,
-        exp_format='prefix')
+        exp_format="prefix")
     offset_formatter = Formatter(
-        sign_mode='+',
+        sign_mode="+",
         exp_mode=exp_mode,
-        exp_format='prefix')
+        exp_format="prefix")
 
-    ax.ticklabel_format(axis=axis, style='sci')
+    ax.ticklabel_format(axis=axis, style="sci")
 
-    if axis == 'x':
+    if axis == "x":
         old_ticklabels = ax.get_xticklabels()
-    elif axis == 'y':
+    elif axis == "y":
         old_ticklabels = ax.get_yticklabels()
     else:
-        raise ValueError(f'axis must be \'x\' or \'y\', not \'{axis}\'.')
+        raise ValueError(f"axis must be \'x\' or \'y\', not \'{axis}\'.")
 
     scale, offset = get_scale_and_offset_from_offset_str(ax, axis)
 
@@ -75,13 +75,13 @@ def prefix_exp_ticks(ax: plt.Axes, axis: Literal['x', 'y'],
     new_tick_labels = list()
[…]
