+++ examples/fit_plot_with_sciform.py
@@ -1,29 +1,31 @@
-import json
-from typing import Literal
-import re
+from __future__ import annotations
 
-import numpy as np
+import json
+import re
+from pathlib import Path
+from typing import TYPE_CHECKING, Literal
+
 import matplotlib.pyplot as plt
+import numpy as np
+from sciform import Formatter
 from scipy.optimize import curve_fit
 from tabulate import tabulate
 
-from sciform import Formatter
+if TYPE_CHECKING:
+    from numpy.typing import NDArray
 
 
 def get_scale_and_offset_from_offset_str(
         ax: plt.Axes, axis: Literal["x", "y"]) -> tuple[float, float]:
-    """
-    Extract the scale and offset for a particular axis from the existing
-    offset text when the axis is formatted in scientific mode.
-    """
+    """Get the scale and offset for an axis formatted in scientific mode."""
     plt.draw()
     if axis == "x":
         offset_text_obj = ax.xaxis.get_offset_text()
     elif axis == "y":
         offset_text_obj = ax.yaxis.get_offset_text()
     else:
-        raise ValueError(f"axis must be \'x\' or \'y\', not "
-                         f"\'{axis}\'.")
+        msg = f"axis must be 'x' or 'y', not '{axis}'."
+        raise ValueError(msg)
 
     ax.ticklabel_format(axis=axis, style="sci")
     ax.get_figure().canvas.draw()  # Redraw canvas to update offset text
@@ -40,8 +42,11 @@ def get_scale_and_offset_from_offset_str(
     return scale, offset
 
 
-def prefix_exp_ticks(ax: plt.Axes, axis: Literal["x", "y"],
-                     shifted: bool = False) -> None:
+def prefix_exp_ticks(
+        ax: plt.Axes, axis: Literal["x", "y"],
+        *,
+        shifted: bool = False,
+) -> None:
     """
     Use prefix notation for axis tick labels. Scale the tick labels by
     the multiplier that appears in the offset text and format the labels
@@ -52,9 +57,11 @@ def prefix_exp_ticks(ax: plt.Axes, axis: Literal["x", "y"],
         exp_mode = "engineering"
     else:
         exp_mode = "engineering_shifted"
+    # noinspection PyTypeChecker
     tick_formatter = Formatter(
         exp_mode=exp_mode,
         exp_format="prefix")
+    # noinspection PyTypeChecker
     offset_formatter = Formatter(
         sign_mode="+",
         exp_mode=exp_mode,
@@ -67,12 +74,13 @@ def prefix_exp_ticks(ax: plt.Axes, axis: Literal["x", "y"],
     elif axis == "y":
         old_ticklabels = ax.get_yticklabels()
     else:
-        raise ValueError(f"axis must be \'x\' or \'y\', not \'{axis}\'.")
+        msg = f"axis must be 'x' or 'y', not '{axis}'."
+        raise ValueError(msg)
 
     scale, offset = get_scale_and_offset_from_offset_str(ax, axis)
 
-    new_tick_locations = list()
-    new_tick_labels = list()
+    new_tick_locations = []
+    new_tick_labels = []
     for old_ticklabel in old_ticklabels:
         x, y = old_ticklabel.get_position()
         if axis == "x":
@@ -99,22 +107,23 @@ def prefix_exp_ticks(ax: plt.Axes, axis: Literal["x", "y"],
         ax.text(x=0, y=1.01, s=offset_str, transform=ax.transAxes)
 
 
-def quadratic(x, c, x0, y0):
+def quadratic(x: NDArray, c: float, x0: float, y0: float) -> NDArray:
     return (c / 2) * (x - x0) ** 2 + y0
 
 
-def main():
[…]
