+++ examples/fit_plot_no_sciform.py
@@ -1,21 +1,29 @@
-import json
+from __future__ import annotations
 
-import numpy as np
+import json
+from pathlib import Path
+from typing import TYPE_CHECKING
+
 import matplotlib.pyplot as plt
+import numpy as np
 from scipy.optimize import curve_fit
 from tabulate import tabulate
 
+if TYPE_CHECKING:
+    from numpy.typing import NDArray
 
-def quadratic(x, c, x0, y0):
+
+def quadratic(x: NDArray, c: float, x0: float, y0: float) -> NDArray:
     return (c / 2) * (x - x0) ** 2 + y0
 
 
-def main():
-    with open("data/fit_data.json", "r") as f:
+def main() -> None:
+    data_path = Path("data", "fit_data.json")
+    with data_path.open() as f:
         data_dict = json.load(f)
 
     color_list = ["red", "blue", "purple"]
-    fit_results_list = list()
+    fit_results_list = []
 
     fig, ax = plt.subplots(1, 1)
 
@@ -24,12 +32,12 @@ def main():
         y = single_data_dict["y"]
         y_err = single_data_dict["y_err"]
 
-        fit_results_dict = dict()
+        fit_results_dict = {}
 
         color = color_list[idx]
-        ax.errorbar(x, y, y_err, marker="o", linestyle="none", color=color,
-                    label=color)
+        ax.errorbar(x, y, y_err, marker="o", linestyle="none", color=color, label=color)
 
+        # noinspection PyTupleAssignmentBalance
         popt, pcov = curve_fit(quadratic, x, y, sigma=y_err, p0=(2e13, 0, 1e9))
 
         model_x = np.linspace(min(x), max(x), 100)
@@ -49,17 +57,22 @@ def main():
 
         fit_results_list.append(fit_results_dict)
 
-    ax.grid(True)
+    ax.grid(True)  # noqa: FBT003
     ax.legend()
 
     fig.savefig("outputs/fit_plot_no_sciform.png", facecolor="white")
     plt.show()
 
-    table_str = tabulate(fit_results_list, tablefmt="grid", headers="keys",
-                         floatfmt="#.2g")
-    with open("outputs/fit_plot_no_sciform_table.txt", "w") as f:
+    table_str = tabulate(
+        fit_results_list,
+        tablefmt="grid",
+        headers="keys",
+        floatfmt="#.2g",
+    )
+    table_path = Path("outputs", "fit_plot_no_sciform_table.txt")
+    with table_path.open("w") as f:
         f.write(table_str)
-    print(table_str)
+    print(table_str)  # noqa: T201
 
 
 if __name__ == "__main__":
