+++ pyproject.toml
@@ -46,6 +46,7 @@ ignore = [
 [tool.ruff.per-file-ignores]
 # Also ignore `E402` in all `__init__.py` files.
 "tests/*.py" = ["D", "ANN", "PT"]
+"examples/*.py" = ["D"]
 
 [tool.ruff.lint.pydocstyle]
 convention = "numpy"
