### 2023-12-10 d377fa7
try to add ruff check to github actions (remove flake8)
