### 2023-12-10 6ccf93d
fix ruff test code and add to changelog
