+++ src/sciform/scinum.py
@@ -8,7 +8,7 @@ from typing import TYPE_CHECKING
 from sciform.formatting import format_num, format_val_unc
 from sciform.fsml import format_options_from_fmt_spec
 
-if TYPE_CHECKING:
+if TYPE_CHECKING:  # pragma: no cover
     from sciform.format_utils import Number
 
 
