+++ src/sciform/rendered_options.py
@@ -12,7 +12,7 @@ from enum import Enum
 from pprint import pformat
 from typing import TYPE_CHECKING
 
-if TYPE_CHECKING:
+if TYPE_CHECKING:  # pragma: no cover
     from sciform import modes
 
 
