+++ tests/test_float_fsml.py
@@ -10,7 +10,7 @@ class TestFormatting(unittest.TestCase):
         for num, formats_list in cases_list:
             for format_spec, expected_num_str in formats_list:
                 snum = SciNum(num)
-                snum_str = f"{snum:{format_spec}}"
+                snum_str = f'{snum:{format_spec}}'
                 with self.subTest(
                     num=num,
                     format_spec=format_spec,
@@ -24,45 +24,45 @@ class TestFormatting(unittest.TestCase):
             (
                 123.456,
                 [
-                    ("f", "123.456"),
-                    (".-3f", "0"),
-                    (".-2f", "100"),
-                    (".-1f", "120"),
-                    (".0f", "123"),
-                    (".1f", "123.5"),
-                    (".2f", "123.46"),
-                    (".3f", "123.456"),
-                    (".4f", "123.4560"),
-                    ("!1f", "100"),
-                    ("!2f", "120"),
-                    ("!3f", "123"),
-                    ("!4f", "123.5"),
-                    ("!5f", "123.46"),
-                    ("!6f", "123.456"),
-                    ("!7f", "123.4560"),
+                    ('f', '123.456'),
+                    ('.-3f', '0'),
+                    ('.-2f', '100'),
+                    ('.-1f', '120'),
+                    ('.0f', '123'),
+                    ('.1f', '123.5'),
+                    ('.2f', '123.46'),
+                    ('.3f', '123.456'),
+                    ('.4f', '123.4560'),
+                    ('!1f', '100'),
+                    ('!2f', '120'),
+                    ('!3f', '123'),
+                    ('!4f', '123.5'),
+                    ('!5f', '123.46'),
+                    ('!6f', '123.456'),
+                    ('!7f', '123.4560'),
                 ],
             ),
             (
                 0.00062607,
                 [
-                    ("f", "0.00062607"),
-                    (".-1f", "0"),
-                    (".0f", "0"),
-                    (".1f", "0.0"),
-                    (".2f", "0.00"),
-                    (".3f", "0.001"),
-                    (".4f", "0.0006"),
-                    (".5f", "0.00063"),
-                    (".6f", "0.000626"),
-                    (".7f", "0.0006261"),
-                    (".8f", "0.00062607"),
-                    (".9f", "0.000626070"),
-                    ("!1f", "0.0006"),
-                    ("!2f", "0.00063"),
-                    ("!3f", "0.000626"),
-                    ("!4f", "0.0006261"),
-                    ("!5f", "0.00062607"),
-                    ("!6f", "0.000626070"),
+                    ('f', '0.00062607'),
+                    ('.-1f', '0'),
+                    ('.0f', '0'),
+                    ('.1f', '0.0'),
+                    ('.2f', '0.00'),
+                    ('.3f', '0.001'),
+                    ('.4f', '0.0006'),
+                    ('.5f', '0.00063'),
+                    ('.6f', '0.000626'),
+                    ('.7f', '0.0006261'),
+                    ('.8f', '0.00062607'),
+                    ('.9f', '0.000626070'),
+                    ('!1f', '0.0006'),
+                    ('!2f', '0.00063'),
+                    ('!3f', '0.000626'),
+                    ('!4f', '0.0006261'),
+                    ('!5f', '0.00062607'),
+                    ('!6f', '0.000626070'),
                 ],
             ),
         ]
@@ -74,45 +74,45 @@ class TestFormatting(unittest.TestCase):
[…]
