+++ src/sciform/__init__.py
@@ -17,19 +17,19 @@ from sciform.modes import AutoDigits, AutoExpVal
 from sciform.scinum import SciNum, SciNumUnc
 
 __all__ = [
-    "Formatter",
-    "GlobalDefaultsContext",
-    "global_add_c_prefix",
-    "global_add_ppth_form",
-    "global_add_small_si_prefixes",
-    "global_reset_iec_prefixes",
-    "global_reset_parts_per_forms",
-    "global_reset_si_prefixes",
-    "print_global_defaults",
-    "reset_global_defaults",
-    "set_global_defaults",
-    "AutoDigits",
-    "AutoExpVal",
-    "SciNum",
-    "SciNumUnc",
+    'Formatter',
+    'GlobalDefaultsContext',
+    'global_add_c_prefix',
+    'global_add_ppth_form',
+    'global_add_small_si_prefixes',
+    'global_reset_iec_prefixes',
+    'global_reset_parts_per_forms',
+    'global_reset_si_prefixes',
+    'print_global_defaults',
+    'reset_global_defaults',
+    'set_global_defaults',
+    'AutoDigits',
+    'AutoExpVal',
+    'SciNum',
+    'SciNumUnc',
 ]
