+++ docs/source/conf.py
@@ -9,9 +9,9 @@ from importlib import metadata
 # -- Project information -----------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#project-information
 
-project = "sciform"
-copyright = "2023, Justin Gerber"
-author = "Justin Gerber"
+project = 'sciform'
+copyright = '2023, Justin Gerber'
+author = 'Justin Gerber'
 
 version = metadata.version(project)
 release = version
@@ -20,27 +20,27 @@ release = version
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#general-configuration
 
 extensions = [
-    "sphinx.ext.autodoc",
-    "sphinx.ext.doctest",
-    "sphinx.ext.todo",
-    "sphinx_toolbox.collapse",
+    'sphinx.ext.autodoc',
+    'sphinx.ext.doctest',
+    'sphinx.ext.todo',
+    'sphinx_toolbox.collapse',
 ]
 
-templates_path = ["_templates"]
+templates_path = ['_templates']
 exclude_patterns = []
 
 # -- Options for HTML output -------------------------------------------------
 # https://www.sphinx-doc.org/en/master/usage/configuration.html#options-for-html-output
 
-html_theme = "sphinx_rtd_theme"
-html_static_path = ["_static"]
+html_theme = 'sphinx_rtd_theme'
+html_static_path = ['_static']
 html_theme_options = {
-    "navigation_depth": 4,
+    'navigation_depth': 4,
 }
 todo_include_todos = True
 todo_emit_warnings = True
 
 # For wrapping of text in tables
 html_css_files = [
-    "css/custom.css",
+    'css/custom.css',
 ]

[…]
