+++ tests/test_float_formatter.py
@@ -23,8 +23,8 @@ class TestFormatting(unittest.TestCase):
                 789,
                 [
                     (
-                        Formatter(exp_mode='scientific', superscript_exp=True),
-                        '7.89×10²',
+                        Formatter(exp_mode="scientific", superscript_exp=True),
+                        "7.89×10²",
                     ),
                 ],
             ),
@@ -34,18 +34,18 @@ class TestFormatting(unittest.TestCase):
                 [
                     (
                         Formatter(
-                            exp_mode='scientific',
-                            exp_format='prefix',
+                            exp_mode="scientific",
+                            exp_format="prefix",
                             superscript_exp=True,
                         ),
-                        '7.89×10²',
+                        "7.89×10²",
                     ),
                 ],
             ),
             (
                 1024,
                 [
-                    (Formatter(exp_mode='binary', superscript_exp=True), '1×2¹⁰'),
+                    (Formatter(exp_mode="binary", superscript_exp=True), "1×2¹⁰"),
                 ],
             ),
         ]
@@ -59,21 +59,21 @@ class TestFormatting(unittest.TestCase):
                 [
                     (
                         Formatter(
-                            upper_separator='_',
-                            lower_separator='_',
-                            fill_mode='0',
+                            upper_separator="_",
+                            lower_separator="_",
+                            fill_mode="0",
                             top_dig_place=14,
                         ),
-                        '000_000_123_456_789.654_321',
+                        "000_000_123_456_789.654_321",
                     ),
                     (
                         Formatter(
-                            upper_separator='_',
-                            lower_separator='_',
-                            fill_mode=' ',
+                            upper_separator="_",
+                            lower_separator="_",
+                            fill_mode=" ",
                             top_dig_place=14,
                         ),
-                        '      123_456_789.654_321',
+                        "      123_456_789.654_321",
                     ),
                 ],
             ),
@@ -82,21 +82,21 @@ class TestFormatting(unittest.TestCase):
                 [
                     (
                         Formatter(
-                            upper_separator='_',
-                            lower_separator='_',
-                            fill_mode='0',
+                            upper_separator="_",
+                            lower_separator="_",
+                            fill_mode="0",
                             top_dig_place=14,
                         ),
-                        '000_000_004_567_899.765_432_1',
+                        "000_000_004_567_899.765_432_1",
                     ),
                     (
                         Formatter(
-                            upper_separator='_',
-                            lower_separator='_',
-                            fill_mode=' ',
+                            upper_separator="_",
+                            lower_separator="_",
+                            fill_mode=" ",
                             top_dig_place=14,
                         ),
-                        '        4_567_899.765_432_1',
+                        "        4_567_899.765_432_1",
                     ),
                 ],
             ),
@@ -110,17 +110,17 @@ class TestFormatting(unittest.TestCase):
                 789,
                 [
                     (
-                        Formatter(exp_mode='scientific', latex=True),
-                        r'7.89\times 10^{+2}',
+                        Formatter(exp_mode="scientific", latex=True),
+                        r"7.89\times 10^{+2}",
                     ),
                     # Latex mode takes precedence over superscript_exp
                     (
                         Formatter(
-                            exp_mode='scientific',
+                            exp_mode="scientific",
                             latex=True,
                             superscript_exp=True,
                         ),
-                        r'7.89\times 10^{+2}',
+                        r"7.89\times 10^{+2}",
                     ),
                 ],
             ),
@@ -129,21 +129,21 @@ class TestFormatting(unittest.TestCase):
[…]
