+++ pyproject.toml
@@ -44,7 +44,6 @@ ignore = [
 ]
 
 [tool.ruff.per-file-ignores]
-# Also ignore `E402` in all `__init__.py` files.
 "tests/*.py" = ["D", "ANN", "PT"]
 "examples/*.py" = ["D"]
 
