+++ tests/test_val_unc_fsml.py
@@ -1,3 +1,5 @@
+from __future__ import annotations
+
 import unittest
 
 from sciform import GlobalDefaultsContext, SciNumUnc
