### 2023-12-11 #121 @jagerber48
Ok, merged a big commit into `main` branch (not released yet though) where I lint and reformat using ruff: https://github.com/jagerber48/sciform/pull/88. Most of it I'm happy with. I was a little annoyed to move to double quotes from single quotes since it takes an extra keystroke to type double quotes. Guess I can type single quotes and let the tool format to double later. Ruff supports single quotes for some functionality but not all, and Black has made the opiniated choice (that I disagree with) to use double quotes, so there is just more tooling support for double quotes. Once Ruff gets support for single quotes on more features I might move back to them.

Trying to wrap up some more checkboxes. @Batalex you suggest removing exception raising in private function. Most of the exceptions in `sciform` arise as `else` clauses in `if... elif...  else...` exhaustive checks. See [this discussion](https://discuss.python.org/t/how-should-i-terminate-if-elif-elif-when-all-options-are-exhausted/31796/1). See this block of code

```
from enum import Enum

class Options(Enum):
    OPTION_1 = 1
    OPTION_2 = 2

def foo(option: Options):
    if option is Options.OPTION_1:
        res = 1
    elif option is Options.OPTION_2:
        res = 2
    return res
```
Pycharm gives me a warning on the last line saying `Local variable 'res' might be referenced before assignment.`. This motivated a lot of the `else` clauses with exceptions. Do you think this is a bad practice? I should check Ruff's behavior on this. Ruff might be more intelligent about static type checking exhaustiveness checks.
