### 2023-12-11 #121 @jagerber48
Ok, thanks for the response. In that case, the list is indeed getting short. I'm going to collect the remaining items I have

- [ ] Annotated FSML examples. These would be useful. They'll take some time to get in. I think the separator configurations, and the exponent value forcing settings are probably the most new/confusing.
- [ ] Rename `top_dig_place` to something like `left_pad_digit` or `left_pad_to_digit`
- [ ] Rename `superscript_exp` to `superscript`
- [ ] Think about clarity of the `unc` abbreviation. Consider revising the interface so that options specific to value/uncertainty formatting sit in a separate place than the rest of the options. This might allow for some improved options names
- [ ] Think about `SciNum`/`SciNumUnc` naming.
- [ ] Consider using precommit hooks for linting/formatting/tests (in lieu of running github actions locally)
- [ ] Add development install instructions. @isabelizimm, you were the one who suggested I end develop install instructions, curious for your and others' thoughts on the following. I'm curious what is typical here. My first guess would be to recommend devs fork and `git clone` the repo then `pip install -r requirements-dev.txt` or something. I know I could configure `pyproject.toml` to allow `python -m pip install sciform[dev]` to work. But I think I'd prefer the former approach so they already have it set up as a git repo and are able to push and make pull requests. And then what should dev requirements be? `sciform` itself has no requirements for installation and use. For simple dev I guess devs would want `ruff` installed. For editing docs I need a couple `sphinx` and `sphinx` utility packages. For the examples I need the heavyweight `numpy`/`scipy`/`matplotlib` requirements. Should I include ALL of these as dev requirements? Should I have separate requirements files for each of these different dev use-cases?
- [ ] Research `tox` or `act` for running github actions locally. I think the main challenge here is spinning up environments that use all the different versions of python I test against the way github actions do. But I wonder if this is what `act` can do. Looks like `act` uses docker which sounds like a pain on Windows. Maybe I could run `act` on a Linux VM or something. But I think I'll likely conclude this is not worth the effort. I can just use github action on github. The worst case is I end up with some throwaway commits in PRs whose only purpose was to test some github action behavior.
- [ ] Consider the `FormattedValue` suggestion to support converting formatted values to a number of output formats including, e.g. string literal, Latex formatted string, html formatted string.
