### 2023-12-12 #121 @isabelizimm
These changes and updates look great! If you would like second eyes on any changes going in, feel free to tag me in PRs 😄

Re-development instructions:

**Minimum:** [add a line in installation](https://github.com/jagerber48/sciform/tree/5385f4a7ad2104588413b9f8ae6da07950468902#installation) to show people how to get latest version, eg

```
pip install git+https://github.com/jagerber48/sciform.git
```

This will show people how to get latest development changes, even if they're not released yet. ([example](https://github.com/rstudio/vetiver-python#installation))

**Slightly more effort:** some documentation, either in a README or CONTRIBUTING file to show mainly 1) how to get to a local install, 2) how to build docs and 3) how to run tests. [One example is the CONTRIBUTING.md guide in the package `pins`.](https://github.com/rstudio/pins-python/blob/main/CONTRIBUTING.md).

How exactly you want to configure those is up to you. My experience is that `requirements.txt` files tend to get stale, so I generally lean towards `sciform[dev]` via pyproject.toml route. _However_, the sciform package itself doesn't have any dependencies, so I think that using requirements files is a perfectly valid move in this scenario if it works better for you! If you want to make it different requirements files, splitting it out to requirements-test/requirements-docs makes sense, but probably no more modular than that.
