### 2023-12-16 86f338f
add dev install dependencies
