+++ CONTRIBUTING.rst
@@ -21,3 +21,21 @@ ease-of-use for the user interface as well as naming suggestions for the
 various objects and options that users interact with.
 While stabilizing the interface will take precedence for the time being,
 new feature requests are always welcome!
+
+Development Details
+-------------------
+
+* Begin development work on ``sciform`` by forking and installing the
+  `git repository <https://github.com/jagerber48/sciform>`_ and
+  installing the development dependencies in editable mode with::
+
+     python -m pip install -e .[dev]
+* Tests can be run using::
+
+     python -m unittest
+* ``sciform`` is formatted using the
+  `ruff linter and formatter <https://docs.astral.sh/ruff/>`_.
+  Code should pass the following checks with no errors::
+
+     ruff check .
+     ruff format . --check
