### 2023-12-16 #121 @jagerber48
@isabelizimm Thanks for all that info. Ok, it sounds like your main suggestion is to include instructions for how people can *install* the latest development/pre-release version of the code to e.g. test it out. That is easy to do. The other part that we've been discussing is to include instructions for people who want to *develop* `sciform`. These two things have different requirements.

In any case, I have a PR with both of them I would appreciate you having a look at if you're able to: https://github.com/jagerber48/sciform/pull/90. I think it's relatively straight forward.

I decided to go with the `pip install -e .[dev]` approach, and also to not break things down any further than this. The `numpy/scipy/matplotlib` dev requirements really aren't that burdensome, and so few people are going to be developing `sciform` I think it's just not worth any added overhead.
