+++ tests/test_print.py
@@ -26,7 +26,7 @@ class TestPrint(unittest.TestCase):
             " 'extra_iec_prefixes': {},\n"
             " 'extra_parts_per_forms': {},\n"
             " 'capitalize': False,\n"
-            " 'superscript_exp': False,\n"
+            " 'superscript': False,\n"
             " 'latex': False,\n"
             " 'nan_inf_exp': False,\n"
             " 'bracket_unc': False,\n"
@@ -59,7 +59,7 @@ class TestPrint(unittest.TestCase):
             " 'extra_iec_prefixes': {},\n"
             " 'extra_parts_per_forms': {},\n"
             " 'capitalize': True,\n"
-            " 'superscript_exp': False,\n"
+            " 'superscript': False,\n"
             " 'latex': False,\n"
             " 'nan_inf_exp': False,\n"
             " 'bracket_unc': False,\n"
