+++ tests/test_float_formatter.py
@@ -17,13 +17,13 @@ class TestFormatting(unittest.TestCase):
                 ):
                     self.assertEqual(snum_str, expected_num_str)
 
-    def test_superscript_exp(self):
+    def test_superscript(self):
         cases_list = [
             (
                 789,
                 [
                     (
-                        Formatter(exp_mode="scientific", superscript_exp=True),
+                        Formatter(exp_mode="scientific", superscript=True),
                         "7.89×10²",
                     ),
                 ],
@@ -36,7 +36,7 @@ class TestFormatting(unittest.TestCase):
                         Formatter(
                             exp_mode="scientific",
                             exp_format="prefix",
-                            superscript_exp=True,
+                            superscript=True,
                         ),
                         "7.89×10²",
                     ),
@@ -45,7 +45,7 @@ class TestFormatting(unittest.TestCase):
             (
                 1024,
                 [
-                    (Formatter(exp_mode="binary", superscript_exp=True), "1×2¹⁰"),
+                    (Formatter(exp_mode="binary", superscript=True), "1×2¹⁰"),
                 ],
             ),
         ]
@@ -113,12 +113,12 @@ class TestFormatting(unittest.TestCase):
                         Formatter(exp_mode="scientific", latex=True),
                         r"7.89\times 10^{+2}",
                     ),
-                    # Latex mode takes precedence over superscript_exp
+                    # Latex mode takes precedence over superscript
                     (
                         Formatter(
                             exp_mode="scientific",
                             latex=True,
-                            superscript_exp=True,
+                            superscript=True,
                         ),
                         r"7.89\times 10^{+2}",
                     ),
