+++ src/sciform/global_configuration.py
@@ -35,7 +35,7 @@ def set_global_defaults(  # noqa: PLR0913
     extra_iec_prefixes: dict[int, str] | None = None,
     extra_parts_per_forms: dict[int, str] | None = None,
     capitalize: bool | None = None,
-    superscript_exp: bool | None = None,
+    superscript: bool | None = None,
     latex: bool | None = None,
     nan_inf_exp: bool | None = None,
     bracket_unc: bool | None = None,
@@ -68,7 +68,7 @@ def set_global_defaults(  # noqa: PLR0913
         extra_iec_prefixes=extra_iec_prefixes,
         extra_parts_per_forms=extra_parts_per_forms,
         capitalize=capitalize,
-        superscript_exp=superscript_exp,
+        superscript=superscript,
         latex=latex,
         nan_inf_exp=nan_inf_exp,
         bracket_unc=bracket_unc,
@@ -175,7 +175,7 @@ class GlobalDefaultsContext:
         extra_iec_prefixes: dict[int, str] | None = None,
         extra_parts_per_forms: dict[int, str] | None = None,
         capitalize: bool | None = None,
-        superscript_exp: bool | None = None,
+        superscript: bool | None = None,
         latex: bool | None = None,
         nan_inf_exp: bool | None = None,
         bracket_unc: bool | None = None,
@@ -203,7 +203,7 @@ class GlobalDefaultsContext:
             extra_iec_prefixes=extra_iec_prefixes,
             extra_parts_per_forms=extra_parts_per_forms,
             capitalize=capitalize,
-            superscript_exp=superscript_exp,
+            superscript=superscript,
             latex=latex,
             nan_inf_exp=nan_inf_exp,
             bracket_unc=bracket_unc,
