### 2023-12-17 4c0c1b7
top_dig_place -> left_pad_dec_place
