+++ src/sciform/global_configuration.py
@@ -29,7 +29,7 @@ def set_global_defaults(  # noqa: PLR0913
     lower_separator: modes.UserLowerSeparators | None = None,
     sign_mode: modes.UserSignMode | None = None,
     fill_mode: modes.UserFillMode | None = None,
-    top_dig_place: int | None = None,
+    left_pad_dec_place: int | None = None,
     exp_format: modes.UserExpFormat | None = None,
     extra_si_prefixes: dict[int, str] | None = None,
     extra_iec_prefixes: dict[int, str] | None = None,
@@ -62,7 +62,7 @@ def set_global_defaults(  # noqa: PLR0913
         lower_separator=lower_separator,
         sign_mode=sign_mode,
         fill_mode=fill_mode,
-        top_dig_place=top_dig_place,
+        left_pad_dec_place=left_pad_dec_place,
         exp_format=exp_format,
         extra_si_prefixes=extra_si_prefixes,
         extra_iec_prefixes=extra_iec_prefixes,
@@ -169,7 +169,7 @@ class GlobalDefaultsContext:
         lower_separator: modes.UserLowerSeparators | None = None,
         sign_mode: modes.UserSignMode | None = None,
         fill_mode: modes.UserFillMode | None = None,
-        top_dig_place: int | None = None,
+        left_pad_dec_place: int | None = None,
         exp_format: modes.UserExpFormat | None = None,
         extra_si_prefixes: dict[int, str] | None = None,
         extra_iec_prefixes: dict[int, str] | None = None,
@@ -197,7 +197,7 @@ class GlobalDefaultsContext:
             lower_separator=lower_separator,
             sign_mode=sign_mode,
             fill_mode=fill_mode,
-            top_dig_place=top_dig_place,
+            left_pad_dec_place=left_pad_dec_place,
             exp_format=exp_format,
             extra_si_prefixes=extra_si_prefixes,
             extra_iec_prefixes=extra_iec_prefixes,
