+++ src/sciform/global_options.py
@@ -13,7 +13,7 @@ PKG_DEFAULT_OPTIONS = RenderedOptions(
     lower_separator=modes.Separator.NONE,
     sign_mode=modes.SignMode.NEGATIVE,
     fill_mode=modes.FillMode.SPACE,
-    top_dig_place=0,
+    left_pad_dec_place=0,
     exp_format=modes.ExpFormat.STANDARD,
     extra_si_prefixes={},
     extra_iec_prefixes={},
