+++ tests/test_float_formatter.py
@@ -62,7 +62,7 @@ class TestFormatting(unittest.TestCase):
                             upper_separator="_",
                             lower_separator="_",
                             fill_mode="0",
-                            top_dig_place=14,
+                            left_pad_dec_place=14,
                         ),
                         "000_000_123_456_789.654_321",
                     ),
@@ -71,7 +71,7 @@ class TestFormatting(unittest.TestCase):
                             upper_separator="_",
                             lower_separator="_",
                             fill_mode=" ",
-                            top_dig_place=14,
+                            left_pad_dec_place=14,
                         ),
                         "      123_456_789.654_321",
                     ),
@@ -85,7 +85,7 @@ class TestFormatting(unittest.TestCase):
                             upper_separator="_",
                             lower_separator="_",
                             fill_mode="0",
-                            top_dig_place=14,
+                            left_pad_dec_place=14,
                         ),
                         "000_000_004_567_899.765_432_1",
                     ),
@@ -94,7 +94,7 @@ class TestFormatting(unittest.TestCase):
                             upper_separator="_",
                             lower_separator="_",
                             fill_mode=" ",
-                            top_dig_place=14,
+                            left_pad_dec_place=14,
                         ),
                         "        4_567_899.765_432_1",
                     ),

[…]
