### 2023-12-17 #121 @jagerber48
I'm going to rename `top_dig_place` to `left_pad_dec_place`. This is consistent with a renaming of the rounding mode that rounds by "decimal place" (similar to "precision" for the python built-in FSML) to `dec_place`. "Digits place" is not a standard terminology for this concept, "decimal place" is the standard terminology.

Here's an idea on `SciNum` and `SciNumUnc`. I think `SciNum` is an ok name for a number that is going to be converted to a formatted string. I don't like `SciNumUnc` though. I think I could remove the `SciNumUnc` object and just let the `SciNum` `__init__` method take an optional uncertainty input parameter and then perform the formatting accordingly.
