+++ tests/test_print.py
@@ -33,7 +33,7 @@ class TestPrint(unittest.TestCase):
             " 'pdg_sig_figs': False,\n"
             " 'val_unc_match_widths': False,\n"
             " 'bracket_unc_remove_seps': False,\n"
-            " 'unc_pm_whitespace': True}\n"
+            " 'pm_whitespace': True}\n"
         )
         self.assertEqual(actual_printout, expected_printout)
 
@@ -66,6 +66,6 @@ class TestPrint(unittest.TestCase):
             " 'pdg_sig_figs': False,\n"
             " 'val_unc_match_widths': False,\n"
             " 'bracket_unc_remove_seps': False,\n"
-            " 'unc_pm_whitespace': True}\n"
+            " 'pm_whitespace': True}\n"
         )
         self.assertEqual(actual_printout, expected_printout)
