+++ src/sciform/global_configuration.py
@@ -42,7 +42,7 @@ def set_global_defaults(  # noqa: PLR0913
     pdg_sig_figs: bool | None = None,
     val_unc_match_widths: bool | None = None,
     bracket_unc_remove_seps: bool | None = None,
-    unc_pm_whitespace: bool | None = None,
+    pm_whitespace: bool | None = None,
     add_c_prefix: bool = False,
     add_small_si_prefixes: bool = False,
     add_ppth_form: bool = False,
@@ -75,7 +75,7 @@ def set_global_defaults(  # noqa: PLR0913
         pdg_sig_figs=pdg_sig_figs,
         val_unc_match_widths=val_unc_match_widths,
         bracket_unc_remove_seps=bracket_unc_remove_seps,
-        unc_pm_whitespace=unc_pm_whitespace,
+        pm_whitespace=pm_whitespace,
         add_c_prefix=add_c_prefix,
         add_small_si_prefixes=add_small_si_prefixes,
         add_ppth_form=add_ppth_form,
@@ -182,7 +182,7 @@ class GlobalDefaultsContext:
         pdg_sig_figs: bool | None = None,
         val_unc_match_widths: bool | None = None,
         bracket_unc_remove_seps: bool | None = None,
-        unc_pm_whitespace: bool | None = None,
+        pm_whitespace: bool | None = None,
         add_c_prefix: bool = False,
         add_small_si_prefixes: bool = False,
         add_ppth_form: bool = False,
@@ -210,7 +210,7 @@ class GlobalDefaultsContext:
             pdg_sig_figs=pdg_sig_figs,
             val_unc_match_widths=val_unc_match_widths,
             bracket_unc_remove_seps=bracket_unc_remove_seps,
-            unc_pm_whitespace=unc_pm_whitespace,
+            pm_whitespace=pm_whitespace,
             add_c_prefix=add_c_prefix,
             add_small_si_prefixes=add_small_si_prefixes,
             add_ppth_form=add_ppth_form,
