### 2023-12-18 #121 @isabelizimm
**re: development docs**: I reviewed that PR, a few small comments but looks good to me! It does really help users to get started contributing when they don't have to guess about some of the going-though-the-motions pieces of development, so I think this is a great addition. Plus, everything ran on my laptop right out of the box, which is always exciting➕ 😄

**re: CI/pre-commit**: I personally don't use act/tox/etc. I've tried a few, but no project of mine has been heavyweight enough that I found it a net-win to my workflow. The one thing I do use, though, is pre-commit hooks. It might be nice to start using this tool locally to make sure all new code is compliant, and integrate it into CI as you see fit.

A few pre-commit moves that might make it more useful to you:

- You can pass `pre-commit --no-verify` to bypass these hooks, if needed
- You can set up CI for pre-commit, or just have it part of a local workflow (still super useful, imo)
- If you check in the pre-commit hooks into a repository, it can be part of the development install/docs, just like how you have `ruff` currently. Because pre-commit runs, well, _pre_-**commit**, people won't be able to make non-compliant commits.

**re: naming**: These both seem like good moves! The digits place -> decimal place move immediately clicked in my brain, and having uncertainty as a parameter seems reasonable enough to me.
