+++ tests/test_val_unc_fsml.py
@@ -1,6 +1,6 @@
 import unittest
 
-from sciform import GlobalDefaultsContext, SciNumUnc
+from sciform import GlobalDefaultsContext, SciNum
 
 ValUncFSMLCases = list[tuple[tuple[float, float], list[tuple[str, str]]]]
 NAN = float("nan")
@@ -11,7 +11,7 @@ class TestFormatting(unittest.TestCase):
     def run_val_unc_fsml_cases(self, cases_list: ValUncFSMLCases):
         for (val, unc), formats_list in cases_list:
             for format_spec, expected_str in formats_list:
-                snum = SciNumUnc(val, unc)
+                snum = SciNum(val, unc)
                 snum_str = f"{snum:{format_spec}}"
                 with self.subTest(
                     val=val,
