+++ docs/source/fsml.rst
@@ -10,9 +10,9 @@ Specification
 =============
 
 Instead of explicitly constructing :class:`Formatter` instances, users
-can construct :class:`SciNum` or :class:`SciNumUnc` instances and format
-them using string formatting with format specification strings from the
-:mod:`sciform` format specification mini-language (FSML).
+can construct :class:`SciNum` instances and format them using string
+formatting with format specification strings from the :mod:`sciform`
+format specification mini-language (FSML).
 This is analogous to how python :class:`int`, :class:`float`, and
 :class:`Decimal` instances can be formatted using the built-in
 `format specification mini-language <https://docs.python.org/3/library/string.html#format-specification-mini-language>`_.
