+++ CHANGELOG.rst
@@ -13,6 +13,7 @@ Changed
 * Refactored code for adding separators.
 * Refactored formatting and formatting utilities to simplify functions
   and make the algorithm easier to follow.
+* More aggressively filter JetBrains .idea folder from version control.
 
 ----
 
