### 2023-12-24 bbc9891
quotes around .[dev] in dev install instructions
This avoids an issue when this is run on a zsh terminal, see https://stackoverflow.com/questions/30539798/zsh-no-matches-found-requestssecurity
