### 2023-12-24 ef7b5df
Remove separators from FSML
