### 2023-12-24 706fe58
Make formatter based tests for separators
