+++ tests/test_float_fsml.py
@@ -505,65 +505,6 @@ class TestFormatting(unittest.TestCase):
         with GlobalDefaultsContext(nan_inf_exp=True):
             self.run_float_fsml_cases(cases_list)
 
-    # TODO: Replace these with formatter based tests.
-    @unittest.expectedFailure
-    def test_separators(self):
-        cases_list = [
-            (
-                123456.654321,
-                [
-                    ("", "123456.654321"),
-                    (",", "123,456.654321"),
-                    (",.s", "123,456.654 321"),
-                    (",._", "123,456.654_321"),
-                    ("_._", "123_456.654_321"),
-                    ("s.s", "123 456.654 321"),
-                    ("n,n", "123456,654321"),
-                    (".,", "123.456,654321"),
-                    (".,s", "123.456,654 321"),
-                    (".,_", "123.456,654_321"),
-                    ("_,_", "123_456,654_321"),
-                    ("s,s", "123 456,654 321"),
-                ],
-            ),
-            (
-                12345.54321,
-                [
-                    ("", "12345.54321"),
-                    (",", "12,345.54321"),
-                    (",.s", "12,345.543 21"),
-                    (",._", "12,345.543_21"),
-                    ("_._", "12_345.543_21"),
-                    ("s.s", "12 345.543 21"),
-                    ("n,n", "12345,54321"),
-                    (".,", "12.345,54321"),
-                    (".,s", "12.345,543 21"),
-                    (".,_", "12.345,543_21"),
-                    ("_,_", "12_345,543_21"),
-                    ("s,s", "12 345,543 21"),
-                ],
-            ),
-            (
-                1234567.7654321,
-                [
-                    ("", "1234567.7654321"),
-                    (",", "1,234,567.7654321"),
-                    (",.s", "1,234,567.765 432 1"),
-                    (",._", "1,234,567.765_432_1"),
-                    ("_._", "1_234_567.765_432_1"),
-                    ("s.s", "1 234 567.765 432 1"),
-                    ("n,n", "1234567,7654321"),
-                    (".,", "1.234.567,7654321"),
-                    (".,s", "1.234.567,765 432 1"),
-                    (".,_", "1.234.567,765_432_1"),
-                    ("_,_", "1_234_567,765_432_1"),
-                    ("s,s", "1 234 567,765 432 1"),
-                ],
-            ),
-        ]
-
-        self.run_float_fsml_cases(cases_list)
-
     def test_signs(self):
         cases_list = [
             (
