+++ tests/test_float_formatter.py
@@ -255,290 +255,3 @@ class TestFormatting(unittest.TestCase):
     def test_dec_place_auto_round(self):
         sform = Formatter(round_mode="dec_place", ndigits=AutoDigits)
         self.assertEqual(sform(123.456), "123.456")
-
-    # TODO: Move this to separate file
-    def test_separators(self):
-        cases_list = [
-            (
-                123456.654321,
-                [
-                    (
-                        Formatter(),
-                        "123456.654321",
-                    ),
-                    (
-                        Formatter(upper_separator=","),
-                        "123,456.654321",
-                    ),
-                    (
-                        Formatter(
-                            upper_separator=",",
-                            decimal_separator=".",
-                            lower_separator=" ",
-                        ),
-                        "123,456.654 321",
-                    ),
-                    (
-                        Formatter(
-                            upper_separator=",",
-                            decimal_separator=".",
-                            lower_separator="_",
-                        ),
-                        "123,456.654_321",
-                    ),
-                    (
-                        Formatter(
-                            upper_separator="_",
-                            decimal_separator=".",
-                            lower_separator="_",
-                        ),
-                        "123_456.654_321",
-                    ),
-                    (
-                        Formatter(
-                            upper_separator=" ",
-                            decimal_separator=".",
-                            lower_separator=" ",
-                        ),
-                        "123 456.654 321",
-                    ),
-                    (
-                        Formatter(
-                            upper_separator="",
-                            decimal_separator=".",
-                            lower_separator="",
-                        ),
-                        "123456.654321",
-                    ),
-                    (
-                        Formatter(
-                            upper_separator=".",
-                            decimal_separator=",",
-                        ),
-                        "123.456,654321",
-                    ),
-                    (
-                        Formatter(
-                            upper_separator=".",
-                            decimal_separator=",",
-                            lower_separator=" ",
-                        ),
-                        "123.456,654 321",
-                    ),
-                    (
-                        Formatter(
-                            upper_separator=".",
-                            decimal_separator=",",
-                            lower_separator="_",
-                        ),
-                        "123.456,654_321",
-                    ),
-                    (
-                        Formatter(
-                            upper_separator="_",
-                            decimal_separator=",",
-                            lower_separator="_",
-                        ),
-                        "123_456,654_321",
-                    ),
-                    (
-                        Formatter(
-                            upper_separator=" ",
-                            decimal_separator=",",
-                            lower_separator=" ",
-                        ),
-                        "123 456,654 321",
-                    ),
-                ],
-            ),
-            (
-                12345.54321,
-                [
-                    (
-                        Formatter(),
-                        "12345.54321",
-                    ),
-                    (
-                        Formatter(
-                            upper_separator=",",
-                        ),
-                        "12,345.54321",
-                    ),
-                    (
-                        Formatter(
-                            upper_separator=",",
-                            decimal_separator=".",
-                            lower_separator=" ",
-                        ),
-                        "12,345.543 21",
-                    ),
-                    (
-                        Formatter(
-                            upper_separator=",",
-                            decimal_separator=".",
-                            lower_separator="_",
-                        ),
-                        "12,345.543_21",
-                    ),
-                    (
-                        Formatter(
-                            upper_separator="_",
-                            decimal_separator=".",
-                            lower_separator="_",
-                        ),
-                        "12_345.543_21",
-                    ),
-                    (
-                        Formatter(
-                            upper_separator=" ",
-                            decimal_separator=".",
-                            lower_separator=" ",
-                        ),
-                        "12 345.543 21",
-                    ),
-                    (
-                        Formatter(
-                            upper_separator="",
-                            decimal_separator=".",
-                            lower_separator="",
-                        ),
[…]
