### 2023-12-26 jagerber48/sciform#29 @jagerber48
Fixed by #95 which removes separator configuration from the FSML
