+++ src/sciform/scinum.py
@@ -30,7 +30,7 @@ class SciNum:
     + 12.346E+03
     >>> snum = SciNum(123456.654321, 0.0234)
     >>> print(f"{snum:#!2r()}")
-    (0.1234566543(23))e+06
+    (0.123456654(23))e+06
     """
 
     def __init__(
