+++ docs/source/fsml.rst
@@ -41,8 +41,7 @@ In this example ``!4`` indicates the number should be formatted with
 four significant figures and ``f`` indicates the number should be
 formatted in fixed point mode.
 
->>> from sciform import SciNumUnc
->>> print(f'{SciNumUnc(12345, 789):!1r}')
+>>> print(f'{SciNum(12345, 789):!1r}')
 (12.3 ± 0.8)e+03
 
 In this example ``!1`` indicates the number will be formatted so that
