+++ src/sciform/global_options.py
@@ -22,7 +22,7 @@ PKG_DEFAULT_OPTIONS = RenderedOptions(
     superscript=False,
     latex=False,
     nan_inf_exp=False,
-    bracket_unc=False,
+    paren_uncertainty=False,
     pdg_sig_figs=False,
     val_unc_match_widths=False,
     bracket_unc_remove_seps=False,
