+++ tests/test_val_unc_formatter.py
@@ -18,30 +18,30 @@ class TestFormatting(unittest.TestCase):
                 ):
                     self.assertEqual(snum_str, expected_val_unc_str)
 
-    def test_bracket_unc(self):
+    def test_paren_uncertainty(self):
         cases_list = [
             (
                 (123.456, 0.789),
                 [
-                    (Formatter(bracket_unc=True), "123.456(789)"),
+                    (Formatter(paren_uncertainty=True), "123.456(789)"),
                     (
-                        Formatter(exp_mode="scientific", bracket_unc=True),
+                        Formatter(exp_mode="scientific", paren_uncertainty=True),
                         "(1.23456(789))e+02",
                     ),
                     (
-                        Formatter(exp_mode="engineering", bracket_unc=True),
+                        Formatter(exp_mode="engineering", paren_uncertainty=True),
                         "(123.456(789))e+00",
                     ),
                     (
-                        Formatter(exp_mode="engineering_shifted", bracket_unc=True),
+                        Formatter(exp_mode="engineering_shifted", paren_uncertainty=True),
                         "(0.123456(789))e+03",
                     ),
                     (
-                        Formatter(exp_mode="scientific", exp_val=+1, bracket_unc=True),
+                        Formatter(exp_mode="scientific", exp_val=+1, paren_uncertainty=True),
                         "(12.3456(789))e+01",
                     ),
                     (
-                        Formatter(exp_mode="scientific", exp_val=-1, bracket_unc=True),
+                        Formatter(exp_mode="scientific", exp_val=-1, paren_uncertainty=True),
                         "(1234.56(7.89))e-01",
                     ),
                 ],
@@ -55,31 +55,31 @@ class TestFormatting(unittest.TestCase):
             (
                 (123, 0),
                 [
-                    (Formatter(bracket_unc=True), "123(0)"),
+                    (Formatter(paren_uncertainty=True), "123(0)"),
                 ],
             ),
             (
                 (-123, 0),
                 [
-                    (Formatter(bracket_unc=True), "-123(0)"),
+                    (Formatter(paren_uncertainty=True), "-123(0)"),
                 ],
             ),
             (
                 (123, float("nan")),
                 [
-                    (Formatter(bracket_unc=True), "123(nan)"),
+                    (Formatter(paren_uncertainty=True), "123(nan)"),
                 ],
             ),
             (
                 (123, float("inf")),
                 [
-                    (Formatter(bracket_unc=True), "123(inf)"),
+                    (Formatter(paren_uncertainty=True), "123(inf)"),
                 ],
             ),
             (
                 (0, 0),
                 [
-                    (Formatter(bracket_unc=True), "0(0)"),
+                    (Formatter(paren_uncertainty=True), "0(0)"),
                 ],
             ),
         ]
@@ -93,7 +93,7 @@ class TestFormatting(unittest.TestCase):
                 [
                     (
                         Formatter(
-                            bracket_unc=True,
+                            paren_uncertainty=True,
                             exp_mode="engineering",
                             exp_format="prefix",
                         ),
@@ -101,7 +101,7 @@ class TestFormatting(unittest.TestCase):
                     ),
                     (
                         Formatter(
-                            bracket_unc=True,
+                            paren_uncertainty=True,
                             exp_mode="engineering_shifted",
                             exp_format="prefix",
                         ),
@@ -114,7 +114,7 @@ class TestFormatting(unittest.TestCase):
                 [
                     (
                         Formatter(
-                            bracket_unc=True,
+                            paren_uncertainty=True,
                             exp_mode="fixed_point",
                             exp_format="prefix",
                         ),
@@ -122,7 +122,7 @@ class TestFormatting(unittest.TestCase):
                     ),
[…]
