+++ src/sciform/global_configuration.py
@@ -38,7 +38,7 @@ def set_global_defaults(  # noqa: PLR0913
     superscript: bool | None = None,
     latex: bool | None = None,
     nan_inf_exp: bool | None = None,
-    bracket_unc: bool | None = None,
+    paren_uncertainty: bool | None = None,
     pdg_sig_figs: bool | None = None,
     val_unc_match_widths: bool | None = None,
     bracket_unc_remove_seps: bool | None = None,
@@ -71,7 +71,7 @@ def set_global_defaults(  # noqa: PLR0913
         superscript=superscript,
         latex=latex,
         nan_inf_exp=nan_inf_exp,
-        bracket_unc=bracket_unc,
+        paren_uncertainty=paren_uncertainty,
         pdg_sig_figs=pdg_sig_figs,
         val_unc_match_widths=val_unc_match_widths,
         bracket_unc_remove_seps=bracket_unc_remove_seps,
@@ -178,7 +178,7 @@ class GlobalDefaultsContext:
         superscript: bool | None = None,
         latex: bool | None = None,
         nan_inf_exp: bool | None = None,
-        bracket_unc: bool | None = None,
+        paren_uncertainty: bool | None = None,
         pdg_sig_figs: bool | None = None,
         val_unc_match_widths: bool | None = None,
         bracket_unc_remove_seps: bool | None = None,
@@ -206,7 +206,7 @@ class GlobalDefaultsContext:
             superscript=superscript,
             latex=latex,
             nan_inf_exp=nan_inf_exp,
-            bracket_unc=bracket_unc,
+            paren_uncertainty=paren_uncertainty,
             pdg_sig_figs=pdg_sig_figs,
             val_unc_match_widths=val_unc_match_widths,
             bracket_unc_remove_seps=bracket_unc_remove_seps,
