+++ tests/test_print.py
@@ -29,7 +29,7 @@ class TestPrint(unittest.TestCase):
             " 'superscript': False,\n"
             " 'latex': False,\n"
             " 'nan_inf_exp': False,\n"
-            " 'bracket_unc': False,\n"
+            " 'paren_uncertainty': False,\n"
             " 'pdg_sig_figs': False,\n"
             " 'val_unc_match_widths': False,\n"
             " 'bracket_unc_remove_seps': False,\n"
@@ -62,7 +62,7 @@ class TestPrint(unittest.TestCase):
             " 'superscript': False,\n"
             " 'latex': False,\n"
             " 'nan_inf_exp': False,\n"
-            " 'bracket_unc': False,\n"
+            " 'paren_uncertainty': False,\n"
             " 'pdg_sig_figs': False,\n"
             " 'val_unc_match_widths': False,\n"
             " 'bracket_unc_remove_seps': False,\n"
