+++ src/sciform/global_configuration.py
@@ -41,7 +41,7 @@ def set_global_defaults(  # noqa: PLR0913
     paren_uncertainty: bool | None = None,
     pdg_sig_figs: bool | None = None,
     val_unc_match_widths: bool | None = None,
-    bracket_unc_remove_seps: bool | None = None,
+    paren_uncertainty_separators: bool | None = None,
     pm_whitespace: bool | None = None,
     add_c_prefix: bool = False,
     add_small_si_prefixes: bool = False,
@@ -74,7 +74,7 @@ def set_global_defaults(  # noqa: PLR0913
         paren_uncertainty=paren_uncertainty,
         pdg_sig_figs=pdg_sig_figs,
         val_unc_match_widths=val_unc_match_widths,
-        bracket_unc_remove_seps=bracket_unc_remove_seps,
+        paren_uncertainty_separators=paren_uncertainty_separators,
         pm_whitespace=pm_whitespace,
         add_c_prefix=add_c_prefix,
         add_small_si_prefixes=add_small_si_prefixes,
@@ -181,7 +181,7 @@ class GlobalDefaultsContext:
         paren_uncertainty: bool | None = None,
         pdg_sig_figs: bool | None = None,
         val_unc_match_widths: bool | None = None,
-        bracket_unc_remove_seps: bool | None = None,
+        paren_uncertainty_separators: bool | None = None,
         pm_whitespace: bool | None = None,
         add_c_prefix: bool = False,
         add_small_si_prefixes: bool = False,
@@ -209,7 +209,7 @@ class GlobalDefaultsContext:
             paren_uncertainty=paren_uncertainty,
             pdg_sig_figs=pdg_sig_figs,
             val_unc_match_widths=val_unc_match_widths,
-            bracket_unc_remove_seps=bracket_unc_remove_seps,
+            paren_uncertainty_separators=paren_uncertainty_separators,
             pm_whitespace=pm_whitespace,
             add_c_prefix=add_c_prefix,
             add_small_si_prefixes=add_small_si_prefixes,
