+++ src/sciform/format_utils.py
@@ -501,7 +501,7 @@ def construct_val_unc_str(  # noqa: PLR0913
     paren_uncertainty: bool,
     latex: bool,
     pm_whitespace: bool,
-    bracket_unc_remove_seps: bool,
+    paren_uncertainty_separators: bool,
 ) -> str:
     """Construct the value/uncertainty part of the formatted string."""
     if not paren_uncertainty:
@@ -510,25 +510,32 @@ def construct_val_unc_str(  # noqa: PLR0913
             pm_symb = f" {pm_symb} "
         val_unc_str = f"{val_mantissa_str}{pm_symb}{unc_mantissa_str}"
     else:
-        if unc_mantissa.is_finite() and val_mantissa.is_finite():
-            if unc_mantissa == 0:
-                unc_mantissa_str = "0"
-            elif unc_mantissa < abs(val_mantissa):
+        if unc_mantissa.is_finite():
+            if val_mantissa.is_finite() and 0 < unc_mantissa < abs(val_mantissa):
+                """
+                Don't left strip the unc_mantissa_str if val_mantissa is non-finite.
+                Don't left strip the unc_mantissa_str if unc_mantissa == 0 (because then
+                the empty string would remain).
+                Don't left strip the unc_mantissa_str if unc_mantissa >= val_mantissa
+                """
                 unc_mantissa_str = unc_mantissa_str.lstrip("0.,_ ")
-        if bracket_unc_remove_seps:
-            for separator in Separator:
-                if separator == decimal_separator:
-                    continue
-                unc_mantissa_str = unc_mantissa_str.replace(separator, "")
-                # TODO: bracket_unc_remove_seps unit test in tests, not just doctest.
-            if unc_mantissa < abs(val_mantissa):
-                # TODO: I think this raises an error if paren_uncertainty=True but
-                #  either unc_mantissa or val_mantissa is non-finite.
-                # Only removed "embedded" decimal symbol for unc < val
-                unc_mantissa_str = unc_mantissa_str.replace(
-                    decimal_separator,
-                    "",
-                )
+            if not paren_uncertainty_separators:
+                for separator in Separator:
+                    if separator == decimal_separator:
+                        if val_mantissa.is_finite():
+                            if unc_mantissa >= abs(val_mantissa):
+                                """
+                                Don't remove the decimal separator if the uncertainty is
+                                larger than the value.
+                                """
+                                continue
+                        else:
+                            """
+                            Don't remove the decimal separator if the value is 
+                            non-finite.
+                            """
+                            continue
+                    unc_mantissa_str = unc_mantissa_str.replace(separator, "")
         val_unc_str = f"{val_mantissa_str}({unc_mantissa_str})"
     return val_unc_str
 
