+++ tests/test_val_unc_fsml.py
@@ -284,26 +284,6 @@ class TestFormatting(unittest.TestCase):
 
         self.run_val_unc_fsml_cases(cases_list)
 
-    def test_match_width(self):
-        cases_list = [
-            (
-                (123.456, 0.789),
-                [
-                    ("0!2f", "123.46 ±   0.79"),
-                    ("0=0!2f", "123.46 ± 000.79"),
-                ],
-            ),
-            (
-                (0.789, 123.456),
-                [
-                    ("0!2f", "  0 ± 120"),
-                    ("0=0!2f", "000 ± 120"),
-                ],
-            ),
-        ]
-
-        self.run_val_unc_fsml_cases(cases_list)
-
     def test_nan_inf(self):
         cases_list = [
             (
@@ -415,26 +395,44 @@ class TestFormatting(unittest.TestCase):
         ]
         self.run_val_unc_fsml_cases(cases_list)
 
-    def test_match_widths(self):
+    def test_left_pad(self):
         cases_list = [
             (
                 (123.456, 0.789),
                 [
-                    ("0=0", "123.456 ± 000.789"),
-                    ("0= 0", " 123.456 ± 000.789"),
-                    (" 0", " 123.456 ±   0.789"),
-                    ("+0", "+123.456 ±   0.789"),
-                    ("0", "123.456 ±   0.789"),
+                    ("0=0", "123.456 ± 0.789"),
+                    ("0= 0", " 123.456 ± 0.789"),
+                    (" 0", " 123.456 ± 0.789"),
+                    ("+0", "+123.456 ± 0.789"),
+                    ("0", "123.456 ± 0.789"),
+                    ("0!2f", "123.46 ± 0.79"),
+                    ("0=0!2f", "123.46 ± 0.79"),
+                    ("0=4", "00123.456 ± 00000.789"),
+                    ("0= 4", " 00123.456 ± 00000.789"),
+                    (" 4", "   123.456 ±     0.789"),
+                    ("+4", "+  123.456 ±     0.789"),
+                    ("4", "  123.456 ±     0.789"),
+                    ("4!2f", "  123.46 ±     0.79"),
+                    ("0=4!2f", "00123.46 ± 00000.79"),
                 ],
             ),
             (
                 (0.789, 123.456),
                 [
-                    ("0=0", "000.789 ± 123.456"),
-                    ("0= 0", " 000.789 ± 123.456"),
-                    (" 0", "   0.789 ± 123.456"),
-                    ("+0", "+  0.789 ± 123.456"),
-                    ("0", "  0.789 ± 123.456"),
+                    ("0=0", "0.789 ± 123.456"),
+                    ("0= 0", " 0.789 ± 123.456"),
+                    (" 0", " 0.789 ± 123.456"),
+                    ("+0", "+0.789 ± 123.456"),
+                    ("0", "0.789 ± 123.456"),
+                    ("0!2f", "0 ± 120"),
+                    ("0=0!2f", "0 ± 120"),
+                    ("0=4", "00000.789 ± 00123.456"),
+                    ("0= 4", " 00000.789 ± 00123.456"),
+                    (" 4", "     0.789 ±   123.456"),
+                    ("+4", "+    0.789 ±   123.456"),
+                    ("4", "    0.789 ±   123.456"),
+                    ("4!2f", "    0 ±   120"),
+                    ("0=4!2f", "00000 ± 00120"),
                 ],
             ),
         ]
