+++ src/sciform/global_options.py
@@ -24,7 +24,7 @@ PKG_DEFAULT_OPTIONS = RenderedOptions(
     nan_inf_exp=False,
     paren_uncertainty=False,
     pdg_sig_figs=False,
-    val_unc_match_widths=False,
+    left_pad_matching=False,
     paren_uncertainty_separators=True,
     pm_whitespace=True,
 )
