+++ src/sciform/global_configuration.py
@@ -40,7 +40,7 @@ def set_global_defaults(  # noqa: PLR0913
     nan_inf_exp: bool | None = None,
     paren_uncertainty: bool | None = None,
     pdg_sig_figs: bool | None = None,
-    val_unc_match_widths: bool | None = None,
+    left_pad_matching: bool | None = None,
     paren_uncertainty_separators: bool | None = None,
     pm_whitespace: bool | None = None,
     add_c_prefix: bool = False,
@@ -73,7 +73,7 @@ def set_global_defaults(  # noqa: PLR0913
         nan_inf_exp=nan_inf_exp,
         paren_uncertainty=paren_uncertainty,
         pdg_sig_figs=pdg_sig_figs,
-        val_unc_match_widths=val_unc_match_widths,
+        left_pad_matching=left_pad_matching,
         paren_uncertainty_separators=paren_uncertainty_separators,
         pm_whitespace=pm_whitespace,
         add_c_prefix=add_c_prefix,
@@ -180,7 +180,7 @@ class GlobalDefaultsContext:
         nan_inf_exp: bool | None = None,
         paren_uncertainty: bool | None = None,
         pdg_sig_figs: bool | None = None,
-        val_unc_match_widths: bool | None = None,
+        left_pad_matching: bool | None = None,
         paren_uncertainty_separators: bool | None = None,
         pm_whitespace: bool | None = None,
         add_c_prefix: bool = False,
@@ -208,7 +208,7 @@ class GlobalDefaultsContext:
             nan_inf_exp=nan_inf_exp,
             paren_uncertainty=paren_uncertainty,
             pdg_sig_figs=pdg_sig_figs,
-            val_unc_match_widths=val_unc_match_widths,
+            left_pad_matching=left_pad_matching,
             paren_uncertainty_separators=paren_uncertainty_separators,
             pm_whitespace=pm_whitespace,
             add_c_prefix=add_c_prefix,
