### 2023-12-27 jagerber48/sciform#73 @jagerber48
One proposed desired behavior would be:

- When formatting a value (on its own with no uncertainty) then `ndigits` is always respected and the state of `pdg_sig_figs` is ignored.
- When formatting value/uncertainty pairs then `pdg_sig_figs` is always respected, however, if `ndigits != AutoDigits` then a warning is raised.
