### 2023-12-27 8fbd151
fill_mode -> fill_char
