+++ src/sciform/global_configuration.py
@@ -28,7 +28,7 @@ def set_global_defaults(  # noqa: PLR0913
     decimal_separator: modes.UserDecimalSeparators | None = None,
     lower_separator: modes.UserLowerSeparators | None = None,
     sign_mode: modes.UserSignMode | None = None,
-    fill_mode: modes.UserFillMode | None = None,
+    fill_char: modes.UserFillMode | None = None,
     left_pad_dec_place: int | None = None,
     exp_format: modes.UserExpFormat | None = None,
     extra_si_prefixes: dict[int, str] | None = None,
@@ -61,7 +61,7 @@ def set_global_defaults(  # noqa: PLR0913
         decimal_separator=decimal_separator,
         lower_separator=lower_separator,
         sign_mode=sign_mode,
-        fill_mode=fill_mode,
+        fill_char=fill_char,
         left_pad_dec_place=left_pad_dec_place,
         exp_format=exp_format,
         extra_si_prefixes=extra_si_prefixes,
@@ -168,7 +168,7 @@ class GlobalDefaultsContext:
         decimal_separator: modes.UserDecimalSeparators | None = None,
         lower_separator: modes.UserLowerSeparators | None = None,
         sign_mode: modes.UserSignMode | None = None,
-        fill_mode: modes.UserFillMode | None = None,
+        fill_char: modes.UserFillMode | None = None,
         left_pad_dec_place: int | None = None,
         exp_format: modes.UserExpFormat | None = None,
         extra_si_prefixes: dict[int, str] | None = None,
@@ -196,7 +196,7 @@ class GlobalDefaultsContext:
             decimal_separator=decimal_separator,
             lower_separator=lower_separator,
             sign_mode=sign_mode,
-            fill_mode=fill_mode,
+            fill_char=fill_char,
             left_pad_dec_place=left_pad_dec_place,
             exp_format=exp_format,
             extra_si_prefixes=extra_si_prefixes,
