+++ src/sciform/global_options.py
@@ -12,7 +12,7 @@ PKG_DEFAULT_OPTIONS = RenderedOptions(
     decimal_separator=modes.Separator.POINT,
     lower_separator=modes.Separator.NONE,
     sign_mode=modes.SignMode.NEGATIVE,
-    fill_mode=modes.FillMode.SPACE,
+    fill_char=modes.FillMode.SPACE,
     left_pad_dec_place=0,
     exp_format=modes.ExpFormat.STANDARD,
     extra_si_prefixes={},
