### 2023-12-27 9206977
FillMode -> FillChar
