+++ src/sciform/formatter.py
@@ -51,7 +51,7 @@ class Formatter:
         decimal_separator: modes.UserDecimalSeparators | None = None,
         lower_separator: modes.UserLowerSeparators | None = None,
         sign_mode: modes.UserSignMode | None = None,
-        fill_char: modes.UserFillMode | None = None,
+        fill_char: modes.UserFillChar | None = None,
         left_pad_dec_place: int | None = None,
         exp_format: modes.UserExpFormat | None = None,
         extra_si_prefixes: dict[int, str] | None = None,
