### 2023-12-28 64ede5a
Under construction notes, add pyopensci reference
