### 2023-12-28 #121 @jagerber48
@isabelizimm ok yes, `pre-commit --no-verify` was the trick to make me comfortable setting up `pre-commit` and adding `pre-commit` instructions. Fortunately git kraken (the tool I use for git) has an easy button to do the `--no-verify` option, so this is all very convenient for me.

Ok I'm working on another release (`0.31.0`) that addresses all of the remaining concerns in the review except the lazy formatting `FormattedValue` suggestion. Here's the PR for the release https://github.com/jagerber48/sciform/pull/96. There are a number of notable changes you can find in the changelog, but the highlights are:

* Remove `SciNumUnc` class, now `SciNum` class accepts optional second input for value/uncertainty formatted
* No longer possible to configure separators using FSML, this just made the FSML too complicated. This made it easier to write annotated FSML examples.
* Many options were renamed. Notably the unintuitive and ugly `unc` abbreviation has been eliminated. See changelog for all the changes.
* Linted/reformatted code using ruff. This makes the diff huge

Anyways, if any of you are inclined, @Batalex @machow  @isabelizimm I would appreciate if any of you have comments on this jump from `0.30.1` to `0.31.0` before I release it. Otherwise, I'm curious about what next steps on the review look like.

There are a couple improvements I see that can be made, but I consider them to be typical package improvements that should happen over time and I don't know that they have bearing on the review.

* https://github.com/jagerber48/sciform/issues/73 Resolving this issue would improve the interface
* I have a todo item to review the behaviors of the options/functions to add e.g. the "centi-" prefix.
* Look at implementing the `FormattedValue` lazy formatting suggestion
* Make `Formatter` accept `SciNum` input
* Add functionality to print out options from the `Formatter` object.
