### 2023-12-29 jagerber48/sciform#97 @jagerber48
The behavior/semantics of extra prefixes may be confusing.
Consider

```
from sciform import Formatter, GlobalDefaultsContext

with GlobalDefaultsContext(add_c_prefix=True):
    sform = Formatter(
        exp_mode="scientific",
        exp_val=2,
        extra_si_prefixes={1: 'da'},
        exp_format='prefix',
    )
    print(sform(123))
# 1.23e+02

sform = Formatter(
    exp_mode="scientific",
    exp_val=2,
    extra_si_prefixes={1: 'da'},
    exp_format='prefix',
)

with GlobalDefaultsContext(add_c_prefix=True):
    print(sform(123))
# 1.23e+02
```

In both cases the `add_c_prefix` does not have the effect of configuring the formatter to use the `c` prefix. This is because the formatter includes input for `extra_si_prefixes`. The current behavior is that, like all other options, the global `extra_si_prefixes` is used only if the local `extra_si_prefixes=None`.

An alternative, possibly expected, behavior would be that the local `extra_si_prefixes` are always merged into the global `extra_si_prefixes` at format time.
