+++ README.rst
@@ -64,7 +64,8 @@ Under Construction
 The ``sciform`` package is still in early stages of development.
 The API is not stable.
 Class, function and parameter names and usages have undergone changes
-and may continue to change until version ``1.0.0`` is released.
+and may continue to change relatively freely until version ``1.0.0`` is
+released.
 API changes will be announced after new releases in the
 `changelog <https://sciform.readthedocs.io/en/stable/project.html#changelog>`_.
 If you have an idea or opinion about how ``sciform`` should be designed,
