### 2024-01-03 #121 @jagerber48
@Batalex great! Thank you!

I've been asking myself the `1.0.0` question lately as well. It's definitely still my goal to get `1.0.0` released. This pyopensci review has been the main thing I've been trying to check off on my release `1.0.0` list, but now that this is wrapping up I need to step back and think about if there is anything else.

* There are definitely a couple minor changes to behavior I have planned, but I'm not sure if these need to block `1.0.0`, I need to list these out.
* The deferred formatting for html and latex may result in removing the `latex` option. That makes me tempted to remove the `latex` option from the user-facing interface for now so that there's not a regression when that feature is deprecated in favor of the deferred formatting.
* I need to get myself a little bit more convinced that the names of everything are good. I'm feeling better about it after this last round of renamings.
