+++ CHANGELOG.rst
@@ -9,7 +9,7 @@ Unreleased
 
 ----
 
-0.31.0 (2023-12-28)
+0.31.0 (2024-01-06)
 -------------------
 
 Removed
