### 2024-01-07 #121 @jagerber48
I just released version `0.31.1` which implements all of the changes discussed above.

My current action is compiling my `sciform` to do list from a few different sources, then I'll know what steps are still needed for `1.0.0`.
