### 2024-01-08 #121 @jagerber48
Ok, I got my thoughts collected. Previously I had to do lists in various places. Now most of the todos are captured as tagged (labeled) [issues](https://github.com/jagerber48/sciform/issues). Most of the issues get the `Status: To Investigate` label which means I need to think about whether I want to do anything about that issue or not.

However, two of the issues got tagged for the [Release 1.0.0 milestone](https://github.com/jagerber48/sciform/milestone/1). They both correspond to what I would consider to be either strange or broken behavior for a couple of options. I'll just link the issues here. I would appreciate any thoughts on the issues that anyone here has.

* https://github.com/jagerber48/sciform/issues/73
* https://github.com/jagerber48/sciform/issues/97

The latter one especially might result in, e.g. removing the global settings functions like `global_add_c_prefix`.

The resolution to this issue

* https://github.com/jagerber48/sciform/issues/99

May also constitute a small breaking change to the behavior.

Regarding naming: I'm pretty happy with where things are at at the moment. There's maybe a couple names I don't love (e.g. `extra_parts_per_forms`) but I think the really ugly ones have been cleaned up.

I do have some internal debates about what `1.0.0` should mean for `sciform`. I can imagine wanting to make small breaking changes, such as https://github.com/jagerber48/sciform/issues/99 after `1.0.0` is released. Do I really need to bump from `1.0.0` to `2.0.0` because the formatting output is slightly changed for some very specific combination of inputs? Or even if I remove a function like `global_add_c_prefix` that no one is likely using? And even if they are (1) it's probably not super critical since it's just formatting results for humans to read and (2) the fix will be very easy. If I take the conservative approach suggested by answers to [Is a small breaking change enough to do a Major Version Upgrade while using SemVer?](https://stackoverflow.com/questions/71813328/is-a-small-breaking-change-enough-to-do-a-major-version-upgrade-while-using-semv) then the answer is yes, I should major version bump. But if this is the approach I take then I'll want to be even more conservative before releasing `1.0.0` and it will take even longer to get it out!

My feeling is that given (1) `sciform` is new and small and I'm not sure if anyone is even really using it yet, (2) the nature of `sciform` as a being an endpoint for making text human readable and (3) the scope of the planned changes, I think it would be reasonable to take a more liberal interpretation of SemVer major version bumps to be for changes that majorly breaking... but I would be open to pushback here.

@Batalex @isabelizimm @machow curious for your feedback on

* The issues linked above (preferably in the issues themselves
* What philosophy I should take towards releasing `sciform` at `1.0.0` (maybe I should just resolve the three issues linked above which are the only open breaking issues before `1.0.0`)
* Any other thoughts on the state of `sciform` re this PyOpenSci review.
