### 2024-01-09 jagerber48/sciform#73 @jagerber48
I'm going to move forward with the strategy discussed in the last post. But there will be no warning about the possibly confusing combination of options. Rather, it will simply be clearly stated in the documentation for `pdg_sig_figs` that `pdg_sig_figs=True` takes precedence over `ndigits` when formatting value/uncertainty pairs.
