### 2024-01-10 jagerber48/sciform#73 @jagerber48
Taking the suggested approach will in fact not be breaking. I was worried that `pdg_sig_figs=True` while formatting a value resulted in the value being rounded according to pdg sig fig rounding rules but I was thinking about https://github.com/jagerber48/sciform/issues/71 in which the value **of a value/uncertainty pair** was erroneously getting rounded according to pdg sig fig rounding rules. The current behavior with `pdg_sig_figs=True` and `ndigits=AutoDigits` is that formatting a value alone result in the value being rounded according to `AutoDigits`.

Instead the proposed change is simply going to **add** the functionality for the formatter to be configured with `pdg_sig_figs=True` and `ndigits!=AutoDigits`, a case which currently always just raises an error.
