### 2024-01-10 31f436b
modify get_round_digit logic and remove check during options construction
