### 2024-01-10 b142d5b
remove now invalid tests checking for pdg_sig_figs=True and ndigits!=AutoDigits
