### 2024-01-10 6b7319c
remove global_add_c_prefix etc. from tests
