### 2024-01-10 705d7fa
remove global_add_c_prefix etc. from global_configuration.py
