### 2024-01-10 ee5bb1c
add clarification that helper flags don't overwrite values in extras dicts
