### 2024-01-10 cfdee25
documentation about the interaction with global settings
