+++ tests/test_config.py
@@ -72,3 +72,21 @@ class TestConfig(unittest.TestCase):
         with GlobalDefaultsContext(add_ppth_form=True):
             self.assertEqual(formatter(num), "2.4 ppth")
         self.assertEqual(formatter(num), "2.4e-03")
+
+    def test_add_c_prefix_no_overwrite(self):
+        sform = Formatter(
+            exp_mode="scientific",
+            exp_format="prefix",
+            extra_si_prefixes={-2: "zzz"},
+            add_c_prefix=True,
+        )
+        self.assertEqual(sform(0.012), "1.2 zzz")
+
+    def test_global_add_c_prefix_no_overwrite(self):
+        sform = Formatter(
+            exp_mode="scientific",
+            exp_format="prefix",
+            extra_si_prefixes={-4: "zzz"},
+        )
+        with GlobalDefaultsContext(add_c_prefix=True):
+            self.assertEqual(sform(0.012), "1.2e-02")
