### 2024-01-10 bcc100b
rework local/global merging counter-example including doctest examples
