### 2024-01-10 jagerber48/sciform#99 @jagerber48
Yes this change should be implemented as described.
