### 2024-01-10 d3057ec
document 10^0
