+++ CHANGELOG.rst
@@ -5,6 +5,11 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 Unreleased
 ----------
 
+* There are no unreleased changes
+
+0.32.0 (2024-01-10)
+-------------------
+
 Added
 ^^^^^
 
