### 2024-01-10 #121 @jagerber48
Addressed the 3 issues linked in the previous post to release [0.32.0](https://github.com/jagerber48/sciform/releases/tag/0.32.0).

That's all the concrete stuff I had on my list before releasing `1.0.0`. Open to suggestions on when I should **actually** bump to `1.0.0`. I'm curious to sit with it a little longer myself and see how I feel about it. I may also try to advertise to others and try to get a little more feedback from more people who actually download and try to use the code for some real application. My belief/hope is that the core functionality should actually be pretty solid and has been for a while, and the issues I've been thinking about lately are pretty fringe edge cases that I only know about because I'm so deep in the code.
