### 2024-01-10 jagerber48/sciform#114 @jagerber48
This change may likely remove the `latex` `Formatter` option in favor of handling latex conversion downstream in the `FormattedValue` class.
