### 2024-01-11 jagerber48/sciform#114 @jagerber48
One issue with this pattern is that right now the output of the `Formatter` is a string, and people might use it as a string. One example I can think of offhand would be appending units

```
from sciform import Formatter

sform = Formatter(exp_mode="engineering", exp_format="prefix")
num_str = sform(12345)
units = "Hz"
unit_str = f'{num_str}{units}'
print(unit_str)
# 12.346 kHz
```

Would this still work as expected? Does passing an object into an f string formatter always cast it to a string? What about

```
from sciform import Formatter

sform = Formatter(exp_mode="engineering", exp_format="prefix")
num_str = sform(12345)
units = "Hz"
unit_str = num_str + units
print(unit_str)
# 12.346 kHz
```
