### 2024-01-11 #121 @machow
> What philosophy I should take towards releasing sciform at 1.0.0. ... I may also try to advertise to others and try to get a little more feedback from more people who actually download and try to use the code for some real application.

Regardless of when you bump to v1, I'd definitely +1 advertising it broadly to folks! I'd think about it like this: if you bump to v1 without people using it, you have to guess what the right behaviors / naming is. If you can get people using it, even with the caveat it's v0, you'll have good feedback on potential changes for a v1. Because v0 explicitly signals you might make changes, in a way you're in a good position to advertise it AND think about important tweaks!
