### 2024-01-11 #121 @jagerber48
> Regardless of when you bump to v1, I'd definitely +1 advertising it broadly to folks! I'd think about it like this: if you bump to v1 without people using it, you have to guess what the right behaviors / naming is. If you can get people using it, even with the caveat it's v0, you'll have good feedback on potential changes for a v1. Because v0 explicitly signals you might make changes, in a way you're in a good position to advertise it AND think about important tweaks!

Yes, this is exactly what I'm thinking! I don't want to have to guess at right behaviors / naming which is what it feels like I would be doing if I bumped to v1.

I don't think v0 is at all an impediment to getting people to use it. That is, I think the fractional intersection of people who might be interested in `sciform` and people who wouldn't use `sciform` because it is <v1 is very small. I think the bigger impediments are getting people to be aware of `sciform` (advertising) and whether `sciform` is actually broadly useful to people once they see it.

So if that's really true maybe I should just cool my horses on v1 and instead commit to something like (and indicate this in the docs) "I'm going to use and advertise `sciform` for N(=12?) months during which time I'll collect feedback. At the end of that time I'll revisit the v1 question".
