+++ docs/source/examples.rst
@@ -289,3 +289,52 @@ visualizations.
 The relative scaling of parameters between datasets and the relative
 scaling between the value and uncertainty for each entry are
 immediately clear.
+
+Mapping Over Collections
+========================
+
+It may be convenient to apply ``sciform`` mapping to collections (sets,
+lists, arrays, etc.) of numbers.
+
+>>> from sciform import Formatter
+>>>
+>>> sform = Formatter(
+...     exp_mode="engineering",
+...     exp_format="prefix",
+...     pdg_sig_figs=True,
+...     paren_uncertainty=True,
+... )
+>>> val_list = [1000, 2000, 3000]
+>>> err_list = [200, 400, 600]
+>>>
+>>> val_str_list = list(map(sform, val_list))
+>>> print(val_str_list)
+['1 k', '2 k', '3 k']
+>>>
+>>> val_err_str_list = list(map(sform, val_list, err_list))
+>>> print(val_err_str_list)
+['1.00(20) k', '2.0(4) k', '3.0(6) k']
+
+We may also want to map over higher dimensional numpy arrays.
+`numpy.vectorize() <https://numpy.org/doc/stable/reference/generated/numpy.vectorize.html>`_
+makes this easy.
+
+>>> import numpy as np
+>>>
+>>> vec_sform = np.vectorize(sform)
+>>> arr = np.array([[1e6, 2e6, 3e6],
+...                 [4e6, 5e6, 6e6],
+...                 [7e6, 8e6, 9e6]])
+>>>
+>>> arr_err = np.array([[9e4, 8e4, 7e4],
+...                     [6e4, 5e4, 4e4],
+...                     [3e4, 2e4, 1e4]])
+>>>
+>>> print(vec_sform(arr))
+[['1 M' '2 M' '3 M']
+ ['4 M' '5 M' '6 M']
+ ['7 M' '8 M' '9 M']]
+>>> print(vec_sform(arr, arr_err))
+[['1.00(9) M' '2.00(8) M' '3.00(7) M']
+ ['4.00(6) M' '5.00(5) M' '6.00(4) M']
+ ['7.000(30) M' '8.000(20) M' '9.000(10) M']]
