### 2024-01-12 8a9e1a2
Remove roadmap and clean up project.rst
