+++ CHANGELOG.rst
@@ -1,3 +1,6 @@
+Changelog
+=========
+
 This project adheres to `Semantic Versioning <https://semver.org/>`_.
 
 ----
@@ -13,6 +16,7 @@ Added
   Changes included adding links to the ``sciform`` feedback survey.
 * Added examples in the documentation demonstrating how ``sciform``
   formatting can be mapped over collections of numbers.
+  [`#120 <https://github.com/jagerber48/sciform/issues/120>`_]
 
 Changed
 ^^^^^^^
