### 2024-01-12 #121 @jagerber48
Before going into my thoughts/questions about advertising/getting feedback: What is the status of the pyopensci review at this point? I greatly appreciate the continued feedback I'm getting but also don't want to veer the conversation out of scope or off topic. @Batalex? My rough thinking on `1.0.0` is that I don't want to do that until I get feedback. I imagine it will **at least** take weeks if not months to get some valuable feedback. The PyOpenSci review is a major milestone towards `1.0.0`, but I don't think it gets me all the way there, so I think `1.0.0` should be decoupled from the review at this point.

____

Suggestions for advertising `sciform` to get users and get feedback? Here's my plan so far

- I've added a link to a google forms survey where I can gather feedback on the readme which shows up on the main page of the documentation: https://sciform.readthedocs.io/en/stable/
- I'm going through relevant stack overflow questions that sciform pertains to (mainly stuff like "how to format numbers using engineering notation in python" or "how to round numbers based on sig figs in python") and answering them, providing links to sciform.
- After this review is accepted I may request the pyopensci announcement to include a note requesting feedback if anyone wants to check it out and give feedback, or I can add that note on myself in slack.
- I was working on something like `sciform` a long time ago, and then [this thread](https://discuss.python.org/t/new-format-specifiers-for-string-formatting-of-floats-with-si-and-iec-prefixes/26914/20) came up. The upshot was that there should be a PyPi package to to do the type of formatting requested. This gave me the motivation to beef what I was working on into the full package `sciform`. I've already linked `sciform` in that thread. Considering bumping that thread with the update that `sciform` does a lot of what was requested and is more mature now. **HOWEVER I don't know if this would be an appropriate advertisement or if it would be seen as bumping a thread for too much self promotion**.
    - A slight side note: There is still a part of me that is interested in getting some of the `sciform` features into the built in format specification mini language. Notably I think using `!` to indicate reproducible sig fig rounding and using an `r` format type to indicate engineering notation would be highly impactful for scientists and I think could be compatible *extensions* to the built in FSML. (Noting that the `sciform` FSML is not strictly an extension of the python FSML). Were I to bump that thread I would probably also include a statement of my ideas in this regard as well.
- I am not active on `r/python` but it seems that that could be a natural place to post "Hey here is my open source package, what do people think?"
- I will likely share with my former research group and maybe some other precision metrology groups I know who may be interested in this package.

How do these ideas look? Where else can/should open source packages be announced?
