### 2024-01-12 #121 @jagerber48
Another breaking variable rename recently occurred to me while I was answering a stack overflow question
```
fill_char -> left_pad_char
fill_char -> left_pad_fill_char
```
https://github.com/jagerber48/sciform/issues/126

This would be much more consistent with its option partners `left_pad_dec_place` and `left_pad_matching`. A more concise name that retain the consistency would be `left_pad_char`.
