### 2024-01-13 jagerber48/sciform#114 @Batalex
Yes, using f-strings would cast the result as a string. You might be onto something with the `+` operator, though. `Formatter` could define `__add__` and `__radd__` so that it can append units.
