### 2024-01-13 #121 @Batalex
Hey @jagerber48,
If you are set on us finishing this review before releasing `1.0.0` so that it come after you gather feedback, that is fine by me.

@machow, @isabelizimm I need you to check

```
Final approval (post-review)

- [  ] The author has responded to my review and made changes to my satisfaction. I recommend approving this package.
```
in your reviews above.

As for how to advertise `sciform` and how to gather feedback:
- pyopensci's announcement is a great place to start
- r/python as well
- you could try to reach the folks @ realpython to include `sciform` in their newsletter "Pycoders"
- there might be specialized discourse for engineering with Python?
