### 2024-01-14 jagerber48/sciform#114 @jagerber48
@Batalex ok, I've thought about this a little bit. The main thing that I like about it is that it provides hooks for `__repr_html__` that can be used for `jupyter` or `__repr_latex__` that could be used to send strings to Latex.

However, I don't think those things require deferred formatting. Consider

```
class Formatter:
    ...

    def __call__(
        self: Formatter,
        value: Number,
        uncertainty: Number | None = None,
        /,
    ) -> str:
        ...
        rendered_options = self._user_options.render()
        if uncertainty is None:
            output = format_num(Decimal(str(value)), rendered_options)
        else:
            output = format_val_unc(
                Decimal(str(value)),
                Decimal(str(uncertainty)),
                rendered_options,
            )
        return FormattedValue(output)

class FormattedValue(str):
    def __repr_html__(self):
        return f'<b>{self}</b>'

    def __repr_latex__(self):
        return fr'\text{{{self}}}'
```

Of course the `__repr_html__` and `__repr_latex__` functions would be built out to something meaningful.

So @Batalex, does this realize the main thrust of this recommendation? Or was the deferred formatting actually critical to the suggestion?
