### 2024-01-14 jagerber48/sciform#114 @Batalex
I proposed the lazy formatting so that special characters would be rendered appropriately:
> \textsuperscript in a LaTeX document or `<sup></sup>` in a notebook

Your suggestion above could work; it seems more complex to replace those characters afterward compared to the following approach

```py
class FormattedValue(str):

    def __init__(self, fmt: Formatter, value: Number, uncertainty: Optional[Number] = None, /) -> None:
        self.fmt = fmt
        self.value = value
        self.uncertainty = uncertainty

    def __str__(self) -> str:
        rendered_options = self.fmt.user_options.render()
        if self.uncertainty is None:
            return format_num(Decimal(str(self.value)), rendered_options)
        else:
            return format_val_unc(Decimal(str(self.value)),
                                  Decimal(str(self.uncertainty)),
                                  rendered_options)

    def _repr_html_(self):
        rendered_options = self.fmt.user_options.render()
        rendered_options.html = True
        if self.uncertainty is None:
            return format_num(Decimal(str(self.value)), rendered_options)
        else:
            return format_val_unc(Decimal(str(self.value)),
                                  Decimal(str(self.uncertainty)),
                                  rendered_options)

    def _repr_latex_(self):
        rendered_options = self.fmt.user_options.render()
        rendered_options.latex = True
        if self.uncertainty is None:
            return format_num(Decimal(str(self.value)), rendered_options)
        else:
            return format_val_unc(Decimal(str(self.value)),
                                  Decimal(str(self.uncertainty)),
                                  rendered_options)

```

(without all the copy paste)
