### 2024-01-16 #121 @isabelizimm
Hey there-- I have given a ✅. The plans and notes above for advertising seem like a solid start; if you're involved in other Python discord/slack/etcs, there's often a place to show off work. ++ on v1 being a personal choice, but that it is never too early to solicit feedback!

This package is in a solid state, and @jagerber48 should be really excited and proud of building it up!
