### 2024-01-19 951a01c
fill_char -> left_pad_char
