+++ src/sciform/format_utils.py
@@ -338,11 +338,11 @@ def get_round_digit(
     return round_digit
 
 
-def get_fill_str(fill_char: str, top_digit: int, top_padded_digit: int) -> str:
+def get_fill_str(left_pad_char: str, top_digit: int, top_padded_digit: int) -> str:
     """Get the string filling from top_digit place to top_padded_digit place."""
     if top_padded_digit > top_digit:
         pad_len = top_padded_digit - max(top_digit, 0)
-        pad_str = fill_char * pad_len
+        pad_str = left_pad_char * pad_len
     else:
         pad_str = ""
     return pad_str
@@ -353,7 +353,7 @@ def format_num_by_top_bottom_dig(
     target_top_digit: int,
     target_bottom_digit: int,
     sign_mode: SignModeEnum,
-    fill_char: str,
+    left_pad_char: str,
 ) -> str:
     """Format a number according to specified top and bottom digit places."""
     print_prec = max(0, -target_bottom_digit)
@@ -362,7 +362,7 @@ def format_num_by_top_bottom_dig(
     sign_str = get_sign_str(num, sign_mode)
 
     num_top_digit = get_top_digit(num)
-    fill_str = get_fill_str(fill_char, num_top_digit, target_top_digit)
+    fill_str = get_fill_str(left_pad_char, num_top_digit, target_top_digit)
     return f"{sign_str}{fill_str}{abs_mantissa_str}"
 
 
