+++ tests/test_print.py
@@ -19,7 +19,7 @@ class TestPrint(unittest.TestCase):
             " 'decimal_separator': '.',\n"
             " 'lower_separator': '',\n"
             " 'sign_mode': '-',\n"
-            " 'fill_char': ' ',\n"
+            " 'left_pad_char': ' ',\n"
             " 'left_pad_dec_place': 0,\n"
             " 'exp_format': 'standard',\n"
             " 'extra_si_prefixes': {},\n"
@@ -52,7 +52,7 @@ class TestPrint(unittest.TestCase):
             " 'decimal_separator': '.',\n"
             " 'lower_separator': '',\n"
             " 'sign_mode': '-',\n"
-            " 'fill_char': ' ',\n"
+            " 'left_pad_char': ' ',\n"
             " 'left_pad_dec_place': 3,\n"
             " 'exp_format': 'standard',\n"
             " 'extra_si_prefixes': {},\n"
