+++ src/sciform/global_configuration.py
@@ -28,7 +28,7 @@ def set_global_defaults(  # noqa: PLR0913
     decimal_separator: modes.DecimalSeparators | None = None,
     lower_separator: modes.LowerSeparators | None = None,
     sign_mode: modes.SignMode | None = None,
-    fill_char: modes.FillChar | None = None,
+    left_pad_char: modes.LeftPadChar | None = None,
     left_pad_dec_place: int | None = None,
     exp_format: modes.ExpFormat | None = None,
     extra_si_prefixes: dict[int, str] | None = None,
@@ -61,7 +61,7 @@ def set_global_defaults(  # noqa: PLR0913
         decimal_separator=decimal_separator,
         lower_separator=lower_separator,
         sign_mode=sign_mode,
-        fill_char=fill_char,
+        left_pad_char=left_pad_char,
         left_pad_dec_place=left_pad_dec_place,
         exp_format=exp_format,
         extra_si_prefixes=extra_si_prefixes,
@@ -113,7 +113,7 @@ class GlobalDefaultsContext:
         decimal_separator: modes.DecimalSeparators | None = None,
         lower_separator: modes.LowerSeparators | None = None,
         sign_mode: modes.SignMode | None = None,
-        fill_char: modes.FillChar | None = None,
+        left_pad_char: modes.LeftPadChar | None = None,
         left_pad_dec_place: int | None = None,
         exp_format: modes.ExpFormat | None = None,
         extra_si_prefixes: dict[int, str] | None = None,
@@ -141,7 +141,7 @@ class GlobalDefaultsContext:
             decimal_separator=decimal_separator,
             lower_separator=lower_separator,
             sign_mode=sign_mode,
-            fill_char=fill_char,
+            left_pad_char=left_pad_char,
             left_pad_dec_place=left_pad_dec_place,
             exp_format=exp_format,
             extra_si_prefixes=extra_si_prefixes,
