+++ src/sciform/global_options.py
@@ -12,7 +12,7 @@ PKG_DEFAULT_OPTIONS = RenderedOptions(
     decimal_separator=modes.SeparatorEnum.POINT,
     lower_separator=modes.SeparatorEnum.NONE,
     sign_mode=modes.SignModeEnum.NEGATIVE,
-    fill_char=modes.FillCharEnum.SPACE,
+    left_pad_char=modes.LeftPadCharEnum.SPACE,
     left_pad_dec_place=0,
     exp_format=modes.ExpFormatEnum.STANDARD,
     extra_si_prefixes={},
