+++ CHANGELOG.rst
@@ -12,6 +12,7 @@ Changed
 ^^^^^^^
 
 * **[BREAKING]** Renamed ``fill_char`` to ``left_pad_char``.
+  [`#126 <https://github.com/jagerber48/sciform/issues/126>`_]
 
 ----
 
