### 2024-01-21 3ad31ac
cleanup to exponent string resolution code and fix bug
