+++ .github/workflows/python-package.yml
@@ -41,5 +41,9 @@ jobs:
         pytest --cov --doctest-glob="*.rst"
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v3
-      env:
-        CODECOV_TOKEN: ${{ secrets.CODECOV_TOKEN }}
+      with:
+        token: ${{ secrets.CODECOV_TOKEN }}
+        files: ./coverage.xml
+        flags: ${{ matrix.python-version }}
+        fail_ci_if_error: true
+        verbose: true
