+++ .github/workflows/python-package.yml
@@ -43,7 +43,7 @@ jobs:
       uses: codecov/codecov-action@v3
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
-        files: ./coverage.xml
+        files: coverage.xml
         flags: ${{ matrix.python-version }}
         fail_ci_if_error: true
         verbose: true
