### 2024-01-21 be27537
latex output conversion
