+++ src/sciform/__init__.py
@@ -8,6 +8,7 @@ from sciform.global_configuration import (
     set_global_defaults,
 )
 from sciform.modes import AutoDigits, AutoExpVal
+from sciform.output_conversion import sciform_to_latex
 from sciform.scinum import SciNum
 
 __all__ = [
@@ -18,5 +19,6 @@ __all__ = [
     "set_global_defaults",
     "AutoDigits",
     "AutoExpVal",
+    "sciform_to_latex",
     "SciNum",
 ]
