### 2024-01-22 f7ad328
no \left or \right
