+++ tests/test_latex_conversion.py
@@ -21,21 +21,22 @@ class TestLatexConversion(unittest.TestCase):
             ("6.26070e-04", r"6.26070\times10^{-4}"),
             (
                 "(0.000000(1.234560))e+02",
-                r"\left(0.000000\left(1.234560\right)\right)\times10^{2}",
+                r"(0.000000(1.234560))\times10^{2}",
             ),
             ("000_000_004_567_899.765_432_1", r"000\_000\_004\_567\_899.765\_432\_1"),
-            ("(nan)%", r"\left(\text{nan}\right)\%"),
+            ("(nan)%", r"(\text{nan})\%"),
             ("123000 ppm", r"123000\:\text{ppm}"),
             ("0b+00", r"0\times2^{0}"),
             ("16.18033E+03", r"16.18033\times10^{3}"),
             ("    1.20e+01", r"\:\:\:\:1.20\times10^{1}"),
-            ("(-INF)E+00", r"\left(-\text{INF}\right)\times10^{0}"),
+            ("(-INF)E+00", r"(-\text{INF})\times10^{0}"),
             (
                 "(0.123456(789))e+03",
-                r"\left(0.123456\left(789\right)\right)\times10^{3}",
+                r"(0.123456(789))\times10^{3}",
             ),
             ("  123.46 ±     0.79", r"\:\:123.46\:\pm\:\:\:\:\:0.79"),
-            ("(7.8900 ± 0.0001)×10²", r"\left(7.8900\:\pm\:0.0001\right)\times10^{2}"),
+            ("(7.8900 ± 0.0001)×10²", r"(7.8900\:\pm\:0.0001)\times10^{2}"),
+            ("(0.123456 ± 0.000789) k", r"(0.123456\:\pm\:0.000789)\:\text{k}"),
         ]
 
         self.run_latex_conversion_cases(cases_list)
