### 2024-01-22 8d4cf02
Remove old latex functionality
