### 2024-01-22 4c26d89
docstrings and minor reorganization
