+++ src/sciform/format_utils.py
@@ -154,7 +154,7 @@ def get_standard_exp_str(base: int, exp_val: int, *, capitalize: bool = False) -
 
 
 def get_superscript_exp_str(base: int, exp_val: int) -> str:
-    """Get superscript (e.g. '×10⁺²') exponent string."""  # noqa: RUF002
+    """Get superscript (e.g. '×10⁺²') exponent string."""
     sup_trans = str.maketrans("+-0123456789", "⁺⁻⁰¹²³⁴⁵⁶⁷⁸⁹")
     exp_val_str = f"{exp_val}".translate(sup_trans)
     return f"×{base}{exp_val_str}"
