### 2024-01-22 #121 @jagerber48
> If you are set on us finishing this review before releasing 1.0.0 so that it comes after you gather feedback, that is fine by me.

To be clear, yes, I don't want to wait for 1.0.0 before the review is finished.

I still have open questions about various idea/features etc. But at this point I think they're out of scope for this review and should be treated as typical issues for the `sciform` package. I may move on to asking them in pyopensci discourse or slack (and elsewhere on the internet) if that makes sense. Does that sound right?
