### 2024-01-23 9ca141a
docstring ruff formatting
