+++ pyproject.toml
@@ -63,6 +63,9 @@ ignore = [
     'RET504',
 ]
 
+[tool.ruff.format]
+docstring-code-format = true
+
 [tool.ruff.per-file-ignores]
 "tests/*.py" = ["D", "ANN", "PT"]
 "examples/*.py" = ["D"]
