+++ tests/test_html_conversion.py
@@ -2,16 +2,16 @@ from __future__ import annotations
 
 import unittest
 
-from sciform import Formatter, sciform_to_latex
+from sciform import Formatter, sciform_to_html
 
 ValFormatterCases = list[tuple[float, list[tuple[Formatter, str]]]]
 ValUncFormatterCases = list[tuple[tuple[float, float], list[tuple[Formatter, str]]]]
 
 
-class TestLatexConversion(unittest.TestCase):
+class TestHTMLConversion(unittest.TestCase):
     def run_direct_conversions(self, cases_list: list[tuple[str, str]]):
         for input_str, expected_str in cases_list:
-            converted_str = sciform_to_latex(input_str)
+            converted_str = sciform_to_html(input_str)
             with self.subTest(
                 input_str=input_str,
                 expected_str=expected_str,
@@ -23,47 +23,47 @@ class TestLatexConversion(unittest.TestCase):
         for val, format_list in cases_list:
             for formatter, expected_output in format_list:
                 sciform_output = formatter(val)
-                latex_output = sciform_output.as_latex()
+                html_output = sciform_output.as_html()
                 with self.subTest(
                     val=val,
                     expected_output=expected_output,
-                    actual_output=latex_output,
+                    actual_output=html_output,
                 ):
-                    self.assertEqual(latex_output, expected_output)
+                    self.assertEqual(html_output, expected_output)
 
     def run_val_unc_formatter_conversions(self, cases_list: ValUncFormatterCases):
         for (val, unc), format_list in cases_list:
             for formatter, expected_output in format_list:
                 sciform_output = formatter(val, unc)
-                latex_output = sciform_output.as_latex()
+                html_output = sciform_output.as_html()
                 with self.subTest(
                     val=val,
                     expected_output=expected_output,
-                    actual_output=latex_output,
+                    actual_output=html_output,
                 ):
-                    self.assertEqual(latex_output, expected_output)
+                    self.assertEqual(html_output, expected_output)
 
     def test_direct_cases(self):
         cases_list = [
-            ("6.26070e-04", r"6.26070\times10^{-4}"),
+            ("6.26070e-04", r"6.26070×10<sup>-4</sup>"),
             (
                 "(0.000000(1.234560))e+02",
-                r"(0.000000(1.234560))\times10^{2}",
+                r"(0.000000(1.234560))×10<sup>2</sup>",
             ),
-            ("000_000_004_567_899.765_432_1", r"000\_000\_004\_567\_899.765\_432\_1"),
-            ("(nan)%", r"(\text{nan})\%"),
-            ("123000 ppm", r"123000\:\text{ppm}"),
-            ("0b+00", r"0\times2^{0}"),
-            ("16.18033E+03", r"16.18033\times10^{3}"),
-            ("    1.20e+01", r"\:\:\:\:1.20\times10^{1}"),
-            ("(-INF)E+00", r"(-\text{INF})\times10^{0}"),
+            ("000_000_004_567_899.765_432_1", r"000_000_004_567_899.765_432_1"),
+            ("(nan)%", r"(nan)%"),
+            ("123000 ppm", r"123000 ppm"),
+            ("0b+00", r"0×2<sup>0</sup>"),
+            ("16.18033E+03", r"16.18033×10<sup>3</sup>"),
+            ("    1.20e+01", r"    1.20×10<sup>1</sup>"),
+            ("(-INF)E+00", r"(-INF)×10<sup>0</sup>"),
             (
                 "(0.123456(789))e+03",
-                r"(0.123456(789))\times10^{3}",
[…]
