### 2024-01-24 jagerber48/sciform#114 @jagerber48
@Batalex see #134. I've come up with much cleaner way to generate latex compatible strings post-facto. I just had to hit my head against some regexes and string replacements. With this I'm tentatively deciding to drop the `latex=` option. This gives some decent simplifications throughout the formatting algorithm.

This also made it very easy to implement a `FormattedNumber` object with `_repr_latex_` and `_repr_html_` and produce output like the following image in a jupyter notebook.
![image](https://github.com/jagerber48/sciform/assets/48464524/a28ab19c-c298-448c-822e-c0d04c6e904f)

I'm somehow comfortable having the latex and HTML conversion be outside the scope of the formatter. the `Formatter` converts python number objects into unicode strings, but then the choice of display in unicode/latex/html depends on the context where the output is going to be used. It is also really nice to REMOVE the latex option rather than ADDING an html option to the `Formatter`. I'm a bit uneasy about how long the list of options for the `Formatter` already is.
