### 2024-01-24 b5a3389
add strip_env_symbs flag
