### 2024-01-24 jagerber48/sciform#114 @Batalex
Wow that's perfect!
> are you aware of other tools (other than IPython/jupyter) that would hook into _repr_latex_ or _repr_html_? If so I'd be curious to test out compatibility with those tools.

Yes, I would love to see how this fits into [quarto](https://quarto.org/)
