### 2024-01-25 330f9fb
refactor conversion code and add ASCII conversion
