+++ src/sciform/output_conversion.py
@@ -1,54 +1,81 @@
 """Convert sciform outputs into latex commands."""
+from __future__ import annotations
 
 import re
+from typing import Literal, get_args
 
-from sciform.format_utils import get_superscript_exp_str
+ascii_exp_pattern = re.compile(
+    r"^(?P<mantissa>.*)(?P<ascii_base>[eEbB])(?P<exp>[+-]\d+)$"
+)
+ascii_base_dict = {"e": 10, "E": 10, "b": 2, "B": 2}
 
-times_str = "×"
+unicode_exp_pattern = re.compile(
+    r"^(?P<mantissa>.*)×(?P<base>10|2)(?P<super_exp>[⁺⁻⁰¹²³⁴⁵⁶⁷⁸⁹])$"
+)
+superscript_translation = str.maketrans("⁺⁻⁰¹²³⁴⁵⁶⁷⁸⁹", "+-0123456789")
 
-
-def standard_exp_str_to_superscript_exp_str(match: re.Match) -> str:
-    """Convert matched ascii exp_str to unicode superscript exp_str."""
-    exp_symbol = match.group("exp_symbol")
-    symbol_to_base_dict = {"e": 10, "b": 2}
-    base = symbol_to_base_dict[exp_symbol.lower()]
-
-    exp_val_str = match.group("exp_val")
-    exp_val = int(exp_val_str)
-
-    superscript_exp_str = get_superscript_exp_str(base, exp_val)
-    return superscript_exp_str
+output_formats = Literal["latex", "html", "ascii"]
 
 
-def make_latex_superscript(match: re.Match) -> str:
-    """Convert matched superscript exp_str to latex exp_str."""
-    sup_trans = str.maketrans("⁺⁻⁰¹²³⁴⁵⁶⁷⁸⁹", "+-0123456789")
-    exp_val_non_sup = match.group(0).translate(sup_trans)
-    return rf"^{{{exp_val_non_sup}}}"
+def _make_exp_str(
+        base: int,
+        exp: int,
+        output_format: output_formats,
+        *,
+        capitalize: bool = False
+) -> str:
+    if output_format == "latex":
+        return rf"\times{base}^{{{exp}}}"
+    if output_format == "html":
+        return f"×{base}<sup>{exp}</sup>"
+    if output_format == "ascii":
+        if base == 10:
+            exp_str = f"e{exp:+03d}"
+        elif base == 2:
+            exp_str = f"b{exp:+03d}"
+        else:
+            msg = f"base must be 10 or 2, not {base}"
+            raise ValueError(msg)
+        if capitalize:
+            exp_str = exp_str.upper()
+        return exp_str
+    msg = (
+        f"output_format must be in {get_args(output_formats)}, not {output_format}"
+    )
+    raise ValueError(msg)
 
 
-def make_html_superscript(match: re.Match) -> str:
-    """Convert matched superscript exp_str to html exp_str."""
-    sup_trans = str.maketrans("⁺⁻⁰¹²³⁴⁵⁶⁷⁸⁹", "+-0123456789")
-    exp_val_non_sup = match.group(0).translate(sup_trans)
-    return rf"<sup>{exp_val_non_sup}</sup>"
+def _string_replacements(input_str: str, replacements: list[tuple[str, str]]) -> str:
+    result_str = input_str
+    for old_chars, new_chars in replacements:
+        result_str = result_str.replace(old_chars, new_chars)
+    return result_str
 
 
-def sciform_to_latex(formatted_str: str) -> str:
+def convert_sciform_format(
+        formatted_str: str,
+        output_format: output_formats,
+) -> str:
     r"""
-    Convert a sciform output string into a latex string.
+    Convert sciform output to new format for different output contexts.
 
-    conversion proceeds by
[…]
