+++ docs/source/usage.rst
@@ -216,33 +216,50 @@ If the user wishes to configure these options, but also use the
 :ref:`FSML <fsml>`, then they must do so by modifying the global default
 settings.
 
-.. _latex_conversion:
+.. _output_conversion:
 
-Latex Conversion
-================
+Output Conversion
+=================
 
-The :func:`sciform_to_latex` function can be used to convert ``sciform``
-output strings into latex commands.
+Typically the output of the :class:`Formatter` is used as a regular
+python string.
+However, the :class:`Formatter` actually returns a
+:class:`FormattedNumber` instance.
+The :class:`FormattedNumber` class subclasses ``str`` and in many cases
+is used like a normal python string.
+However, the :class:`FormattedNumber` class exposes methods to convert
+the standard string representation into LaTex, HTML, or ASCII
+representations.
+The LaTeX and HTML representations may be useful when :mod:`sciform`
+outputs are being used in contexts outside of e.g. text terminals such
+as `Matplotlib <https://matplotlib.org/>`_ plots,
+`Jupyter <https://jupyter.org/>`_ notebooks, or
+`Quarto <https://quarto.org/>`_ documents which support richer display
+functionality than unicode text.
+The ASCII representation may be useful if :mod:`sciform` outputs are
+being used in contexts in which only ASCII, and not unicode, text is
+supported or preferred.
 
->>> from sciform import sciform_to_latex
+These conversions can be accessed via the :func:`as_latex()`,
+:func:`as_html()`, :func:`as_ascii()` methods on the
+:class:`FormattedNumber` instance.
+
 >>> sform = Formatter(
 ...     exp_mode="scientific",
 ...     exp_val=-1,
 ...     upper_separator="_",
 ... )
 >>> formatted_str = sform(12345)
->>> latex_str = sciform_to_latex(formatted_str)
->>> print(f"{formatted_str} -> {latex_str}")
-123_450e-01 -> 123\_450\times10^{-1}
+>>> print(f"{formatted_str} -> {formatted_str.as_latex()}")
+123_450e-01 -> $123\_450\times10^{-1}$
 
 >>> sform = Formatter(
 ...     exp_mode="percent",
 ...     lower_separator="_",
 ... )
 >>> formatted_str = sform(0.12345678, 0.00000255)
->>> latex_str= sciform_to_latex(formatted_str)
->>> print(f"{formatted_str} -> {latex_str}")
-(12.345_678 ± 0.000_255)% -> (12.345\_678\:\pm\:0.000\_255)\%
+>>> print(f"{formatted_str} -> {formatted_str.as_latex()}")
+(12.345_678 ± 0.000_255)% -> $(12.345\_678\:\pm\:0.000\_255)\%$
 
 >>> sform = Formatter(
 ...     exp_mode="engineering",
@@ -250,9 +267,16 @@ output strings into latex commands.
 ...     ndigits=4
 ... )
 >>> formatted_str = sform(314.159e-6, 2.71828e-6)
->>> latex_str = sciform_to_latex(formatted_str)
->>> print(f"{formatted_str} -> {latex_str}")
-(314.159 ± 2.718) μ -> (314.159\:\pm\:2.718)\:\text{\textmu}
+>>> print(f"{formatted_str} -> {formatted_str.as_latex()}")
+(314.159 ± 2.718) μ -> $(314.159\:\pm\:2.718)\:\text{\textmu}$
+
+.. _latex_conversion:
+
+Latex Conversion
+================
+
+The :func:`sciform_to_latex` function can be used to convert ``sciform``
+output strings into latex commands.
 
 The latex format makes the following changes:
 
