### 2024-01-26 dd59969
fix unicode exp pattern regex
