+++ docs/source/usage.rst
@@ -224,12 +224,13 @@ Output Conversion
 Typically the output of the :class:`Formatter` is used as a regular
 python string.
 However, the :class:`Formatter` actually returns a
-:class:`FormattedNumber` instance.
-The :class:`FormattedNumber` class subclasses ``str`` and in many cases
-is used like a normal python string.
-However, the :class:`FormattedNumber` class exposes methods to convert
-the standard string representation into LaTex, HTML, or ASCII
-representations.
+:class:`FormattedNumber <formatter.FormattedNumber>` instance.
+The :class:`FormattedNumber <formatter.FormattedNumber>` class
+subclasses ``str`` and in many cases is used like a normal python
+string.
+However, the :class:`FormattedNumber <formatter.FormattedNumber>` class
+exposes methods to convert the standard string representation into
+LaTex, HTML, or ASCII representations.
 The LaTeX and HTML representations may be useful when :mod:`sciform`
 outputs are being used in contexts outside of e.g. text terminals such
 as `Matplotlib <https://matplotlib.org/>`_ plots,
@@ -240,18 +241,25 @@ The ASCII representation may be useful if :mod:`sciform` outputs are
 being used in contexts in which only ASCII, and not unicode, text is
 supported or preferred.
 
-These conversions can be accessed via the :func:`as_latex()`,
-:func:`as_html()`, :func:`as_ascii()` methods on the
-:class:`FormattedNumber` instance.
+These conversions can be accessed via the
+:meth:`as_latex() <formatter.FormattedNumber.as_latex>`,
+:meth:`as_html() <formatter.FormattedNumber.as_html>`, and
+:meth:`as_ascii() <formatter.FormattedNumber.as_ascii>` methods on the
+:class:`FormattedNumber <formatter.FormattedNumber>` class.
 
 >>> sform = Formatter(
 ...     exp_mode="scientific",
 ...     exp_val=-1,
 ...     upper_separator="_",
+...     superscript=True,
 ... )
 >>> formatted_str = sform(12345)
 >>> print(f"{formatted_str} -> {formatted_str.as_latex()}")
-123_450e-01 -> $123\_450\times10^{-1}$
+123_450×10⁻¹ -> $123\_450\times10^{-1}$
+>>> print(f"{formatted_str} -> {formatted_str.as_html()}")
+123_450×10⁻¹ -> 123_450×10<sup>-1</sup>
+>>> print(f"{formatted_str} -> {formatted_str.as_ascii()}")
+123_450×10⁻¹ -> 123_450e-01
 
 >>> sform = Formatter(
 ...     exp_mode="percent",
@@ -260,6 +268,10 @@ These conversions can be accessed via the :func:`as_latex()`,
 >>> formatted_str = sform(0.12345678, 0.00000255)
 >>> print(f"{formatted_str} -> {formatted_str.as_latex()}")
 (12.345_678 ± 0.000_255)% -> $(12.345\_678\:\pm\:0.000\_255)\%$
+>>> print(f"{formatted_str} -> {formatted_str.as_html()}")
+(12.345_678 ± 0.000_255)% -> (12.345_678 ± 0.000_255)%
+>>> print(f"{formatted_str} -> {formatted_str.as_ascii()}")
+(12.345_678 ± 0.000_255)% -> (12.345_678 +/- 0.000_255)%
 
 >>> sform = Formatter(
 ...     exp_mode="engineering",
@@ -269,29 +281,10 @@ These conversions can be accessed via the :func:`as_latex()`,
[…]
