### 2024-01-26 66298dd
_repr_html_ and _repr_latex_ to use as_html and as_latex (instead of the other way around)
