### 2024-01-26 a9e081c
consolidate output conversion tests
