+++ tests/test_latex_conversion.py
@@ -5,66 +5,248 @@ import unittest
 from sciform import Formatter
 from sciform.output_conversion import convert_sciform_format
 
-ValFormatterCases = list[tuple[float, list[tuple[Formatter, str]]]]
-ValUncFormatterCases = list[tuple[tuple[float, float], list[tuple[Formatter, str]]]]
+ValFormatterCases = list[
+    tuple[float, list[tuple[Formatter, tuple[str, str, str, str]]]]
+]
+ValUncFormatterCases = list[
+    tuple[tuple[float, float], list[tuple[Formatter, tuple[str, str, str, str]]]]
+]
 
 
 class TestLatexConversion(unittest.TestCase):
-    def run_direct_conversions(self, cases_list: list[tuple[str, str]]):
-        for input_str, expected_str in cases_list:
-            converted_str = convert_sciform_format(input_str, "latex")
+    def run_direct_conversions(
+        self,
+        cases_list: list[tuple[str, tuple[str, str, str]]],
+    ):
+        for input_str, (expected_ascii, expected_html, expected_latex) in cases_list:
+            ascii_str = convert_sciform_format(input_str, "ascii")
             with self.subTest(
                 input_str=input_str,
-                expected_str=expected_str,
-                actual_str=converted_str,
+                expected_str=expected_ascii,
+                actual_str=ascii_str,
             ):
-                self.assertEqual(converted_str, expected_str)
+                self.assertEqual(ascii_str, expected_ascii)
+
+            html_str = convert_sciform_format(input_str, "html")
+            with self.subTest(
+                input_str=input_str,
+                expected_str=expected_html,
+                actual_str=html_str,
+            ):
+                self.assertEqual(html_str, expected_html)
+
+            latex_str = convert_sciform_format(input_str, "latex")
+            with self.subTest(
+                input_str=input_str,
+                expected_str=expected_latex,
+                actual_str=latex_str,
+            ):
+                self.assertEqual(latex_str, expected_latex)
 
     def run_val_formatter_conversions(self, cases_list: ValFormatterCases):
         for val, format_list in cases_list:
-            for formatter, expected_output in format_list:
+            for formatter, expected_outputs in format_list:
+                (
+                    expected_str,
+                    expected_ascii,
+                    expected_html,
+                    expected_latex,
+                ) = expected_outputs
                 sciform_output = formatter(val)
+
+                str_output = sciform_output.as_str()
+                with self.subTest(
+                    val=val,
+                    expected_output=expected_str,
+                    actual_output=str_output,
+                ):
+                    self.assertEqual(str_output, expected_str)
+
+                ascii_output = sciform_output.as_ascii()
+                with self.subTest(
+                    val=val,
+                    expected_output=expected_ascii,
+                    actual_output=ascii_output,
+                ):
+                    self.assertEqual(ascii_output, expected_ascii)
+
+                html_output = sciform_output.as_html()
+                with self.subTest(
+                    val=val,
+                    expected_output=expected_html,
+                    actual_output=html_output,
+                ):
+                    self.assertEqual(html_output, expected_html)
+
                 latex_output = sciform_output.as_latex(strip_env_symbs=True)
                 with self.subTest(
                     val=val,
-                    expected_output=expected_output,
+                    expected_output=expected_latex,
                     actual_output=latex_output,
                 ):
-                    self.assertEqual(latex_output, expected_output)
+                    self.assertEqual(latex_output, expected_latex)
 
     def run_val_unc_formatter_conversions(self, cases_list: ValUncFormatterCases):
         for (val, unc), format_list in cases_list:
-            for formatter, expected_output in format_list:
[…]
