### 2024-01-26 bb8cca4
test latex/html repr
