### 2024-01-26 695e08b
test invalid cases
