### 2024-01-26 5671994
strip_math_mode rename and documentation
