### 2024-01-26 1e448b3
_repr_html_ and _repr_latex_ docs
