### 2024-01-26 770dc08
jupyter/IPython display example
