### 2024-01-26 b34bdbc
break out "examples" dependencies
