### 2024-01-27 jagerber48/sciform#114 @jagerber48
> Yes, I would love to see how this fits into [quarto](https://quarto.org/)

Quarto works somewhat as expected. I have this jupyter notebook cell (the same as appears in [the latest docs](https://sciform.readthedocs.io/en/latest/usage.html#output-conversion) that looks like
![image](https://github.com/jagerber48/sciform/assets/48464524/d04ace54-5827-4d52-84c6-50c91205e9bf)
When I render this into a quarto html doc I get
![image](https://github.com/jagerber48/sciform/assets/48464524/4974c7c7-b01c-49d9-9b2a-af0374bc8fb4)
So the `display` functions aren't interleaved as I would naively expect with the `print` statements. But I think that's a bit of a standard `IPython` issue that probably has workaround. So not a `sciform` issue, but rather an `IPython` thing to figure out.

I'm curious how people want to use the `LaTeX` features. The `display` function is one of the most integrated ways to show LaTeX output with python. maptlotlib plots can also show LaTeX. But other than these two cases, I can't imagine how people would be getting LaTeX out of python and into a LaTeX enabled document without a copy and paste step.

Should probably open another issue about this.
