+++ .github/workflows/python-package.yml
@@ -27,7 +27,6 @@ jobs:
     - name: Install dependencies
       run: |
         python -m pip install --upgrade pip
-        python -m pip install ruff blackdoc pytest pytest-cov
         python -m pip install -e .[dev]
     - name: Lint/format with ruff
       run: |
@@ -38,7 +37,9 @@ jobs:
         blackdoc .
     - name: Test with pytest
       run: |
-        pytest --cov --doctest-glob="*.rst"
+        coverage run -m unittest
+        coverage report
+        python -m sphinx -b doctest docs/source/ docs/build/doctest/
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v3
       with:
