+++ pyproject.toml
@@ -31,6 +31,8 @@ documentation = "https://sciform.readthedocs.io"
 [project.optional-dependencies]
 dev = [
     "ruff",
+    "blackdoc",
+    "coverage",
     "sphinx",
     "sphinx-rtd-theme",
     "sphinx-toolbox",
