### 2024-01-27 b200882
change CI order and specify coverage file
