+++ .github/workflows/python-package.yml
@@ -29,22 +29,25 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install -e .[dev]
-    - name: Lint/format with ruff
+    - name: Lint/format source code
       run: |
         ruff check .
         ruff format . --check
-    - name: Lint/format docs
-      run: |
-        blackdoc .
-    - name: Test with pytest
+    - name: Test source code and measure coverage
       run: |
         coverage run -m unittest
         coverage report
+    - name: Lint/format docs
+      run: |
+        blackdoc .
+    - name: Test docs
+      run: |
         python -m sphinx -b doctest docs/source/ docs/build/doctest/
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v3
       with:
         token: ${{ secrets.CODECOV_TOKEN }}
         flags: ${{ matrix.python-version }}
+        files: .coverage
         fail_ci_if_error: true
         verbose: true
