+++ .github/workflows/python-package.yml
@@ -29,19 +29,15 @@ jobs:
       run: |
         python -m pip install --upgrade pip
         python -m pip install -e .[dev]
-    - name: Lint/format source code
+    - name: Lint/format source code and docs
       run: |
         ruff check .
         ruff format . --check
-    - name: Test source code and measure coverage
+        blackdoc . --check
+    - name: Test source code and docs
       run: |
         coverage run -m unittest
         coverage report
-    - name: Lint/format docs
-      run: |
-        blackdoc .
-    - name: Test docs
-      run: |
         python -m sphinx -b doctest docs/source/ docs/build/doctest/
     - name: Upload coverage reports to Codecov
       uses: codecov/codecov-action@v3
