### 2024-01-28 04bbdca
remove blackdoc requirement (it has some weird behaviors)
