### 2024-01-28 #121 @jagerber48
@Batalex has let me know the review will be wrapping up and asked if there's anything else I'd like to add to the codebase first. A this point there's a continuous stream of small to medium sized improvements I'm making, but I don't think any of them need to tie up or be tied up in the review.

There is one thing I'd mention though. @Batalex suggested a while ago that `sciform` return a class which defines `_repr_latex_` and `_repr_html_` functions so that tools like `IPython` or maybe quarto could hook into these functions to allow richer display than pure unicode in contexts where that is possible. Essentially a beefed up version of the old `latex` option. After some thought I was able to get this implemented. This involved removing the `latex` option and introducing the `FormattedNumber` class which is returned by the `Formatter`.

See

* https://github.com/jagerber48/sciform/pull/134
* https://github.com/jagerber48/sciform/issues/114
* https://github.com/jagerber48/sciform/discussions/83
* https://sciform.readthedocs.io/en/latest/usage.html#output-conversion (Latest docs, not stable)

The feature is merged into `main` now but has yet to be released. I'll probably release it this week into version `0.33.0`.
