### 2024-01-29 2420e9f
add format_to_target_length and tests
