### 2024-01-29 6f25a71
test_cases to look more like gformat tests
