+++ tests/test_target_length.py
@@ -33,63 +33,63 @@ class TestFormatToTargetLength(unittest.TestCase):
         cases = [
             (
                 (11301.3646429824, None),
-                "11301.3646",
+                " 11301.3646",
             ),
             (
                 (7.55438813238573, None),
-                "7.55438813",
+                " 7.55438813",
             ),
             (
                 (3037.18756328971, None),
-                "3037.18756",
+                " 3037.18756",
             ),
             (
                 (3058.4404411981735, None),
-                "3058.44044",
+                " 3058.44044",
             ),
             (
                 (13.8904759429872987, None),
-                "13.8904759",
+                " 13.8904759",
             ),
             (
                 (5.440263872938719, None),
-                "5.44026387",
+                " 5.44026387",
             ),
             (
                 (0.124643890130481, None),
-                "0.12464389",
+                " 0.12464389",
             ),
             (
                 (0.009963634432798437298, None),
-                "0.00996363",
+                " 0.00996363",
             ),
             (
                 (0.2441075332894732, None),
-                "0.24410753",
+                " 0.24410753",
             ),
             (
                 (0.0141610629879, None),
-                "0.01416106",
+                " 0.01416106",
             ),
             (
                 (0.0241421029837978, None),
-                "0.02414210",
+                " 0.02414210",
             ),
             (
                 (0.0002027524899875, None),
-                "0.00020275",
+                " 0.00020275",
             ),
             (
                 (0.00002027524899875, None),
-                "2.0275e-05",
+                " 2.0275e-05",
             ),
         ]
 
         self.run_val_cases(
             cases,
-            target_length=10,
+            target_length=11,
             allowed_exp_modes=["fixed_point", "scientific"],
-            base_formatter=None,
+            base_formatter=Formatter(sign_mode=" "),
         )
 
     def test_lmfit_gformat_data(self):
