### 2024-01-31 #121 @jagerber48
The previously mentioned `FormattedNumber` updates and others were just released in `0.33.0`. https://github.com/jagerber48/sciform/releases/tag/0.33.0
