### 2024-02-01 7e65564
allow integer 0 for left_pad_char
