### 2024-02-01 39a0e23
add components for dict display
OptionsDict for pretty printing dicts
rendered options and user options can return dict versions of themselves.
Formatter uses the user and rendered options getters.
tests
