### 2024-02-02 352c751
refactor render options
