+++ src/sciform/user_options.py
@@ -3,8 +3,9 @@
 
 from __future__ import annotations
 
-from dataclasses import InitVar, asdict, dataclass
-from typing import Literal, get_args
+from dataclasses import asdict, dataclass
+from pprint import pformat
+from typing import Any, Literal, get_args
 
 from sciform import global_options, modes
 from sciform.rendered_options import RenderedOptions
@@ -37,154 +38,154 @@ class UserOptions:
     paren_uncertainty_separators: bool | None = None
     pm_whitespace: bool | None = None
 
-    add_c_prefix: InitVar[bool] = False
-    add_small_si_prefixes: InitVar[bool] = False
-    add_ppth_form: InitVar[bool] = False
+    add_c_prefix: bool = None
+    add_small_si_prefixes: bool = None
+    add_ppth_form: bool = None
 
-    def __post_init__(
-        self: UserOptions,
-        add_c_prefix: bool,  # noqa: FBT001
-        add_small_si_prefixes: bool,  # noqa: FBT001
-        add_ppth_form: bool,  # noqa: FBT001
-    ) -> None:
-        if self.left_pad_char == 0:
-            super().__setattr__("left_pad_char", "0")
-        self.populate_dicts(
-            add_c_prefix=add_c_prefix,
-            add_small_si_prefixes=add_small_si_prefixes,
-            add_ppth_form=add_ppth_form,
-        )
-        self.validate(self)
+    def __post_init__(self: UserOptions) -> None:
+        validate_options(self)
 
-    def populate_dicts(  # noqa: C901
-        self: UserOptions,
-        *,
-        add_c_prefix: bool,
-        add_small_si_prefixes: bool,
-        add_ppth_form: bool,
-    ) -> None:
-        """Populate extra prefix translations from user input flags."""
-        if add_c_prefix:
-            if self.extra_si_prefixes is None:
-                super().__setattr__("extra_si_prefixes", {})
-            if -2 not in self.extra_si_prefixes:
-                self.extra_si_prefixes[-2] = "c"
+    def as_dict(self: UserOptions) -> dict[str, Any]:
+        """Return a dict representation of the inputs used to construct UserOptions."""
+        options_dict = asdict(self)
+        for key in list(options_dict.keys()):
+            if options_dict[key] is None:
+                del options_dict[key]
+        return options_dict
 
-        if add_small_si_prefixes:
-            if self.extra_si_prefixes is None:
-                super().__setattr__("extra_si_prefixes", {})
-            if -2 not in self.extra_si_prefixes:
-                self.extra_si_prefixes[-2] = "c"
-            if -1 not in self.extra_si_prefixes:
-                self.extra_si_prefixes[-1] = "d"
-            if +1 not in self.extra_si_prefixes:
-                self.extra_si_prefixes[+1] = "da"
-            if +2 not in self.extra_si_prefixes:
-                self.extra_si_prefixes[+2] = "h"
+    def __str__(self: UserOptions) -> str:
+        return pformat(self.as_dict(), sort_dicts=False)
 
-        if add_ppth_form:
-            if self.extra_parts_per_forms is None:
-                super().__setattr__("extra_parts_per_forms", {})
-            if -3 not in self.extra_parts_per_forms:
-                self.extra_parts_per_forms[-3] = "ppth"
 
-    @staticmethod
-    def validate(options: UserOptions | RenderedOptions) -> None:
-        """Validate user inputs."""
+def validate_options(options: UserOptions | RenderedOptions) -> None:
+    """Validate user inputs."""
+    if (
+        options.round_mode == "sig_fig"
+        and isinstance(options.ndigits, int)
+        and options.ndigits < 1
+    ):
[…]
