+++ src/sciform/options/__init__.py
@@ -1,0 +1,2 @@
+from sciform.options.input_options import InputOptions
+from sciform.options.finalized_options import FinalizedOptions
