+++ src/sciform/options/__init__.py
@@ -1,2 +1,0 @@
-from sciform.options.input_options import InputOptions
-from sciform.options.finalized_options import FinalizedOptions
