+++ src/sciform/formatter.py
@@ -231,49 +231,6 @@ class Formatter:
         :param uncertainty: Optional uncertainty to be formatted.
         :type uncertainty: ``Decimal | float | int | str | None``
         """
-        return FormattedNumber(
-            format_from_options(value, uncertainty, input_options=self._user_options)
+        return format_from_options(
+            value, uncertainty, input_options=self._user_options
         )
-
-
-class FormattedNumber(str):
-    """
-    Representation (typically string) of a formatted number.
-
-    The ``FormattedNumber`` class is returned by ``sciform`` formatting
-    methods. In most cases it behaves like a regular python string, but
-    it provides the possibility for post-converting the string to
-    various other formats such as latex or html. This allows the
-    formatted number to be displayed in a range of contexts other than
-    e.g. text terminals.
-
-    """
-
-    __slots__ = ()
-
-    def as_str(self: FormattedNumber) -> str:
-        """Return the string representation of the formatted number."""
-        return self.__str__()
-
-    def as_ascii(self: FormattedNumber) -> str:
-        """Return the ascii representation of the formatted number."""
-        return convert_sciform_format(self, "ascii")
-
-    def as_html(self: FormattedNumber) -> str:
-        """Return the html representation of the formatted number."""
-        return convert_sciform_format(self, "html")
-
-    def as_latex(self: FormattedNumber, *, strip_math_mode: bool = False) -> str:
-        """Return the latex representation of the formatted number."""
-        latex_repr = convert_sciform_format(self, "latex")
-        if strip_math_mode:
-            latex_repr = latex_repr.strip("$")
-        return latex_repr
-
-    def _repr_html_(self: FormattedNumber) -> str:
-        """Hook for HTML display."""  # noqa: D401
-        return self.as_html()
-
-    def _repr_latex_(self: FormattedNumber) -> str:
-        """Hook for LaTeX display."""  # noqa: D401
-        return self.as_latex()
