+++ src/sciform/options/validation.py
@@ -13,7 +13,7 @@ if TYPE_CHECKING:
 
 
 def validate_options(
-        options: InputOptions | PopulatedOptions | FinalizedOptions,
+    options: InputOptions | PopulatedOptions | FinalizedOptions,
 ) -> None:
     """Validate user inputs."""
     validate_sig_fig_round_mode(options)
@@ -22,7 +22,7 @@ def validate_options(
 
 
 def validate_sig_fig_round_mode(
-        options: InputOptions | PopulatedOptions | FinalizedOptions,
+    options: InputOptions | PopulatedOptions | FinalizedOptions,
 ) -> None:
     r"""Validate ndigits if round_mode == "sig_fig"."""
     if (
@@ -35,7 +35,7 @@ def validate_sig_fig_round_mode(
 
 
 def validate_exp_val(
-        options: InputOptions | PopulatedOptions | FinalizedOptions,
+    options: InputOptions | PopulatedOptions | FinalizedOptions,
 ) -> None:
     """Validate exp_val."""
     if options.exp_val is not modes.AutoExpVal and options.exp_val is not None:
@@ -63,7 +63,7 @@ def validate_exp_val(
 
 
 def validate_separator_options(
-        options: InputOptions | PopulatedOptions | FinalizedOptions,
+    options: InputOptions | PopulatedOptions | FinalizedOptions,
 ) -> None:
     """Validate separator user input."""
     if options.upper_separator is not None:
