+++ src/sciform/options/input_options.py
@@ -9,7 +9,7 @@ from typing import TYPE_CHECKING, Any, Literal
 
 from sciform.options.validation import validate_options
 
-if TYPE_CHECKING:
+if TYPE_CHECKING:  # pragma: no cover
     from sciform import modes
 
 
