+++ src/sciform/formatting.py
@@ -30,7 +30,7 @@ from sciform.modes import (
 from sciform.options.conversion import finalize_input_options
 from sciform.output_conversion import convert_sciform_format
 
-if TYPE_CHECKING:
+if TYPE_CHECKING:  # pragma: no cover
     from sciform.options.finalized_options import FinalizedOptions
     from sciform.options.input_options import InputOptions
 
