### 2024-02-03 a9717b5
update options handling in Formatter
