### 2024-02-03 f133699
Rename global defaults -> global options. Drop printing in package
