### 2024-02-03 eaa9bc3
string comparison tests instead of prints
