### 2024-02-03 12e816c
add test for input_options string repr
