### 2024-02-03 e19fe3a
formatter populated options test
