### 2024-02-03 d921235
exposed FormattedNumber, InputOptions, PopulatedOptions for documentation
