+++ src/sciform/__init__.py
@@ -1,6 +1,7 @@
 """``sciform`` is used to convert python numbers into scientific formatted strings."""
 
 from sciform.formatter import Formatter
+from sciform.formatting import FormattedNumber
 from sciform.global_configuration import (
     GlobalOptionsContext,
     get_global_options,
@@ -8,10 +9,13 @@ from sciform.global_configuration import (
     set_global_options,
 )
 from sciform.modes import AutoDigits, AutoExpVal
+from sciform.options.input_options import InputOptions
+from sciform.options.populated_options import PopulatedOptions
 from sciform.scinum import SciNum
 
 __all__ = [
     "Formatter",
+    "FormattedNumber",
     "GlobalOptionsContext",
     "get_global_options",
     "reset_global_options",
@@ -19,4 +23,6 @@ __all__ = [
     "AutoDigits",
     "AutoExpVal",
     "SciNum",
+    "InputOptions",
+    "PopulatedOptions",
 ]
