### 2024-02-03 ed68492
changed InputOptions and PopulatedOptions __str__ so they don't look like dicts. Update Formatter docstring.
