+++ src/sciform/options/input_options.py
@@ -56,4 +56,7 @@ class InputOptions:
         return options_dict
 
     def __str__(self: InputOptions) -> str:
-        return pformat(self.as_dict(), sort_dicts=False)
+        options_str = pformat(self.as_dict(), width=-1, sort_dicts=False)
+        options_str = options_str.lstrip('{').rstrip('}')
+        options_str = f"InputOptions(\n {options_str},\n)"
+        return options_str
