### 2024-02-03 f5d3b63
add get_default_global_options
