### 2024-02-04 023c894
shorten Formatter.populated_options docstring
