### 2024-02-04 a05a4f0
get_default_global_options to api.rst
