+++ docs/source/usage.rst
@@ -23,29 +23,30 @@ The :class:`Formatter` object is then called with a number and returns
 a corresponding formatted string.
 
 >>> from sciform import Formatter
->>> sform = Formatter(
+>>> formatter = Formatter(
 ...     round_mode="dec_place", ndigits=6, upper_separator=" ", lower_separator=" "
 ... )
->>> print(sform(51413.14159265359))
+>>> print(formatter(51413.14159265359))
 51 413.141 593
->>> sform = Formatter(round_mode="sig_fig", ndigits=4, exp_mode="engineering")
->>> print(sform(123456.78))
+>>> formatter = Formatter(round_mode="sig_fig", ndigits=4, exp_mode="engineering")
+>>> print(formatter(123456.78))
 123.5e+03
 
 It is not necessary to provide input for all options.
 At format time, any un-populated options will be populated with the
-corresponding options from the global default options.
+corresponding options from the global options.
 See :ref:`global_config` for details about how to view and modify the
-global default options.
+global options.
 
 SciNum
 ------
 
 The :mod:`sciform` :ref:`FSML <fsml>` can be accessed via the
 :class:`SciNum` object.
-Python numbers specified as :class:`string`, :class:`float`, or
-:class:`Decimal` objects are cast to :class:`SciNum` objects which can
-be formatted using the :mod:`sciform` :ref:`FSML <fsml>`.
+Python numbers specified as :class:`string`, :class:`int`,
+:class:`float`, or :class:`Decimal` objects are cast to :class:`SciNum`
+objects which can be formatted using the :mod:`sciform`
+:ref:`FSML <fsml>`.
 
 >>> from sciform import SciNum
 >>> num = SciNum(123456)
@@ -72,19 +73,21 @@ using the :class:`SciNum` object.
 
 >>> val = 84.3
 >>> unc = 0.2
->>> sform = Formatter(ndigits=2)
->>> print(sform(val, unc))
+>>> formatter = Formatter(ndigits=2)
+>>> print(formatter(val, unc))
 84.30 ± 0.20
 >>> from sciform import SciNum
->>> val_unc = SciNum(val, unc)
->>> print(f"{val_unc:!2}")
+>>> num = SciNum(val, unc)
+>>> print(f"{num:!2}")
 84.30 ± 0.20
 
 Value/uncertainty pairs can also be formatted using a parentheses
 notation in which the uncertainty is displayed in parentheses following
 the value.
+See
+`BIPM Guide Section 7.2.2 <https://www.bipm.org/documents/20126/2071204/JCGM_100_2008_E.pdf/cb0ef43f-baa5-11cf-3f85-4dcd86f77bd6#page=37>`_.
 
->>> print(f"{val_unc:!2()}")
+>>> print(f"{num:!2()}")
 84.30(20)
 
 Value/uncertainty pairs are formatted according to the following
@@ -107,21 +110,20 @@ algorithm:
 Global Configuration
 ====================
 
-It is possible to modify the global default configuration for
-:mod:`sciform` to avoid repetition of verbose configuration options or
-format specification strings.
+It is possible to modify the global options for :mod:`sciform` to avoid
+repetition of verbose configuration options or format specification
+strings.
 When the user creates a :class:`Formatter` object or formats a string
 using the :ref:`FSML <fsml>`, they typically do not specify settings for
 all available options.
-In these cases, the unspecified options resolve their values from the
-global default settings at format time.
+In these cases, the unpopulated options resolve their values from the
+global options at format time.
 
-The global default settings can be viewed using
[…]
