### 2024-02-04 163c3fa
sform -> formatter, scinum -> num
