+++ docs/source/options.rst
@@ -36,10 +36,10 @@ Fixed point notation is standard notation in which a number is displayed
 directly with no extra exponent.
 
 >>> from sciform import Formatter
->>> sform = Formatter(exp_mode="fixed_point")
->>> print(sform(123.456))
+>>> formatter = Formatter(exp_mode="fixed_point")
+>>> print(formatter(123.456))
 123.456
->>> print(sform(123.456, 0.001))
+>>> print(formatter(123.456, 0.001))
 123.456 ± 0.001
 
 .. _percent_mode:
@@ -51,10 +51,10 @@ Percent mode is similar to fixed point mode.
 For percent mode, the number is multiplied by 100 and a ``%`` symbol is
 appended to the end of the formatted string.
 
->>> sform = Formatter(exp_mode="percent")
->>> print(sform(0.12345))
+>>> formatter = Formatter(exp_mode="percent")
+>>> print(formatter(0.12345))
 12.345%
->>> print(sform(0.12345, 0.001))
+>>> print(formatter(0.12345, 0.001))
 (12.3 ± 0.1)%
 
 .. _scientific:
@@ -66,10 +66,10 @@ Scientific notation is used to display base-10 decimal numbers.
 In scientific notation, the exponent is uniquely chosen so that the
 mantissa ``m`` satisfies ``1 <= m < 10``.
 
->>> sform = Formatter(exp_mode="scientific")
->>> print(sform(123.456))
+>>> formatter = Formatter(exp_mode="scientific")
+>>> print(formatter(123.456))
 1.23456e+02
->>> print(sform(123.456, 0.001))
+>>> print(formatter(123.456, 0.001))
 (1.23456 ± 0.00001)e+02
 
 By default the exponent is expressed using ASCII characters, e.g.
@@ -99,10 +99,10 @@ Here it may be more convenient to display the number with an exponent of
 to an exponent of 8 which would be chosen for standard scientific
 notation.
 
->>> sform = Formatter(exp_mode="engineering")
->>> print(sform(123.456))
+>>> formatter = Formatter(exp_mode="engineering")
+>>> print(formatter(123.456))
 123.456e+00
->>> print(sform(123.456, 0.001))
+>>> print(formatter(123.456, 0.001))
 (123.456 ± 0.001)e+00
 
 .. _engineering_shifted:
@@ -114,10 +114,10 @@ Shifted engineering notation is the same as engineering notation except
 the exponent is chosen so that the mantissa ``m`` satisfies
 ``0.1 <= m < 100``.
 
->>> sform = Formatter(exp_mode="engineering_shifted")
->>> print(sform(123.456))
+>>> formatter = Formatter(exp_mode="engineering_shifted")
+>>> print(formatter(123.456))
 0.123456e+03
->>> print(sform(123.456, 0.001))
+>>> print(formatter(123.456, 0.001))
 (0.123456 ± 0.000001)e+03
 
 .. _binary:
@@ -128,8 +128,8 @@ Binary
 Binary formatting can be chosen to display a number in scientific
 notation in base-2.
 
->>> sform = Formatter(exp_mode="binary")
->>> print(sform(256))
+>>> formatter = Formatter(exp_mode="binary")
+>>> print(formatter(256))
 1b+08
 
 Here ``b`` exponent symbol indicates base-2 instead of base-10.
@@ -146,8 +146,8 @@ exponent is constrained to be a multiple of 10, consistent with the
 IEC binary prefixes.
 The mantissa ``m`` satisfies ``1 <= m < 1024``.
 
->>> sform = Formatter(exp_mode="binary_iec")
->>> print(sform(2048))
+>>> formatter = Formatter(exp_mode="binary_iec")
[…]

[…]
