### 2024-02-04 1e0427e
add populated_options record to FormattedNumber
