+++ CHANGELOG.rst
@@ -8,6 +8,13 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 Unreleased
 ----------
 
+* There are no unreleased changes.
+
+----
+
+0.34.0 (2024-02-04)
+-------------------
+
 Added
 ^^^^^
 
