### 2024-02-05 f15844d
add paren_uncertainty_trim_digits and rename paren_uncertainty_trim_separators and reverse logic (again)
