+++ src/sciform/format_utils.py
@@ -439,7 +439,8 @@ def construct_val_unc_str(  # noqa: PLR0913
     *,
     paren_uncertainty: bool,
     pm_whitespace: bool,
-    paren_uncertainty_separators: bool,
+    paren_uncertainty_trim_digits: bool,
+    paren_uncertainty_trim_separators: bool,
 ) -> str:
     """Construct the value/uncertainty part of the formatted string."""
     if not paren_uncertainty:
@@ -449,31 +450,32 @@ def construct_val_unc_str(  # noqa: PLR0913
         val_unc_str = f"{val_mantissa_str}{pm_symb}{unc_mantissa_str}"
     else:
         if unc_mantissa.is_finite():
-            if val_mantissa.is_finite() and 0 < unc_mantissa < abs(val_mantissa):
-                """
-                Don't left strip the unc_mantissa_str if val_mantissa is non-finite.
-                Don't left strip the unc_mantissa_str if unc_mantissa == 0 (because then
-                the empty string would remain).
-                Don't left strip the unc_mantissa_str if unc_mantissa >= val_mantissa
-                """
-                unc_mantissa_str = unc_mantissa_str.lstrip("0.,_ ")
-            if not paren_uncertainty_separators:
-                for separator in SeparatorEnum:
-                    if separator == decimal_separator:
-                        if val_mantissa.is_finite():
-                            if unc_mantissa >= abs(val_mantissa):
+            if paren_uncertainty_trim_digits:
+                if val_mantissa.is_finite() and 0 < unc_mantissa < abs(val_mantissa):
+                    """
+                    Don't left strip the unc_mantissa_str if val_mantissa is non-finite.
+                    Don't left strip the unc_mantissa_str if unc_mantissa == 0 (because then
+                    the empty string would remain).
+                    Don't left strip the unc_mantissa_str if unc_mantissa >= val_mantissa
+                    """
+                    unc_mantissa_str = unc_mantissa_str.lstrip("0.,_ ")
+                if paren_uncertainty_trim_separators:
+                    for separator in SeparatorEnum:
+                        if separator == decimal_separator:
+                            if val_mantissa.is_finite():
+                                if unc_mantissa >= abs(val_mantissa):
+                                    """
+                                    Don't remove the decimal separator if the 
+                                    uncertainty is larger than the value.
+                                    """
+                                    continue
+                            else:
                                 """
-                                Don't remove the decimal separator if the uncertainty is
-                                larger than the value.
+                                Don't remove the decimal separator if the value is
+                                non-finite.
                                 """
                                 continue
-                        else:
-                            """
-                            Don't remove the decimal separator if the value is
-                            non-finite.
-                            """
-                            continue
-                    unc_mantissa_str = unc_mantissa_str.replace(separator, "")
+                        unc_mantissa_str = unc_mantissa_str.replace(separator, "")
         val_unc_str = f"{val_mantissa_str}({unc_mantissa_str})"
     return val_unc_str
 
