+++ src/sciform/options/global_options.py
@@ -24,7 +24,8 @@ PKG_DEFAULT_OPTIONS = PopulatedOptions(
     paren_uncertainty=False,
     pdg_sig_figs=False,
     left_pad_matching=False,
-    paren_uncertainty_separators=True,
+    paren_uncertainty_trim_digits=True,
+    paren_uncertainty_trim_separators=False,
     pm_whitespace=True,
 )
 
