+++ tests/test_options_printout.py
@@ -57,7 +57,8 @@ class TestOptionsPrintOut(unittest.TestCase):
             " 'paren_uncertainty': False,\n"
             " 'pdg_sig_figs': False,\n"
             " 'left_pad_matching': False,\n"
-            " 'paren_uncertainty_separators': True,\n"
+            " 'paren_uncertainty_trim_digits': True,\n"
+            " 'paren_uncertainty_trim_separators': False,\n"
             " 'pm_whitespace': True,\n"
             ")"
         )
@@ -87,7 +88,8 @@ class TestOptionsPrintOut(unittest.TestCase):
             " 'paren_uncertainty': False,\n"
             " 'pdg_sig_figs': False,\n"
             " 'left_pad_matching': False,\n"
-            " 'paren_uncertainty_separators': True,\n"
+            " 'paren_uncertainty_trim_digits': True,\n"
+            " 'paren_uncertainty_trim_separators': False,\n"
             " 'pm_whitespace': True,\n"
             ")"
         )
@@ -118,7 +120,8 @@ class TestOptionsPrintOut(unittest.TestCase):
             " 'paren_uncertainty': False,\n"
             " 'pdg_sig_figs': False,\n"
             " 'left_pad_matching': False,\n"
-            " 'paren_uncertainty_separators': True,\n"
+            " 'paren_uncertainty_trim_digits': True,\n"
+            " 'paren_uncertainty_trim_separators': False,\n"
             " 'pm_whitespace': True,\n"
             ")"
         )
