+++ src/sciform/options/finalized_options.py
@@ -40,7 +40,8 @@ class FinalizedOptions:
     paren_uncertainty: bool
     pdg_sig_figs: bool
     left_pad_matching: bool
-    paren_uncertainty_separators: bool
+    paren_uncertainty_trim_digits: bool
+    paren_uncertainty_trim_separators: bool
     pm_whitespace: bool
 
     def __post_init__(self: FinalizedOptions) -> None:

[…]
