### 2024-02-05 21b3ea5
Remove paren_uncertainty separator control entirely
