### 2024-02-05 8bd80ac
trim_digits to trim. Change behavior so that trim removes digits to the left of the most significant uncertainty digit but also removes all non-decimal separators. tests
