+++ tests/test_val_unc_formatter.py
@@ -394,6 +394,141 @@ class TestFormatting(unittest.TestCase):
         self.assertEqual(result, expected_result)
 
     def test_paren_uncertainties_trim_digits(self):
+        cases_list = [
+            (
+                (123, 0.03),
+                [
+                    (
+                        Formatter(
+                            paren_uncertainty=True,
+                            paren_uncertainty_trim=True,
+                        ),
+                        "123.00(3)",
+                    ),
+                    (
+                        Formatter(
+                            paren_uncertainty=True,
+                            paren_uncertainty_trim=False,
+                        ),
+                        "123.00(0.03)",
+                    ),
+                ],
+            ),
+            (
+                (123.456789, 0.0012),
+                [
+                    (
+                        Formatter(
+                            paren_uncertainty=True,
+                            paren_uncertainty_trim=False,
+                            lower_separator="_",
+                        ),
+                        "123.456_8(0.001_2)",
+                    ),
+                    (
+                        # Intermediate separators are stripped.
+                        Formatter(
+                            paren_uncertainty=True,
+                            paren_uncertainty_trim=True,
+                            lower_separator="_",
+                        ),
+                        "123.456_8(12)",
+                    ),
+                ],
+            ),
+            (
+                (123456, 1234),
+                [
+                    (
+                        Formatter(
+                            paren_uncertainty=True,
+                            paren_uncertainty_trim=False,
+                            upper_separator=" ",
+                        ),
+                        "123 456(1 234)",
+                    ),
+                    (
+                        # Intermediate separators are stripped.
+                        Formatter(
+                            paren_uncertainty=True,
+                            paren_uncertainty_trim=True,
+                            upper_separator=" ",
+                        ),
+                        "123 456(1234)",
+                    ),
+                ],
+            ),
+            (
+                (123456, 1234),
+                [
+                    (
+                        Formatter(
+                            paren_uncertainty=True,
+                            paren_uncertainty_trim=False,
+                            upper_separator=",",
+                        ),
+                        "123,456(1,234)",
+                    ),
+                    (
+                        # Intermediate separators are stripped.
+                        Formatter(
+                            paren_uncertainty=True,
+                            paren_uncertainty_trim=True,
+                            upper_separator=",",
+                        ),
+                        "123,456(1234)",
+                    ),
+                ],
+            ),
+            (
+                (0.0012, 123.456789),
+                [
+                    (
+                        Formatter(
+                            paren_uncertainty=True,
+                            paren_uncertainty_trim=False,
+                            lower_separator="_",
+                        ),
+                        "0.001_200(123.456_789)",
+                    ),
+                    (
+                        # No trimming for uncertainty > value
+                        Formatter(
+                            paren_uncertainty=True,
+                            paren_uncertainty_trim=True,
+                            lower_separator="_",
+                        ),
+                        "0.001_200(123.456_789)",
+                    ),
+                ],
+            ),
+            (
+                (0.0012, 0.0012),
+                [
+                    (
+                        Formatter(
+                            paren_uncertainty=True,
+                            paren_uncertainty_trim=False,
+                            lower_separator="_",
+                        ),
+                        "0.001_2(0.001_2)",
+                    ),
+                    (
+                        # No trimming for uncertainty == value
+                        Formatter(
+                            paren_uncertainty=True,
+                            paren_uncertainty_trim=True,
+                            lower_separator="_",
+                        ),
[…]
