+++ src/sciform/global_configuration.py
@@ -47,7 +47,7 @@ def set_global_options(  # noqa: PLR0913
     paren_uncertainty: bool | None = None,
     pdg_sig_figs: bool | None = None,
     left_pad_matching: bool | None = None,
-    paren_uncertainty_trim_digits: bool | None = None,
+    paren_uncertainty_trim: bool | None = None,
     pm_whitespace: bool | None = None,
     add_c_prefix: bool = False,
     add_small_si_prefixes: bool = False,
@@ -79,7 +79,7 @@ def set_global_options(  # noqa: PLR0913
         paren_uncertainty=paren_uncertainty,
         pdg_sig_figs=pdg_sig_figs,
         left_pad_matching=left_pad_matching,
-        paren_uncertainty_trim_digits=paren_uncertainty_trim_digits,
+        paren_uncertainty_trim=paren_uncertainty_trim,
         pm_whitespace=pm_whitespace,
         add_c_prefix=add_c_prefix,
         add_small_si_prefixes=add_small_si_prefixes,
@@ -130,7 +130,7 @@ class GlobalOptionsContext:
         paren_uncertainty: bool | None = None,
         pdg_sig_figs: bool | None = None,
         left_pad_matching: bool | None = None,
-        paren_uncertainty_trim_digits: bool | None = None,
+        paren_uncertainty_trim: bool | None = None,
         pm_whitespace: bool | None = None,
         add_c_prefix: bool = False,
         add_small_si_prefixes: bool = False,
@@ -157,7 +157,7 @@ class GlobalOptionsContext:
             paren_uncertainty=paren_uncertainty,
             pdg_sig_figs=pdg_sig_figs,
             left_pad_matching=left_pad_matching,
-            paren_uncertainty_trim_digits=paren_uncertainty_trim_digits,
+            paren_uncertainty_trim=paren_uncertainty_trim,
             pm_whitespace=pm_whitespace,
             add_c_prefix=add_c_prefix,
             add_small_si_prefixes=add_small_si_prefixes,
