+++ docs/source/options.rst
@@ -685,41 +685,95 @@ value formatting.
 Parentheses Uncertainty
 -----------------------
 
-Instead of displaying ``123.456 ± 0.789``, there is a notation where
-the uncertainty is shown in parentheses after the value as
-``123.456(789)``.
-Here the ``(789)`` in parentheses is meant to be "matched up" with the
-final three digits of the value so that the 9 in the uncertainty is
-understood to appear in the thousandths place.
-This format is described in the
-`BIPM Guide Section 7.2.2 <https://www.bipm.org/documents/20126/2071204/JCGM_100_2008_E.pdf/cb0ef43f-baa5-11cf-3f85-4dcd86f77bd6#page=37>`_.
-We call this format "parentheses uncertainty" mode.
-:mod:`sciform` provides this functionality via the ``paren_uncertainty``
-option:
+The
+`BIPM Guide Section 7.2.2 <https://www.bipm.org/documents/20126/2071204/JCGM_100_2008_E.pdf/cb0ef43f-baa5-11cf-3f85-4dcd86f77bd6#page=37>`_
+Provides three example value/uncertainty formats::
 
->>> formatter = Formatter(paren_uncertainty=True)
->>> print(formatter(123.456, 0.789))
-123.456(789)
+  100,021 47 ± 0,000 35 g
+  100,021 47(0,000 35) g
+  100,021 47(35) g
 
-Or with other options:
+In the first example the value and uncertainty are shown as regular
+numbers separated by a ``±``.
+In the second example the uncertainty is shown after the value inside
+parentheses.
+The third example is like the second but all leading zeros and separator
+characters have been removed.
+In the third example it is clear to see exactly which digits of the
+value are uncertain and by how much.
 
->>> formatter = Formatter(ndigits=2, paren_uncertainty=True)
->>> print(formatter(123.456, 0.789))
-123.46(79)
+:mod:`sciform` provides the ability to realize all three of these
+formatting strategies by using the ``paren_uncertainty`` and
+``paren_uncertainty_trim`` options.
+Note of course that the BIPM examples include a unit, ``g`` but the
+:mod:`sciform` examples do not, because :mod:`sciform` does not support
+appending units to numbers.
+:mod:`sciform` does support
+:ref:`SI prefix translations <exp_str_replacement>` to which the user
+could manually append simple unit strings.
+
+>>> value = 100.02147
+>>> uncertainty = 0.00035
+>>>
 >>> formatter = Formatter(
-...     ndigits=2,
-...     exp_mode="scientific",
+...     paren_uncertainty=False,
+...     decimal_separator=",",
+...     lower_separator=" ",
+... )
+>>> print(formatter(value, uncertainty))
+100,021 47 ± 0,000 35
+>>> formatter = Formatter(
+...     paren_uncertainty=True,
+...     paren_uncertainty_trim=False,
+...     decimal_separator=",",
+...     lower_separator=" ",
+... )
+>>> print(formatter(value, uncertainty))
+100,021 47(0,000 35)
+>>> formatter = Formatter(
+...     paren_uncertainty=True,
+...     paren_uncertainty_trim=True,
+...     decimal_separator=",",
+...     lower_separator=" ",
+... )
+>>> print(formatter(value, uncertainty))
+100,021 47(35)
+
+``paren_uncertainty_trim`` eliminates all separators which are not the
+decimal separator.
+
+>>> value = 100.0215
+>>> uncertainty = 0.0035
+>>> formatter = Formatter(
+...     paren_uncertainty=True,
+...     paren_uncertainty_trim=False,
+...     decimal_separator=",",
+...     lower_separator=" ",
+... )
+>>> print(formatter(value, uncertainty))
+100,021 5(0,003 5)
+>>> formatter = Formatter(
[…]
