### 2024-02-05 02aff47
use GlobalOptionsContext for doctest
