### 2024-02-05 ddaa6ba
options.rst improvements
