+++ tests/test_output_conversion.py
@@ -150,8 +150,8 @@ class TestLatexConversion(unittest.TestCase):
                 "(0.000000(1.234560))e+02",
                 (
                     "(0.000000(1.234560))e+02",
-                    "(0.000000(1.234560))×10<sup>2</sup>",
-                    r"$(0.000000(1.234560))\times10^{2}$",
+                    "0.000000(1.234560)×10<sup>2</sup>",
+                    r"$0.000000(1.234560)\times10^{2}$",
                 ),
             ),
             (
@@ -214,8 +214,8 @@ class TestLatexConversion(unittest.TestCase):
                 "(0.123456(789))e+03",
                 (
                     "(0.123456(789))e+03",
-                    "(0.123456(789))×10<sup>3</sup>",
-                    r"$(0.123456(789))\times10^{3}$",
+                    "0.123456(789)×10<sup>3</sup>",
+                    r"$0.123456(789)\times10^{3}$",
                 ),
             ),
             (
@@ -240,6 +240,14 @@ class TestLatexConversion(unittest.TestCase):
                     "(7.8900 +/- 0.0001)e-02",
                     "(7.8900 ± 0.0001)×10<sup>-2</sup>",
                     r"$(7.8900\:\pm\:0.0001)\times10^{-2}$",
+                ),
+            ),
+            (
+                "7.8900(0.0001)×10⁻²",
+                (
+                    "(7.8900(0.0001))e-02",
+                    "7.8900(0.0001)×10<sup>-2</sup>",
+                    r"$7.8900(0.0001)\times10^{-2}$",
                 ),
             ),
             (
