### 2024-02-06 7d5731b
remove redundant parentheses entirely for paren_uncertainty mode
