+++ tests/test_output_conversion.py
@@ -147,9 +147,9 @@ class TestLatexConversion(unittest.TestCase):
                 ),
             ),
             (
-                "(0.000000(1.234560))e+02",
+                "0.000000(1.234560)e+02",
                 (
-                    "(0.000000(1.234560))e+02",
+                    "0.000000(1.234560)e+02",
                     "0.000000(1.234560)×10<sup>2</sup>",
                     r"$0.000000(1.234560)\times10^{2}$",
                 ),
@@ -211,9 +211,9 @@ class TestLatexConversion(unittest.TestCase):
                 ),
             ),
             (
-                "(0.123456(789))e+03",
+                "0.123456(789)e+03",
                 (
-                    "(0.123456(789))e+03",
+                    "0.123456(789)e+03",
                     "0.123456(789)×10<sup>3</sup>",
                     r"$0.123456(789)\times10^{3}$",
                 ),
@@ -245,7 +245,7 @@ class TestLatexConversion(unittest.TestCase):
             (
                 "7.8900(0.0001)×10⁻²",
                 (
-                    "(7.8900(0.0001))e-02",
+                    "7.8900(0.0001)e-02",
                     "7.8900(0.0001)×10<sup>-2</sup>",
                     r"$7.8900(0.0001)\times10^{-2}$",
                 ),
@@ -414,10 +414,10 @@ class TestLatexConversion(unittest.TestCase):
                             paren_uncertainty=True,
                         ),
                         (
-                            "(12.345_678(255))%",
-                            "(12.345_678(255))%",
-                            "(12.345_678(255))%",
-                            r"(12.345\_678(255))\%",
+                            "12.345_678(255)%",
+                            "12.345_678(255)%",
+                            "12.345_678(255)%",
+                            r"12.345\_678(255)\%",
                         ),
                     ),
                 ],
@@ -447,7 +447,7 @@ class TestLatexConversion(unittest.TestCase):
     def test_special_repr_output(self):
         cases_list = [
             "6.26070e-04",
-            "(0.000000(1.234560))e+02",
+            "0.000000(1.234560)e+02",
             "000_000_004_567_899.765_432_1",
             "(nan)%",
             "123000 ppm",
