### 2024-02-06 8a3e28e
remove code to detect paren uncertainty
