+++ src/sciform/format_utils.py
@@ -504,15 +504,8 @@ def construct_val_unc_exp_str(  # noqa: PLR0913
     )
 
     if exp_str == "":
-        val_unc_exp_str = f"{val_unc_str}{exp_str}"
+        val_unc_exp_str = val_unc_str
     elif paren_uncertainty:
-        """
-        "1234(12)" val_unc_str along with exp_str "k" will be formatted as
-        "1234(12) k whereas 1234(12) with exp_str "e+03" will be
-        formatted as (1234(12))e+03.
-        "1234 ± 12" will be formatted with parentheses as "(1234 ± 12) k" or
-        "(1234 ± 12)e+03"
-        """
         val_unc_exp_str = f"{val_unc_str}{exp_str}"
     else:
         val_unc_exp_str = f"({val_unc_str}){exp_str}"
