### 2024-02-07 #121 @Batalex
---

🎉 `sciform` has been approved by pyOpenSci! Thank you, @jagerber48, for submitting `sciform` and many thanks to @isabelizimm and @machow for reviewing this package! 😸

We took longer than I expected at first glance, but I think it is for the best. The package was already in a good shape, but it is now definitely more welcoming to new contributors, hence more future-proof.

- [ ] Add the badge for pyOpenSci peer-review to the README.md of <package-name-here>. The badge should be `[![pyOpenSci](https://tinyurl.com/y22nb8up)](https://github.com/pyOpenSci/software-review/issues/121)`.

   @jagerber48: Please let me know once you are done.

---
