+++ src/sciform/parser.py
@@ -10,52 +10,85 @@ finite_val_pattern = r"(?:[-+]?[\d.,_ ]*\d)"
 non_finite_val_pattern = r"(?:[+-]?(?:nan|NAN|inf|INF))"
 any_val_pattern = rf"(?:{finite_val_pattern}|{non_finite_val_pattern})"
 pm_pattern = rf"(?:(?P<pm_val>{any_val_pattern})(?: ± |±)(?P<pm_unc>{any_val_pattern}))"
-paren_pattern = rf"(?:(?P<paren_val>{any_val_pattern})\((?P<paren_unc>{any_val_pattern})\))"
+paren_pattern = (
+    rf"(?:(?P<paren_val>{any_val_pattern})\((?P<paren_unc>{any_val_pattern})\))"
+)
 
 ascii_exp_pattern = r"(?P<ascii_exp>(?P<ascii_base>[eEbB])(?P<ascii_exp_val>[+-]\d+))"
 uni_exp_pattern = r"(?P<uni_exp>×(?P<uni_base>10|2)(?P<uni_exp_val>[⁺⁻]?[⁰¹²³⁴⁵⁶⁷⁸⁹]+))"
 prefix_exp_pattern = r"(?:\ (?P<prefix_exp>[a-zA-zμ]+))"
 percent_exp_pattern = r"(?P<percent_exp>%)"
-any_exp_pattern = rf"(?:{ascii_exp_pattern}|{uni_exp_pattern}|{prefix_exp_pattern}|{percent_exp_pattern})"
+any_exp_pattern = (
+    rf"(?:{ascii_exp_pattern}"
+    rf"|{uni_exp_pattern}"
+    rf"|{prefix_exp_pattern}"
+    rf"|{percent_exp_pattern})"
+)
 
 no_exp_pattern = rf"^{pm_pattern}|(?P<non_finite_val>{non_finite_val_pattern})$"
-optional_exp_pattern = rf"^(?:(?P<val>{finite_val_pattern})|{paren_pattern})(?P<exp>{any_exp_pattern})?$"
-always_exp_pattern = rf"^(?:\((?P<non_finite_val>{non_finite_val_pattern})\)|\({pm_pattern}\))(?P<exp>{any_exp_pattern})$"
+optional_exp_pattern = (
+    rf"^(?:(?P<val>{finite_val_pattern})|{paren_pattern})(?P<exp>{any_exp_pattern})?$"
+)
+always_exp_pattern = (
+    rf"^(?:\((?P<non_finite_val>{non_finite_val_pattern})\)"
+    rf"|\({pm_pattern}\))"
+    rf"(?P<exp>{any_exp_pattern})$"
+)
 
 superscript_translation = str.maketrans("⁺⁻⁰¹²³⁴⁵⁶⁷⁸⁹", "+-0123456789")
 
 
-def extract_exp_base_exp_val(
+def _get_ascii_base_exp_val(match: re.Match) -> tuple[int, int]:
+    base = match.group("ascii_base")
+    if base.lower() == "e":
+        base = 10
+    elif base.lower() == "b":
+        base = 2
+    else:
+        raise ValueError
+    exp_val = int(match.group("ascii_exp_val"))
+    return base, exp_val
+
+
+def _get_unicode_base_exp_val(match: re.Match) -> tuple[int, int]:
+    base = int(match.group("uni_base"))
+    exp_val = int(match.group("uni_exp_val").translate(superscript_translation))
+    return base, exp_val
+
+
+def _get_prefix_base_exp_val(prefix_exp: str) -> tuple[int, int]:
+    candidate_base_exp_val_pairs = []
+    for key, value in prefix_module.si_val_to_prefix_dict.items():
+        if prefix_exp == value:
[…]
