+++ src/sciform/parser.py
@@ -1,3 +1,7 @@
+# ruff: noqa: ERA001
+
+"""Parse formatted strings back into value/uncertainty pairs."""
+
 from __future__ import annotations
 
 import re
@@ -6,34 +10,70 @@ from decimal import Decimal
 from sciform import modes
 from sciform import prefix as prefix_module
 
-finite_val_pattern = r"(?:[-+]?[\d.,_ ]*\d)"
-non_finite_val_pattern = r"(?:[+-]?(?:nan|NAN|inf|INF))"
-any_val_pattern = rf"(?:{finite_val_pattern}|{non_finite_val_pattern})"
-pm_pattern = rf"(?:(?P<pm_val>{any_val_pattern})(?: ± |±)(?P<pm_unc>{any_val_pattern}))"
-paren_pattern = (
-    rf"(?:(?P<paren_val>{any_val_pattern})\((?P<paren_unc>{any_val_pattern})\))"
+# language=pythonverboseregexp
+finite_val_pattern = r"""  
+(
+  [ +-]?  # Sign
+  (0*|\ *)  # Leading zeros or spaces
+  (\d|\d\d|\d\d\d)((,\d\d\d)*|(\.\d\d\d)*|(\d\d\d)*)  # Leading digit groups 
+  (  # Start of optional fractional part
+    ((?<!,\d\d\d),|(?<!\.\d\d\d)\.)  # decimal_separator != upper_separator
+    ((\d\d\d_)*|(\d\d\d\ )*|(\d\d\d)*)(\d|\d\d|\d\d\d)  # Trailing digit groups
+  )?  # End of optional fractional part
 )
+"""
+# language=pythonverboseregexp
+non_finite_val_pattern = r"(nan|NAN|([ +-]?(inf|INF)))"
+# language=pythonverboseregexp
+any_val_pattern = rf"({finite_val_pattern}|{non_finite_val_pattern})"
 
+# language=pythonverboseregexp
+pm_pattern = rf"""
+(
+  (?P<pm_val>{any_val_pattern})  # Value
+  (\ (±|\+/-)\ |(±|\+/-))  # +/- symbol (optional whitespace)
+  (?P<pm_unc>{any_val_pattern})  # Uncertainty
+)
+"""
+# language=pythonverboseregexp
+paren_pattern = rf"""
+((?P<paren_val>{any_val_pattern})\((?P<paren_unc>{any_val_pattern})\))
+"""
+
+# language=pythonverboseregexp
 ascii_exp_pattern = r"(?P<ascii_exp>(?P<ascii_base>[eEbB])(?P<ascii_exp_val>[+-]\d+))"
+# language=pythonverboseregexp
 uni_exp_pattern = r"(?P<uni_exp>×(?P<uni_base>10|2)(?P<uni_exp_val>[⁺⁻]?[⁰¹²³⁴⁵⁶⁷⁸⁹]+))"
-prefix_exp_pattern = r"(?:\ (?P<prefix_exp>[a-zA-zμ]+))"
+# language=pythonverboseregexp
+prefix_exp_pattern = r"(\ (?P<prefix_exp>[a-zA-zμ]+))"
+# language=pythonverboseregexp
 percent_exp_pattern = r"(?P<percent_exp>%)"
-any_exp_pattern = (
-    rf"(?:{ascii_exp_pattern}"
-    rf"|{uni_exp_pattern}"
-    rf"|{prefix_exp_pattern}"
-    rf"|{percent_exp_pattern})"
+# language=pythonverboseregexp
+any_exp_pattern = rf"""
+(
+  {ascii_exp_pattern}
+  |{uni_exp_pattern}
+  |{prefix_exp_pattern}
+  |{percent_exp_pattern}
[…]
