### 2024-02-10 5ce7453
scinum input and other updates
