### 2024-02-10 9d670b0
grouping refactor, more tests
