+++ tests/test_string_parser.py
@@ -1,8 +1,11 @@
 import unittest
+from decimal import Decimal
+from math import isnan
 from typing import Optional
 
 from sciform import SciNum
 from sciform.format_utils import Number
+from sciform.parser import parse_val_unc_from_str
 
 CasesList = list[tuple[str, tuple[Number, Optional[Number]]]]
 
@@ -10,233 +13,279 @@ CasesList = list[tuple[str, tuple[Number, Optional[Number]]]]
 NAN = float("nan")
 INF = float("inf")
 
+cases = [
+    (
+        "000_000_123_456_789.654_321",
+        (123456789.654321, None),
+    ),
+    (
+        "123_456_789.654_321",
+        (123456789.654321, None),
+    ),
+    (
+        "nan",
+        (NAN, None),
+    ),
+    (
+        "inf",
+        (INF, None),
+    ),
+    (
+        "-inf",
+        (-INF, None),
+    ),
+    (
+        "NAN",
+        (NAN, None),
+    ),
+    (
+        "INF",
+        (INF, None),
+    ),
+    (
+        "-INF",
+        (-INF, None),
+    ),
+    (
+        "(nan)%",
+        (NAN, None),
+    ),
+    (
+        "(nan)e+00",
+        (NAN, None),
+    ),
+    (
+        "0.000123e+03",
+        (0.123, None),
+    ),
+    (
+        "0.123",
+        (0.123, None),
+    ),
+    (
+        "123e-09",
+        (123e-9, None),
+    ),
+    (
+        "123 ppb",
+        (123e-9, None),
+    ),
+    (
+        "7.89×10²",
+        (7.89e2, None),
+    ),
+    (
+        "1×2¹⁰",
+        (1024, None),
+    ),
+    (
+        "1b+10",
+        (1024, None),
+    ),
+    (
+        "0b+00",
+        (0, None),
+    ),
+    (
+        "1.618033e+04",
+        (16180.33, None),
+    ),
+    (
+        "-1.618033e+04",
+        (-16180.33, None),
+    ),
+    (
+        "12.3456%",
+        (0.123456, None),
+    ),
+    (
+        "3.1415 q",
+        (3.1415e-30, None),
+    ),
+    (
+        "3.1415 Q",
+        (3.1415e30, None),
+    ),
+    (
+        "1 Ki",
+        (1024, None),
+    ),
+    (
+        "1 Pi",
+        (2**50, None),
+    ),
+    (
+        "123.456.789,987_654_321",
+        ("123456789.987654321", None),
+    ),
+    (
+        "123,456,789.987 654 321",
+        ("123456789.987654321", None),
+    ),
+    (
+        "123(0)",
+        (123, 00),
+    ),
+    (
+        "123(nan)",
+        (123, NAN),
+    ),
+    (
+        "123(inf)",
+        (123, INF),
+    ),
+    (
+        "1234.56(7.89)e-01",
+        (123.456, 0.789),
+    ),
+    (
+        "nan(1.2)",
+        (NAN, 1.2),
+    ),
+    (
+        "nan(nan)",
+        (NAN, NAN),
+    ),
+    (
+        "10.000 ± 0.035",
+        (10, 0.035),
+    ),
+    (
+        "(3.1416 ± 0.0016) M",
+        (3.1416e6, 0.0016e6),
+    ),
+    (
+        "(12.345_678 ± 0.000_255)%",
+        (0.12345678, 0.00000255),
+    ),
+    (
+        "123.456(789) k",
+        (123.456e3, 0.789e3),
+    ),
+    (
+        "123.456(789) ppm",
[…]
