### 2024-02-10 3c12826
quick test scinum with fancy inputs for val and uncertainty
