### 2024-02-10 924c6b6
Various
* add merging in of global translations for prefix parsing
* Clean up ambiguous decimal separator parsing
* More verbose exception handling
* Tests to reach 100% coverage again
