+++ tests/test_sci_num.py
@@ -12,3 +12,11 @@ class TestSciNum(unittest.TestCase):
             repr(SciNum(123.456, 0.0000023)),
             "SciNum(123.456, 0.0000023)",
         )
+
+    def test_double_uncertainty_input(self):
+        self.assertRaises(
+            ValueError,
+            SciNum,
+            "1 +/- 2",
+            2,
+        )
