### 2024-02-10 4d710f8
add pyopensci badge
