### 2024-02-11 #121 @jagerber48
@Batalex Ok, I completed all those items. Setup Zenodo and was able to do a release to get a DOI (added a badge to readme). Also added PyOpenSci badge (very excited about that!). Completed the maintainer survey.

@Batalex @machow @isabelizimm Thank you all so much for your help, patience, and time spent on this review. This has been a very valuable learning experience for me and I think `sciform` is very much improved thanks to all of your feedback. My goal is for `sciform` to be something that can help people and I feel like going through this review process really helped me towards that goal!
