+++ CHANGELOG.rst
@@ -8,7 +8,10 @@ This project adheres to `Semantic Versioning <https://semver.org/>`_.
 Unreleased
 ----------
 
-* There are no unreleased changes.
+Added
+^^^^^
+
+* Added badge for Zenodo.
 
 ----
 
