+++ CHANGELOG.rst
@@ -17,24 +17,40 @@ Added
   presentation of the uncertainty in ``paren_uncertainty`` mode in which
   leading zeros and separator characters are **not** stripped from the
   string.
-  E.g. ``paren_uncertainties_trim=True`` will give ``123.002 3(21)``
+  E.g. ``paren_uncertainties_trim=True`` will give
+
+    123.002 3(21)
+
   while ``paren_uncertainties_trim=False`` will give
-  ``123.002 3(0.002 1)``.
+
+    123.002 3(0.002 1)
 
 * Added badge for Zenodo.
 
 Removed
 ^^^^^^^
 
-* **[BREAKING]** Removed the ``paren_uncertainties_separator`` option.
+* **[BREAKING]** Removed the ``paren_uncertainties_separators`` option.
   This options made it possible (when ``False``) to optionally strip all
-  separator characters, including the decimal separator.
+  separator characters, including the decimal separator, from the
+  uncertainty in ``paren_uncertainty`` mode.
   This lead to the possibility of value/uncertainty pairs like
-  ``123 456.789 8 ± 123.456 7`` being represented as
-  ``123 456.789 8(1234567)``.
-  ``sciform`` will now display this as ``123 456.789 8(123.456 7)``
-  if ``paren_uncertainty_strip=False`` or ``123 456.789 8(123.4567)`` if
-  ``paren_uncertainty_strip=True``, but always retaining the decimal
+
+    123 456.789 8 ± 123.456 7
+
+  being represented as
+
+    123 456.789 8(1234567)
+
+  ``sciform`` will now display this as
+
+    123 456.789 8(123.456 7)
+
+  if ``paren_uncertainty_strip=False`` or
+
+    123 456.789 8(123.4567)
+
+  if ``paren_uncertainty_strip=True``, but always retaining the decimal
   separator.
   In most cases many fewer significant digits of the uncertainty are
   displayed and the resulting outputs don't look as egregious when the
