### 2024-02-11 c489c62
move input parser to parser.py
