+++ src/sciform/formatter.py
@@ -7,6 +7,7 @@ from typing import TYPE_CHECKING, Literal
 from sciform.formatting import format_from_options
 from sciform.options.conversion import populate_options
 from sciform.options.input_options import InputOptions
+from sciform.parser import parse_val_unc_from_input
 
 if TYPE_CHECKING:  # pragma: no cover
     from sciform import modes
@@ -341,6 +342,7 @@ class Formatter:
         :param uncertainty: Optional uncertainty to be formatted.
         :type uncertainty: ``Decimal | float | int | str | None``
         """
+        value, uncertainty = parse_val_unc_from_input(value, uncertainty)
         return format_from_options(
             value,
             uncertainty,
