### 2024-02-11 582fa58
normalize all decimal input
