### 2024-02-11 818da81
store value/uncertainty in FormattedNumber
