+++ src/sciform/formatting.py
@@ -54,7 +54,7 @@ def format_from_options(
         formatted_str = format_val_unc(value, uncertainty, finalized_options)
     else:
         formatted_str = format_num(value, finalized_options)
-    return FormattedNumber(formatted_str, populated_options)
+    return FormattedNumber(formatted_str, value, uncertainty, populated_options)
 
 
 def format_non_finite(num: Decimal, options: FinalizedOptions) -> str:
@@ -341,6 +341,10 @@ class FormattedNumber(str):
     """
 
     __slots__ = {
+        "value": "The value that was formatted to generate the "
+        ":class:`FormattedNumber`.",
+        "uncertainty": "The uncertainty that was formatted to generate the "
+        ":class:`FormattedNumber`.",
         "populated_options": "Record of the :class:`PopulatedOptions` used to "
         "generate the :class:`FormattedNumber`.",
     }
@@ -348,10 +352,14 @@ class FormattedNumber(str):
     def __new__(
         cls: type[Self],
         formatted_str: str,
+        value: Number,
+        uncertainty: Number | None,
         populated_options: PopulatedOptions,
     ) -> Self:
         """Get a new string."""
         obj = super().__new__(cls, formatted_str)
+        obj.value = value
+        obj.uncertainty = uncertainty
         obj.populated_options = populated_options
         return obj
 
