+++ docs/source/usage.rst
@@ -105,7 +105,136 @@ algorithm:
 #. The value and uncertainty mantissas are formatted together with the
    exponent according to other user-selected display options.
 
-.. _global_config:
+.. _formatted_input:
+
+Formatted Input
+===============
+
+Both the :class:`Formatter` format function and :class:`SciNum`
+constructor accept :class:`str`, :class:`int`, :class:`float`, and
+:class:`Decimal` input for the value and optionally accept the same
+types, or for the uncertainty.
+
+>>> from decimal import Decimal
+>>> from sciform import Formatter
+>>>
+>>> formatter = Formatter(ndigits=4)
+>>> print(formatter("32", "9"))
+32.000 ± 9.000
+>>> print(formatter(32, 9))
+32.000 ± 9.000
+>>> print(formatter(32.00, 9.00))
+32.000 ± 9.000
+>>> print(formatter(Decimal("32.000"), Decimal("9.000")))
+32.000 ± 9.000
+
+These example strings can be natively cast to numeric types like
+:class:`int`, :class:`float`, or :class:`Decimal`.
+However, the :class:`Formatter` and :class:`SciNum` also accept
+formatted strings which contain the numeric information, but with more
+sophisticated formatting.
+
+>>> print(formatter("+  123_456,789 987 n"))
+0.0001235
+
+Here the string ``"+  123_45,789 987 n"`` is re-interpreted by the
+:class:`Formatter` as the number ``12345.789987e-09``.
+The ``e-09`` arises as the reverse translation of the SI "nano" prefix
+``n``.
+When this is formatted in fixed point mode with 4 significant digits the
+result is ``"0.0001235"``.
+
+Instead of passing in two inputs, one representing the value and one
+representing the value, it is possible to pass in one input which
+contains information about both the value and the uncertainty.
+
+>>> print(formatter("(123.0 ± 0.4) m"))
+0.1230000 ± 0.0004000
+>>> print(formatter("(123 +/- 0.4) m"))
+0.1230000 ± 0.0004000
+
+Note that the first example input string is an example :mod:`sciform`
+output string that would have resulted from various options and inputs
+while the second example input string is not an example :mod:`sciform`
+output string.
+Any :mod:`sciform` output string can also be used as an input string.
+However, some input strings can be supplied which are not valid
+:mod:`sciform` outputs.
+
+  * For inputs, it is not necessary that the value and uncertainty are
+    expressed to the same lowest decimal place.
+    By contrast, :mod:`sciform` outputs will always format the value and
+    uncertainty to the same lowest decimal place.
+  * For inputs, the ASCII ``+/-`` symbols may be used to separate the
+    value from the uncertainty.
+    By contrast, :mod:`sciform` will always use the Unicode ``±`` symbol
+    for direct outputs (noting that the ``±`` symbol can be converted
+    into ``+/-`` afterwards using :meth:`FormattedNumber.as_ascii`, see
+    :ref:`output_conversion`).
+
+Like :mod:`sciform` outputs, the input value/uncertainty pairs must
+always share the exponent string.
+
+.. doctest::
+  :options: -IGNORE_EXCEPTION_DETAIL
+
+  >>> print(formatter("(1.2 +/- 0.1)e+03"))
+  1200.0 ± 100.0
+  >>> print(formatter("1.2e+03 +/- 0.1e+03"))
+  Traceback (most recent call last):
+    ...
+  ValueError: Input string "1.2e+03 +/- 0.1e+03" does not match any expected input format.
+
+Inputs can use parentheses uncertainty notation and will be
+interpreted using the following rules.
+
+  * The value string will always be directly converted to a float.
+  * If the uncertainty string contains a decimal symbol then it will be
+    directly converted to a float.
[…]
