### 2024-02-12 e56ede6
fix parsing for cases like 1.0 +/- 234 (previously this was erroneously caught as a paren uncertainty invalid case)
