+++ tests/test_string_parser.py
@@ -242,6 +242,10 @@ shared_cases = [
         "123.45",
         (123.45, None),
     ),
+    (
+        "1.0 +/- 23",
+        (1, 23),
+    ),
 ]
 
 
