+++ tests/test_string_parser.py
@@ -246,6 +246,10 @@ shared_cases = [
         "1.0 +/- 23",
         (1, 23),
     ),
+    (
+        "1.0+/-23",
+        (1, 23),
+    ),
 ]
 
 
