+++ pyproject.toml
@@ -66,7 +66,6 @@ ignore = [
     'TD003',
     'FIX002',
     'RET504',
-    'Q003',
 ]
 
 [tool.ruff.format]
