### 2024-02-13 b0bafa7
parse value in formatting function, type cast in SciNum
