### 2024-02-13 39630b1
decimal normalization tests
