+++ src/sciform/scinum.py
@@ -2,7 +2,6 @@
 
 from __future__ import annotations
 
-from decimal import Decimal
 from typing import TYPE_CHECKING
 
 from sciform.formatting import FormattedNumber, format_from_options
@@ -10,6 +9,8 @@ from sciform.fsml import format_options_from_fmt_spec
 from sciform.parser import parse_val_unc_from_input
 
 if TYPE_CHECKING:  # pragma: no cover
+    from decimal import Decimal
+
     from sciform.format_utils import Number
 
 
