+++ src/sciform/scinum.py
@@ -34,6 +34,15 @@ class SciNum:
     >>> print(f"{num:#!2r()}")
     0.123456654(23)e+06
 
+    Inputs may be :class:`str`, :class:`int`, :class:`float`, or
+    :class:`Decimal`.
+
+    :class:`float` inputs are first converted to :class:`str` to
+    retrieve the shortest round-trippable decimal representation of the
+    :class:`float`. For more details see :ref:`float_issues`.
+
+    :class:`Decimal` inputs are normalized upon input.
+
     Formatted input can also be passed into :class:`SciNum`. For more
     details see :ref:`formatted_input`.
 
