### 2024-02-13 7c1d426
remove redundant decimal casting
