+++ src/sciform/parser.py
@@ -133,9 +133,12 @@ def _get_prefix_base_exp_val(prefix_exp: str) -> tuple[int, int]:
         msg = f'Unrecognized prefix: "{prefix_exp}". Unable to parse input.'
         raise ValueError(msg)
     if len(set(candidate_base_exp_val_pairs)) > 1:
+        candidate_exps = [
+            candidate_pair[1] for candidate_pair in candidate_base_exp_val_pairs
+        ]
         msg = (
-            f'Multiple translations found for prefix "{prefix_exp}". Unable to '
-            f"parse input."
+            f'Multiple translations found for prefix "{prefix_exp}: {candidate_exps}. '
+            f"Unable to parse input."
         )
         raise ValueError(msg)
     base, exp_val = candidate_base_exp_val_pairs[0]
