+++ src/sciform/formatter.py
@@ -348,14 +348,13 @@ class Formatter:
         Format a value or value/uncertainty pair.
 
         Inputs may be :class:`str`, :class:`int`, :class:`float`, or
-        :class:`Decimal`.
-
-        :class:`float` inputs are first converted to
+        :class:`Decimal`. :class:`float` inputs are first converted to
         :class:`str` to retrieve the shortest round-trippable decimal
         representation of the :class:`float`. For more details see
-        :ref:`float_issues`.
-
-        :class:`Decimal` inputs are normalized upon input.
+        :ref:`float_issues`. :class:`Decimal` inputs are normalized upon
+        input. That is, ``Decimal("1.000")`` is treated the same as
+        ``Decimal("1")``. Formatted input strings are also accepted.
+        See :ref:`formatted_input`.
 
         :param value: Value to be formatted.
         :type value: ``Decimal | float | int | str``
