### 2024-02-14 7a407e7
scinum variable documentation
