### 2024-02-14 13977c6
fix up decimal separator parsing
