### 2024-02-14 7ecd0ae
Clean up decimal separator parsing
