### 2024-02-14 b3e29ee
more decimal symbol fixes
