+++ tests/test_string_parser.py
@@ -392,6 +392,14 @@ class TestStringParser(unittest.TestCase):
                 "123.456,789",
                 (123456.789, None),
             ),
+            (
+                "123.456.789",
+                (123456789, None),
+            ),
+            (
+                "123,456,789",
+                (123456789, None),
+            ),
         ]
         with GlobalOptionsContext(decimal_separator=","):
             self.run_scinum_cases(cases)
@@ -413,6 +421,14 @@ class TestStringParser(unittest.TestCase):
             (
                 "123,456.789",
                 (123456.789, None),
+            ),
+            (
+                "123.456.789",
+                (123456789, None),
+            ),
+            (
+                "123,456,789",
+                (123456789, None),
             ),
         ]
         with GlobalOptionsContext(decimal_separator="."):
