### 2024-02-14 29f5aa2
simplify _infer_decimal_separator
