### 2024-02-16 4dca8b7
0 and nan handling tests
