### 2024-02-16 a9af9d5
refactor get_sign_str and format_non_finite
