### 2024-02-17 #121 @jagerber48
@Batalex how do I add `sciform` and myself to the PyOpenSci website?
