### 2024-02-17 5e3a9c8
move FormattedNumber
