+++ src/sciform/__init__.py
@@ -1,5 +1,6 @@
 """``sciform`` is used to convert python numbers into scientific formatted strings."""
 
+from sciform.api.formatted_number import FormattedNumber
 from sciform.api.formatter import Formatter
 from sciform.api.global_configuration import (
     GlobalOptionsContext,
@@ -9,7 +10,6 @@ from sciform.api.global_configuration import (
     set_global_options,
 )
 from sciform.api.scinum import SciNum
-from sciform.formatting import FormattedNumber
 from sciform.modes import AutoDigits, AutoExpVal
 from sciform.options.input_options import InputOptions
 from sciform.options.populated_options import PopulatedOptions
