### 2024-02-17 cfcf660
rename modes.py -> options/option_types.py
