+++ src/sciform/api/global_configuration.py
@@ -4,14 +4,13 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from sciform.options import global_options
+from sciform.options import global_options, option_types
 from sciform.options.conversion import populate_options
 from sciform.options.input_options import InputOptions
 
 if TYPE_CHECKING:  # pragma: no cover
     from types import TracebackType
 
-    from sciform import modes
     from sciform.options.populated_options import PopulatedOptions
 
 
@@ -27,17 +26,17 @@ def get_global_options() -> PopulatedOptions:
 
 def set_global_options(  # noqa: PLR0913
     *,
-    exp_mode: modes.ExpMode | None = None,
-    exp_val: int | type(modes.AutoExpVal) | None = None,
-    round_mode: modes.RoundMode | None = None,
-    ndigits: int | type(modes.AutoDigits) | None = None,
-    upper_separator: modes.UpperSeparators | None = None,
-    decimal_separator: modes.DecimalSeparators | None = None,
-    lower_separator: modes.LowerSeparators | None = None,
-    sign_mode: modes.SignMode | None = None,
-    left_pad_char: modes.LeftPadChar | None = None,
+    exp_mode: option_types.ExpMode | None = None,
+    exp_val: int | type(option_types.AutoExpVal) | None = None,
+    round_mode: option_types.RoundMode | None = None,
+    ndigits: int | type(option_types.AutoDigits) | None = None,
+    upper_separator: option_types.UpperSeparators | None = None,
+    decimal_separator: option_types.DecimalSeparators | None = None,
+    lower_separator: option_types.LowerSeparators | None = None,
+    sign_mode: option_types.SignMode | None = None,
+    left_pad_char: option_types.LeftPadChar | None = None,
     left_pad_dec_place: int | None = None,
-    exp_format: modes.ExpFormat | None = None,
+    exp_format: option_types.ExpFormat | None = None,
     extra_si_prefixes: dict[int, str] | None = None,
     extra_iec_prefixes: dict[int, str] | None = None,
     extra_parts_per_forms: dict[int, str] | None = None,
@@ -110,17 +109,17 @@ class GlobalOptionsContext:
     def __init__(  # noqa: PLR0913
         self: GlobalOptionsContext,
         *,
-        exp_mode: modes.ExpMode | None = None,
-        exp_val: int | type(modes.AutoExpVal) | None = None,
-        round_mode: modes.RoundMode | None = None,
-        ndigits: int | type(modes.AutoDigits) | None = None,
-        upper_separator: modes.UpperSeparators | None = None,
-        decimal_separator: modes.DecimalSeparators | None = None,
-        lower_separator: modes.LowerSeparators | None = None,
-        sign_mode: modes.SignMode | None = None,
-        left_pad_char: modes.LeftPadChar | None = None,
+        exp_mode: option_types.ExpMode | None = None,
+        exp_val: int | type(option_types.AutoExpVal) | None = None,
+        round_mode: option_types.RoundMode | None = None,
+        ndigits: int | type(option_types.AutoDigits) | None = None,
+        upper_separator: option_types.UpperSeparators | None = None,
+        decimal_separator: option_types.DecimalSeparators | None = None,
+        lower_separator: option_types.LowerSeparators | None = None,
+        sign_mode: option_types.SignMode | None = None,
+        left_pad_char: option_types.LeftPadChar | None = None,
         left_pad_dec_place: int | None = None,
-        exp_format: modes.ExpFormat | None = None,
+        exp_format: option_types.ExpFormat | None = None,
         extra_si_prefixes: dict[int, str] | None = None,
         extra_iec_prefixes: dict[int, str] | None = None,
         extra_parts_per_forms: dict[int, str] | None = None,
