+++ tests/test_string_parser.py
@@ -3,8 +3,9 @@ from decimal import Decimal
 from math import isnan
 from typing import Optional
 
-from sciform import GlobalOptionsContext, SciNum, modes
+from sciform import GlobalOptionsContext, SciNum
 from sciform.format_utils import Number
+from sciform.options import option_types
 from sciform.parser import parse_val_unc_from_str
 
 CasesList = list[tuple[str, tuple[Number, Optional[Number]]]]
@@ -265,7 +266,7 @@ class TestStringParser(unittest.TestCase):
     def run_direct_cases(
         self,
         cases: CasesList,
-        decimal_separator: modes.DecimalSeparators = ".",
+        decimal_separator: option_types.DecimalSeparators = ".",
     ):
         for string, (val, unc) in cases:
             expected_val = Decimal(str(val))
