+++ tests/test_invalid_options.py
@@ -2,13 +2,15 @@ import unittest
 from decimal import Decimal
 
 from sciform import Formatter
-from sciform.format_utils import (
+from sciform.format_utils.exponent_utils import get_prefix_dict
+from sciform.format_utils.number_utils import (
     get_mantissa_exp_base,
-    get_prefix_dict,
-    get_round_digit,
-    get_sign_str,
     get_top_digit,
     get_top_digit_binary,
+)
+from sciform.format_utils.rounding_utils import get_round_digit
+from sciform.format_utils.string_utils import (
+    get_sign_str,
 )
 from sciform.formatting import format_non_finite
 from sciform.options import option_types
