+++ src/sciform/formatting.py
@@ -7,21 +7,22 @@ from typing import TYPE_CHECKING, cast
 from warnings import warn
 
 from sciform.api.formatted_number import FormattedNumber
-from sciform.format_utils import (
+from sciform.format_utils.exponent_utils import get_val_unc_exp
+from sciform.format_utils.grouping_utils import add_separators
+from sciform.format_utils.number_utils import (
+    get_mantissa_exp_base,
+    get_val_unc_top_digit,
+)
+from sciform.format_utils.rounding_utils import get_round_digit, round_val_unc
+from sciform.format_utils.string_utils import (
     Number,
     construct_val_unc_exp_str,
     construct_val_unc_str,
     format_num_by_top_bottom_dig,
     get_exp_str,
-    get_mantissa_exp_base,
-    get_round_digit,
     get_sign_str,
-    get_val_unc_exp,
     get_val_unc_mantissa_strs,
-    get_val_unc_top_digit,
-    round_val_unc,
 )
-from sciform.grouping import add_separators
 from sciform.options.conversion import finalize_populated_options, populate_options
 from sciform.options.option_types import (
     AutoExpVal,
