+++ src/sciform/format_utils/string_utils.py
@@ -1,4 +1,5 @@
 """String assembly utilities."""
+
 from __future__ import annotations
 
 import re
