+++ src/sciform/format_utils/rounding_utils.py
@@ -1,4 +1,5 @@
 """Rounding utilities."""
+
 from __future__ import annotations
 
 from decimal import Decimal
