+++ src/sciform/api/scinum.py
@@ -4,9 +4,9 @@ from __future__ import annotations
 
 from typing import TYPE_CHECKING
 
-from sciform.formatting import format_from_options
-from sciform.fsml import format_options_from_fmt_spec
-from sciform.parser import parse_val_unc_from_input
+from sciform.formatting.fsml import format_options_from_fmt_spec
+from sciform.formatting.number_formatting import format_from_options
+from sciform.formatting.parser import parse_val_unc_from_input
 
 if TYPE_CHECKING:  # pragma: no cover
     from decimal import Decimal
