+++ tests/test_output_conversion.py
@@ -3,10 +3,10 @@ from __future__ import annotations
 import unittest
 
 from sciform import Formatter
-from sciform.formatting import FormattedNumber
+from sciform.formatting.number_formatting import FormattedNumber
 from sciform.options.conversion import populate_options
 from sciform.options.input_options import InputOptions
-from sciform.output_conversion import convert_sciform_format
+from sciform.formatting.output_conversion import convert_sciform_format
 
 ValFormatterCases = list[
     tuple[float, list[tuple[Formatter, tuple[str, str, str, str]]]]
