+++ src/sciform/formatting/__init__.py
@@ -1,0 +1,1 @@
+"""The main formatting algorithms."""
