+++ src/sciform/format_utils/exponent_utils.py
@@ -17,7 +17,7 @@ from sciform.options.option_types import (
     ExpModeEnum,
 )
 
-if TYPE_CHECKING:
+if TYPE_CHECKING:  # pragma: no cover
     from decimal import Decimal
 
 
