### 2024-02-17 #121 @Batalex
It should be good: https://github.com/pyOpenSci/pyopensci.github.io/blob/main/_data/packages.yml#L110-L150 but the CI job failed: https://github.com/pyOpenSci/pyopensci.github.io/actions/runs/7909228977/job/21589918535.

@lwasser I am not familiar with Jekyll, is this a one-off error, or is there something more to it?
