### 2024-02-17 4ceacff
rename translation dicts
