### 2024-02-17 2d1e8b4
refactor mantissa extraction
