### 2024-02-17 11d751e
factor constant definitions out of functions
