+++ src/sciform/format_utils/make_strings.py
@@ -16,7 +16,7 @@ from sciform.options.option_types import (
     SignModeEnum,
 )
 
-if TYPE_CHECKING:
+if TYPE_CHECKING:  # pragma: no cover
     from decimal import Decimal
 
 
