+++ tests/feature/test_val_formatter.py
@@ -2,21 +2,22 @@ import unittest
 from decimal import Decimal
 
 from sciform import AutoDigits, Formatter
+from sciform.format_utils import Number
 
-FloatFormatterCases = list[tuple[float, list[tuple[Formatter, str]]]]
+ValFormatterCases = list[tuple[Number, list[tuple[Formatter, str]]]]
 
 
-class TestFormatting(unittest.TestCase):
-    def run_float_formatter_cases(self, cases_list: FloatFormatterCases):
-        for num, formats_list in cases_list:
-            for formatter, expected_num_str in formats_list:
-                num_str = formatter(num)
+class TestValFormatter(unittest.TestCase):
+    def run_val_formatter_cases(self, cases_list: ValFormatterCases):
+        for number, formatter_cases in cases_list:
+            for formatter, expected_output in formatter_cases:
+                actual_output = formatter(number)
                 with self.subTest(
-                    num=num,
-                    expected_num_str=expected_num_str,
-                    actual_num_str=num_str,
+                    number=number,
+                    expected_output=expected_output,
+                    actual_output=actual_output,
                 ):
-                    self.assertEqual(num_str, expected_num_str)
+                    self.assertEqual(expected_output, actual_output)
 
     def test_superscript(self):
         cases_list = [
@@ -51,7 +52,7 @@ class TestFormatting(unittest.TestCase):
             ),
         ]
 
-        self.run_float_formatter_cases(cases_list)
+        self.run_val_formatter_cases(cases_list)
 
     def test_left_pad_and_separators(self):
         cases_list = [
@@ -121,7 +122,7 @@ class TestFormatting(unittest.TestCase):
             ),
         ]
 
-        self.run_float_formatter_cases(cases_list)
+        self.run_val_formatter_cases(cases_list)
 
     def test_nan(self):
         cases_list = [
@@ -147,7 +148,7 @@ class TestFormatting(unittest.TestCase):
             ),
         ]
 
-        self.run_float_formatter_cases(cases_list)
+        self.run_val_formatter_cases(cases_list)
 
     def test_inf(self):
         cases_list = [
@@ -172,7 +173,7 @@ class TestFormatting(unittest.TestCase):
                 ],
             ),
         ]
-        self.run_float_formatter_cases(cases_list)
+        self.run_val_formatter_cases(cases_list)
 
     def test_zero(self):
         cases_list = [
@@ -195,7 +196,7 @@ class TestFormatting(unittest.TestCase):
                 ],
             ),
         ]
-        self.run_float_formatter_cases(cases_list)
+        self.run_val_formatter_cases(cases_list)
 
     def test_parts_per_exp(self):
         cases_list = [
@@ -281,7 +282,7 @@ class TestFormatting(unittest.TestCase):
             ),
         ]
 
-        self.run_float_formatter_cases(cases_list)
+        self.run_val_formatter_cases(cases_list)
 
     def test_no_options(self):
         formatter = Formatter()
@@ -316,7 +317,7 @@ class TestFormatting(unittest.TestCase):
             ),
         ]
 
-        self.run_float_formatter_cases(cases_list)
+        self.run_val_formatter_cases(cases_list)
 
     def test_decimal_normalization(self):
         formatter = Formatter(ndigits=AutoDigits)
