+++ tests/feature/test_separators.py
@@ -1,34 +1,35 @@
 import unittest
 
 from sciform import Formatter
+from sciform.format_utils import Number
 
-FloatFormatterCases = list[tuple[float, list[tuple[Formatter, str]]]]
-ValUncFormatterCases = list[tuple[tuple[float, float], list[tuple[Formatter, str]]]]
+FloatFormatterCases = list[tuple[Number, list[tuple[Formatter, str]]]]
+ValUncFormatterCases = list[tuple[tuple[Number, Number], list[tuple[Formatter, str]]]]
 
 
 class TestSeparators(unittest.TestCase):
-    def run_float_formatter_cases(self, cases_list: FloatFormatterCases):
-        for num, formats_list in cases_list:
-            for formatter, expected_num_str in formats_list:
-                num_str = formatter(num)
+    def run_val_formatter_cases(self, cases_list: FloatFormatterCases):
+        for number, formatter_cases in cases_list:
+            for formatter, expected_output in formatter_cases:
+                actual_output = formatter(number)
                 with self.subTest(
-                    num=num,
-                    expected_num_str=expected_num_str,
-                    actual_num_str=num_str,
+                    number=number,
+                    expected_output=expected_output,
+                    actual_output=actual_output,
                 ):
-                    self.assertEqual(num_str, expected_num_str)
+                    self.assertEqual(expected_output, actual_output)
 
     def run_val_unc_formatter_cases(self, cases_list: ValUncFormatterCases):
-        for (val, unc), formats_list in cases_list:
-            for formatter, expected_val_unc_str in formats_list:
-                num_str = formatter(val, unc)
+        for (val, unc), formatter_cases in cases_list:
+            for formatter, expected_output in formatter_cases:
+                actual_output = formatter(val, unc)
                 with self.subTest(
-                    val=val,
-                    unc=unc,
-                    expected_num_str=expected_val_unc_str,
-                    actual_num_str=num_str,
+                    value=val,
+                    uncertainty=unc,
+                    expected_output=expected_output,
+                    actual_output=actual_output,
                 ):
-                    self.assertEqual(num_str, expected_val_unc_str)
+                    self.assertEqual(expected_output, actual_output)
 
     def test_val_separators(self):
         cases_list = [
@@ -314,7 +315,7 @@ class TestSeparators(unittest.TestCase):
             ),
         ]
 
-        self.run_float_formatter_cases(cases_list)
+        self.run_val_formatter_cases(cases_list)
 
     def test_val_unc_separators(self):
         cases_list = [
