### 2024-02-18 57abe96
validation on translation dicts
