### 2024-02-18 753c905
test decimal place and rename digit_place -> decimal_place throughout
