### 2024-02-18 b882896
tests for get_bottom_dec_place. Do normalization in the function. Less likely non-normalized input can cause bugs.
