### 2024-02-18 761a606
test get_val_unc_top_dec_place
