### 2024-02-18 fca9c3a
more tests and runtime validation
