+++ tests/unit/test_number_utils.py
@@ -3,6 +3,7 @@ from __future__ import annotations
 import unittest
 from decimal import Decimal
 
+from sciform import AutoExpVal
 from sciform.format_utils import numbers
 
 
@@ -40,6 +41,7 @@ class TestNumberUtils(unittest.TestCase):
 
     def test_get_bottom_dec_place(self):
         cases: list[tuple[Decimal, int]] = [
+            # Unnormalized:
             (Decimal("10000000.0000001"), -7),
             (Decimal("10000000.0000010"), -6),
             (Decimal("10000000.0000100"), -5),
@@ -55,6 +57,7 @@ class TestNumberUtils(unittest.TestCase):
             (Decimal("10100000.0000000"), +5),
             (Decimal("11000000.0000000"), +6),
             (Decimal("10000000.0000000"), +7),
+            # Normalized:
             (Decimal("1.000001"), -6),
             (Decimal("1.00001"), -5),
             (Decimal("1.0001"), -4),
@@ -92,6 +95,15 @@ class TestNumberUtils(unittest.TestCase):
             ((Decimal("123"), Decimal("0.456"), 4), 4),
             ((Decimal("123"), Decimal("0.456"), 5), 5),
             ((Decimal("123"), Decimal("0.456"), 6), 6),
+            ((Decimal("0.456"), Decimal("123"), 0), 0),
+            ((Decimal("0.456"), Decimal("123"), 1), 1),
+            ((Decimal("0.456"), Decimal("123"), 2), 2),
+            ((Decimal("0.456"), Decimal("123"), 3), 3),
+            ((Decimal("0.456"), Decimal("123"), 4), 4),
+            ((Decimal("0.456"), Decimal("123"), 5), 5),
+            ((Decimal("0.456"), Decimal("123"), 6), 6),
+            ((Decimal("0.000123"), Decimal("0.000000456"), -1), 0),
+            ((Decimal("0.000000456"), Decimal("0.000123"), -1), 0),
         ]
 
         left_pad_matching = False
@@ -128,6 +140,8 @@ class TestNumberUtils(unittest.TestCase):
             ((Decimal("0.456"), Decimal("123"), 4), 4),
             ((Decimal("0.456"), Decimal("123"), 5), 5),
             ((Decimal("0.456"), Decimal("123"), 6), 6),
+            ((Decimal("0.000123"), Decimal("0.000000456"), -1), 0),
+            ((Decimal("0.000000456"), Decimal("0.000123"), -1), 0),
         ]
 
         left_pad_matching = True
@@ -143,6 +157,58 @@ class TestNumberUtils(unittest.TestCase):
                 unc_mantissa=unc,
                 input_top_dec_place=input_top_dec_place,
                 left_pad_matching=left_pad_matching,
+                expected_output=expected_output,
+                actual_output=actual_output,
+            ):
+                self.assertEqual(expected_output, actual_output)
+
+    def test_get_fixed_exp(self):
+        self.assertEqual(0, numbers.get_fixed_exp(0))
+        self.assertEqual(0, numbers.get_fixed_exp(AutoExpVal))
+        self.assertRaises(
+            ValueError,
+            numbers.get_fixed_exp,
+            3
+        )
+
+    def test_get_scientific_exp(self):
+        cases: list[tuple[tuple[Decimal, int | type(AutoExpVal)], int]] = [
+            ((Decimal("1000000"), AutoExpVal), 6),
+            ((Decimal("100000"), AutoExpVal), 5),
+            ((Decimal("10000"), AutoExpVal), 4),
+            ((Decimal("1000"), AutoExpVal), 3),
+            ((Decimal("100"), AutoExpVal), 2),
+            ((Decimal("10"), AutoExpVal), 1),
[…]
