### 2024-02-19 f325ddc
add binary tests and mantissa factors for more tests
