+++ tests/unit/test_number_utils.py
@@ -27,17 +27,20 @@ class TestNumberUtils(unittest.TestCase):
             (Decimal("1e+20"), 20),
         ]
 
-        for number, expected_output in cases:
-            actual_output = numbers.get_top_dec_place(number)
-            with self.subTest(
-                number=number,
-                expected_output=expected_output,
-                actual_output=actual_output,
-            ):
-                self.assertEqual(
-                    expected_output,
-                    actual_output,
-                )
+        for base_number, expected_output in cases:
+            for factor in [Decimal(1), Decimal(5)]:
+                test_number = factor * base_number
+                actual_output = numbers.get_top_dec_place(test_number)
+                with self.subTest(
+                    base_number=base_number,
+                    test_number=test_number,
+                    expected_output=expected_output,
+                    actual_output=actual_output,
+                ):
+                    self.assertEqual(
+                        expected_output,
+                        actual_output,
+                    )
 
     def test_get_bottom_dec_place(self):
         cases: list[tuple[Decimal, int]] = [
@@ -74,10 +77,13 @@ class TestNumberUtils(unittest.TestCase):
             (Decimal("1E+7"), +7),
         ]
 
-        for number, expected_output in cases:
-            actual_output = numbers.get_bottom_dec_place(number)
+        for base_number, expected_output in cases:
+            for factor in [Decimal(1), Decimal(5)]:
+                test_number = factor * base_number
+            actual_output = numbers.get_bottom_dec_place(test_number)
             with self.subTest(
-                number=number,
+                base_number=base_number,
+                test_number=test_number,
                 expected_output=expected_output,
                 actual_output=actual_output,
             ):
@@ -201,15 +207,226 @@ class TestNumberUtils(unittest.TestCase):
             ((Decimal("0.000001"), 3), 3),
         ]
 
-        for (number, input_exp), expected_output in cases:
-            actual_output = numbers.get_scientific_exp(
-                number,
-                input_exp,
-            )
-            with self.subTest(
-                number=number,
-                input_exp=input_exp,
-                expected_output=expected_output,
-                actual_output=actual_output,
-            ):
-                self.assertEqual(expected_output, actual_output)
+        for (base_number, input_exp), expected_output in cases:
+            for factor in [Decimal(1), Decimal(5)]:
+                test_number = factor * base_number
+                actual_output = numbers.get_scientific_exp(
+                    test_number,
+                    input_exp,
+                )
+                with self.subTest(
+                    base_number=base_number,
+                    test_number=test_number,
+                    input_exp=input_exp,
+                    expected_output=expected_output,
+                    actual_output=actual_output,
+                ):
+                    self.assertEqual(expected_output, actual_output)
+
+    def test_get_engineering_exp(self):
+        cases: list[tuple[tuple[Decimal, int | type(AutoExpVal)], int]] = [
+            ((Decimal("1000000"), AutoExpVal), 6),
+            ((Decimal("100000"), AutoExpVal), 3),
+            ((Decimal("10000"), AutoExpVal), 3),
+            ((Decimal("1000"), AutoExpVal), 3),
+            ((Decimal("100"), AutoExpVal), 0),
[…]
