### 2024-02-19 4c15076
mantissa_exp_base test cases
