+++ tests/unit/test_number_utils.py
@@ -2,9 +2,18 @@ from __future__ import annotations
 
 import unittest
 from decimal import Decimal
+from typing import Union
 
 from sciform import AutoExpVal
 from sciform.format_utils import numbers
+from sciform.options.option_types import ExpModeEnum
+
+MantissaExpBaseCase = list[
+    tuple[
+        tuple[Decimal, ExpModeEnum, Union[int, type(AutoExpVal)]],
+        tuple[Decimal, int, int],
+    ]
+]
 
 
 class TestNumberUtils(unittest.TestCase):
@@ -430,3 +439,230 @@ class TestNumberUtils(unittest.TestCase):
             5,
             iec=True,
         )
+
+    def test_get_mantissa_exp_base(self):
+        cases: MantissaExpBaseCase = [
+            (
+                (Decimal("123456"), ExpModeEnum.FIXEDPOINT, AutoExpVal),
+                (Decimal("123456"), 0, 10),
+            ),
+            (
+                (Decimal("1234.56"), ExpModeEnum.FIXEDPOINT, AutoExpVal),
+                (Decimal("1234.56"), 0, 10),
+            ),
+            (
+                (Decimal("12.3456"), ExpModeEnum.FIXEDPOINT, AutoExpVal),
+                (Decimal("12.3456"), 0, 10),
+            ),
+            (
+                (Decimal("0.123456"), ExpModeEnum.FIXEDPOINT, AutoExpVal),
+                (Decimal("0.123456"), 0, 10),
+            ),
+            (
+                (Decimal("0.00123456"), ExpModeEnum.FIXEDPOINT, AutoExpVal),
+                (Decimal("0.00123456"), 0, 10),
+            ),
+            (
+                (Decimal("0.0000123456"), ExpModeEnum.FIXEDPOINT, AutoExpVal),
+                (Decimal("0.0000123456"), 0, 10),
+            ),
+            (
+                (Decimal("123456"), ExpModeEnum.PERCENT, AutoExpVal),
+                (Decimal("123456"), 0, 10),
+            ),
+            (
+                (Decimal("1234.56"), ExpModeEnum.PERCENT, AutoExpVal),
+                (Decimal("1234.56"), 0, 10),
+            ),
+            (
+                (Decimal("12.3456"), ExpModeEnum.PERCENT, AutoExpVal),
+                (Decimal("12.3456"), 0, 10),
+            ),
+            (
+                (Decimal("0.123456"), ExpModeEnum.PERCENT, AutoExpVal),
+                (Decimal("0.123456"), 0, 10),
+            ),
+            (
+                (Decimal("0.00123456"), ExpModeEnum.PERCENT, AutoExpVal),
+                (Decimal("0.00123456"), 0, 10),
+            ),
+            (
+                (Decimal("0.0000123456"), ExpModeEnum.PERCENT, AutoExpVal),
+                (Decimal("0.0000123456"), 0, 10),
+            ),
+            (
+                (Decimal("123456"), ExpModeEnum.SCIENTIFIC, AutoExpVal),
+                (Decimal("1.23456"), 5, 10),
+            ),
+            (
+                (Decimal("1234.56"), ExpModeEnum.SCIENTIFIC, AutoExpVal),
+                (Decimal("1.23456"), 3, 10),
+            ),
+            (
+                (Decimal("12.3456"), ExpModeEnum.SCIENTIFIC, AutoExpVal),
+                (Decimal("1.23456"), 1, 10),
+            ),
+            (
+                (Decimal("0.123456"), ExpModeEnum.SCIENTIFIC, AutoExpVal),
+                (Decimal("1.23456"), -1, 10),
+            ),
+            (
[…]
