### 2024-02-20 27be4a2
more mantissa_exp_base tests
