+++ tests/unit/test_number_utils.py
@@ -666,3 +666,230 @@ class TestNumberUtils(unittest.TestCase):
                 self.assertEqual(expected_mantissa, actual_mantissa)
                 self.assertEqual(expected_exp, actual_exp)
                 self.assertEqual(expected_base, actual_base)
+
+    def test_get_mantissa_exp_base_input_exp(self):
+        cases: MantissaExpBaseCase = [
+            (
+                (Decimal("123456"), ExpModeEnum.FIXEDPOINT, 0),
+                (Decimal("123456"), 0, 10),
+            ),
+            (
+                (Decimal("1234.56"), ExpModeEnum.FIXEDPOINT, 0),
+                (Decimal("1234.56"), 0, 10),
+            ),
+            (
+                (Decimal("12.3456"), ExpModeEnum.FIXEDPOINT, 0),
+                (Decimal("12.3456"), 0, 10),
+            ),
+            (
+                (Decimal("0.123456"), ExpModeEnum.FIXEDPOINT, 0),
+                (Decimal("0.123456"), 0, 10),
+            ),
+            (
+                (Decimal("0.00123456"), ExpModeEnum.FIXEDPOINT, 0),
+                (Decimal("0.00123456"), 0, 10),
+            ),
+            (
+                (Decimal("0.0000123456"), ExpModeEnum.FIXEDPOINT, 0),
+                (Decimal("0.0000123456"), 0, 10),
+            ),
+            (
+                (Decimal("123456"), ExpModeEnum.PERCENT, 0),
+                (Decimal("123456"), 0, 10),
+            ),
+            (
+                (Decimal("1234.56"), ExpModeEnum.PERCENT, 0),
+                (Decimal("1234.56"), 0, 10),
+            ),
+            (
+                (Decimal("12.3456"), ExpModeEnum.PERCENT, 0),
+                (Decimal("12.3456"), 0, 10),
+            ),
+            (
+                (Decimal("0.123456"), ExpModeEnum.PERCENT, 0),
+                (Decimal("0.123456"), 0, 10),
+            ),
+            (
+                (Decimal("0.00123456"), ExpModeEnum.PERCENT, 0),
+                (Decimal("0.00123456"), 0, 10),
+            ),
+            (
+                (Decimal("0.0000123456"), ExpModeEnum.PERCENT, 0),
+                (Decimal("0.0000123456"), 0, 10),
+            ),
+            (
+                (Decimal("123456"), ExpModeEnum.SCIENTIFIC, 2),
+                (Decimal("1234.56"), 2, 10),
+            ),
+            (
+                (Decimal("1234.56"), ExpModeEnum.SCIENTIFIC, 2),
+                (Decimal("12.3456"), 2, 10),
+            ),
+            (
+                (Decimal("12.3456"), ExpModeEnum.SCIENTIFIC, 2),
+                (Decimal("0.123456"), 2, 10),
+            ),
+            (
+                (Decimal("0.123456"), ExpModeEnum.SCIENTIFIC, 2),
+                (Decimal("0.00123456"), 2, 10),
+            ),
+            (
+                (Decimal("0.00123456"), ExpModeEnum.SCIENTIFIC, 2),
+                (Decimal("0.0000123456"), 2, 10),
+            ),
+            (
+                (Decimal("0.0000123456"), ExpModeEnum.SCIENTIFIC, 2),
+                (Decimal("0.000000123456"), 2, 10),
+            ),
+            (
+                (Decimal("123456"), ExpModeEnum.ENGINEERING, 3),
+                (Decimal("123.456"), 3, 10),
+            ),
+            (
+                (Decimal("1234.56"), ExpModeEnum.ENGINEERING, 3),
[…]
