### 2024-02-20 f8bbb39
more nan coverage
