### 2024-02-24 f2104b2
more rounding tests and NanTestCase
