### 2024-02-24 ffa63af
Handle infinite cases better
