### 2024-02-25 f27c035
test exp utils
