+++ tests/unit/test_exponent_utils.py
@@ -1,8 +1,11 @@
+from __future__ import annotations
+
 import unittest
-from typing import Literal
+from decimal import Decimal
+from typing import Literal, Union
 
 from sciform.format_utils import exp_translations, exponents
-from sciform.options.option_types import ExpFormatEnum
+from sciform.options.option_types import AutoExpVal, ExpFormatEnum, ExpModeEnum
 
 Base = Literal[10, 2]
 GetTranslationDictCase = tuple[
@@ -14,6 +17,11 @@ GetTranslationDictCase = tuple[
         dict[int, str],
     ],
     dict[int, str],
+]
+GetStandardExpStrCase = tuple[tuple[int, int, bool], str]
+GetValUncExpCase = tuple[
+    tuple[Decimal, Decimal, ExpModeEnum, Union[int, type[AutoExpVal]]],
+    int,
 ]
 
 
@@ -70,3 +78,120 @@ class TestExponentUtils(unittest.TestCase):
         self.assertEqual("Kb", translation_dict[10])
         self.assertEqual("MiB", translation_dict[20])
         self.assertEqual(None, translation_dict[30])
+
+    def test_get_standard_exp_str(self):
+        cases: list[GetStandardExpStrCase] = [
+            ((10, -111, False), "e-111"),
+            ((10, -15, False), "e-15"),
+            ((10, -2, False), "e-02"),
+            ((10, -1, False), "e-01"),
+            ((10, 0, False), "e+00"),
+            ((10, +1, False), "e+01"),
+            ((10, +2, False), "e+02"),
+            ((10, +15, False), "e+15"),
+            ((10, +111, False), "e+111"),
+            ((10, -111, True), "E-111"),
+            ((10, -15, True), "E-15"),
+            ((10, -2, True), "E-02"),
+            ((10, -1, True), "E-01"),
+            ((10, 0, True), "E+00"),
+            ((10, +1, True), "E+01"),
+            ((10, +2, True), "E+02"),
+            ((10, +15, True), "E+15"),
+            ((10, +111, True), "E+111"),
+            ((2, -111, False), "b-111"),
+            ((2, -15, False), "b-15"),
+            ((2, -2, False), "b-02"),
+            ((2, -1, False), "b-01"),
+            ((2, 0, False), "b+00"),
+            ((2, +1, False), "b+01"),
+            ((2, +2, False), "b+02"),
+            ((2, +15, False), "b+15"),
+            ((2, +111, False), "b+111"),
+            ((2, -111, True), "B-111"),
+            ((2, -15, True), "B-15"),
+            ((2, -2, True), "B-02"),
+            ((2, -1, True), "B-01"),
+            ((2, 0, True), "B+00"),
+            ((2, +1, True), "B+01"),
+            ((2, +2, True), "B+02"),
+            ((2, +15, True), "B+15"),
+            ((2, +111, True), "B+111"),
+        ]
+
+        for (base, exp_val, capitalize), expected_output in cases:
+            actual_output = exponents.get_standard_exp_str(
+                base=base,
+                exp_val=exp_val,
+                capitalize=capitalize,
+            )
+            with self.subTest(
+                base=base,
+                exp_val=exp_val,
+                capitalize=capitalize,
+            ):
+                self.assertEqual(expected_output, actual_output)
+
+    def test_get_superscript_exp_str(self):
+        cases: list[tuple[tuple[int, int], str]] = [
+            ((10, -10), "×10⁻¹⁰"),
+            ((10, -1), "×10⁻¹"),
[…]
