### 2024-02-25 f237e8b
more exp_utils tests
