### 2024-02-26 459a676
refacotor out construct_num_str
