### 2024-02-26 f4ad767
test_make_strings
