+++ tests/unit/test_grouping.py
@@ -84,6 +84,9 @@ class TestGrouping(unittest.TestCase):
             ("1234567.76543", "1u234u567d765l43"),
             ("1234567.765432", "1u234u567d765l432"),
             ("1234567.7654321", "1u234u567d765l432l1"),
+            ("+   1234567.7654321", "+   1u234u567d765l432l1"),
+            ("+0001234567.7654321", "+0u001u234u567d765l432l1"),
+            ("    1234567.7654321", "    1u234u567d765l432l1"),
         ]
 
         separator_cases: list[tuple[str, str, str]] = [
