### 2024-02-26 c5fae04
abs_num_str_by_bottom_dec_place test
