### 2024-02-26 97fcf98
test_construct_num_str
