### 2024-02-26 9fac59a
refactor get_exp_str call out of construct_val_unc_exp_str
