### 2024-02-26 f293d49
test construct_val_unc_exp_str
100% coverage of format_utils via unit tests!
