### 2024-02-26 6cfec4c
move format_utils tests into dedicated directory, add some invalid modes tests
