### 2024-02-26 7241444
remove redundant tests from test_invalid_options and migrate into format_utils tests
