### 2024-05-02 #121 @cmarmo
Hello everybody, I see `sciform` has been successfully added to the pyOpenSci website.
Wearing my triaging hat, I'm going to close this issue as successfully completed! :rocket:
